DeclareDesign package

The four main types of functions are to declare a step, to combine steps into designs, and to manipulate designs and designers (functions that return designs).

Design Steps

declare_population

Population step

declare_potential_outcomes

Potential outcomes step

declare_sampling

Sampling step

declare_assignment

Assignment step

reveal_outcomes

Reveal outcomes step

declare_estimand

Estimand step

declare_estimator

Estimator step

Design Objects

+

Add steps to create a design

draw_data

Simulate the DGP

run_design

Simulate the DGP with estimands/estimators

diagnose_design

Diagnose a design

cite_design

Cite a design

Design Editing

modify_design

Add, delete or replace a step

redesign

Modify local variables within a design (advanced)

Designers

expand_design

Generate designs from a designer

designs

See also the DesignLibrary package for designers to use