Deprecated functions in package DeclareDesign.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("-deprecated").

Usage

tidy_estimator(estimator_function)

model_handler(...)

tidy_estimator

For tidy_estimator, use label_estimator.

model_handler

For model_handler, use method_handler.