Declare measurement procedure

Add a step to a design that explicitly indicates what variables are measured by a design. It is always possible to include measured variables directly in your declare_model calls or using declare_step. However, putting latent variables in declare_model and variables-as-measured in declare_measurement helps communicate which parts of your research design are in the model and which parts are in the data strategy.

Usage

declare_measurement(..., handler = measurement_handler, label = NULL)

measurement_handler(data, ...)

potential_outcomes_handler(data, ...)

Arguments

...

arguments to be captured, and later passed to the handler

handler

a tidy-in, tidy-out function

label

a string describing the step

data

A data.frame.

Value

A function that returns a data.frame.

Examples

# declare_measurement a in use a
## Two-arm randomized experiment
design <-
  declare_model(
    N = 500,
    X = rep(c(0, 1), each = N / 2),
    U = rnorm(N, sd = 0.25),
    potential_outcomes(Y ~ 0.2 * Z + X + U)
  ) +
  declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(S = complete_rs(N = N, n = 200)) +
  declare_assignment(Z = complete_ra(N = N, m = 100)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z)) +
  declare_estimator(Y ~ Z, inquiry = "ATE")
  
run_design(design)
#>   inquiry estimand estimator term estimate  std.error statistic   p.value
#> 1     ATE      0.2 estimator    Z 0.125076 0.07869747  1.589327 0.1135821
#>      conf.low conf.high  df outcome
#> 1 -0.03011679 0.2802688 198       Y

# Reveal potential outcomes according to treatment assignment
design <-
  declare_model(N = 100,
                potential_outcomes(Y ~ rbinom(
                  N, size = 1, prob = 0.1 * Z + 0.5
                ))) +
  declare_assignment(Z = complete_ra(N, m = 50)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z))
  
head(draw_data(design))
#>    ID Y_Z_0 Y_Z_1 Z Y
#> 1 001     1     1 0 1
#> 2 002     1     0 0 1
#> 3 003     1     1 0 1
#> 4 004     1     0 1 0
#> 5 005     0     0 1 0
#> 6 006     0     0 1 0

# Generate observed measurement from a latent value
design <- 
  declare_model(N = 100, latent = runif(N)) +
  declare_measurement(observed = rbinom(N, 1, prob = latent))
  
head(draw_data(design))
#>    ID    latent observed
#> 1 001 0.4567889        1
#> 2 002 0.5209913        1
#> 3 003 0.5580132        1
#> 4 004 0.5682925        1
#> 5 005 0.1825915        1
#> 6 006 0.8342294        1

if(require("psych")) {

  # Index creation
  
  library(psych)
  
  design <-
    declare_model(
      N = 500,
      X = rep(c(0, 1), each = N / 2),
      Y_1 = 0.2 * X + rnorm(N, sd = 0.25),
      Y_2 = 0.3 * X + 0.5 * rnorm(N, sd = 0.50),
      Y_3 = 0.1 * X + 0.4 * rnorm(N, sd = 0.75)) +
    declare_measurement(
      index = fa(
        r = cbind(Y_1, Y_2, Y_3),
        nfactors = 1,
        rotate = "varimax"
      )$scores
    )
  
  draw_data(design)
}
#> Loading required package: psych
#>      ID X           Y_1           Y_2           Y_3           MR1
#> 1   001 0  1.436340e-01  0.0100889975  0.5070853746 -0.0501194068
#> 2   002 0  4.352511e-01  0.0667326329  0.0505785018  0.2260224559
#> 3   003 0 -3.931334e-01  0.1260055693  0.5006337273 -0.4459965914
#> 4   004 0 -2.127862e-01  0.3909005890 -0.1726552829 -0.0333183479
#> 5   005 0  3.707238e-02  0.1200579262  0.0789927756 -0.1076631456
#> 6   006 0  4.449162e-01 -0.1786820750  0.0823879274 -0.1194690967
#> 7   007 0 -1.646548e-01  0.4132438880 -0.3350441829  0.0095272717
#> 8   008 0 -3.955433e-02  0.0841076372 -0.5899077160 -0.4107371363
#> 9   009 0 -7.864121e-02  0.2371355526  0.2679584933 -0.0084321140
#> 10  010 0  3.607595e-01 -0.2769142548 -0.4645975189 -0.4919894933
#> 11  011 0  1.902256e-02  0.0856300752 -0.4111506228 -0.3015488596
#> 12  012 0 -3.752159e-01 -0.2036478653  0.0955754484 -1.0180006322
#> 13  013 0  1.120362e-01  0.3154357722 -0.0664709123  0.2242022021
#> 14  014 0 -3.406056e-01 -0.0242375869 -0.0505207581 -0.7525144078
#> 15  015 0  7.520434e-02  0.2005254038 -0.2067361932 -0.0203786697
#> 16  016 0 -1.892312e-01 -0.0263568903  0.7369081644 -0.3970930604
#> 17  017 0 -2.231076e-01 -0.4032972479  0.1921902701 -1.1290809120
#> 18  018 0 -5.256932e-02 -0.1594098262 -0.0736182903 -0.6549374831
#> 19  019 0  1.158012e-02  0.0556118962 -0.0800291396 -0.2702332658
#> 20  020 0 -7.087843e-02  0.0264092801  0.3340643146 -0.2958548624
#> 21  021 0  1.466869e-01 -0.1212543743  0.2336515276 -0.3106380020
#> 22  022 0  4.406545e-01  0.4206105042 -0.4924923112  0.6189912270
#> 23  023 0 -5.685637e-01  0.3378761265 -0.6459162079 -0.6066476946
#> 24  024 0 -2.137801e-01  0.0679226830 -0.0762663886 -0.4886887246
#> 25  025 0  2.764089e-01  0.2566742304  0.0058644116  0.3287259777
#> 26  026 0  5.145359e-01 -0.1908389446  0.2970288869 -0.0098437841
#> 27  027 0 -1.634764e-01  0.2463094785 -0.3423914892 -0.2384934362
#> 28  028 0 -4.381408e-02 -0.2822367729  0.0906542240 -0.7862410162
#> 29  029 0  4.341497e-01  0.5468481203 -0.2552967269  0.8591475325
#> 30  030 0  1.108926e-01  0.0854559204  0.1149337820 -0.0720191046
#> 31  031 0 -1.738337e-01  0.0011241528  0.3400134269 -0.4403965347
#> 32  032 0  1.399974e-01 -0.2886739886 -0.3880361540 -0.7228785413
#> 33  033 0 -1.024948e-01  0.2116799723 -0.0591272580 -0.1539466766
#> 34  034 0  4.275750e-01 -0.1831148314 -0.2462006523 -0.2273385663
#> 35  035 0  7.542894e-02  0.3893385310  0.1910866975  0.3601949183
#> 36  036 0 -5.884516e-03  0.4321735685  0.3046488261  0.3666205477
#> 37  037 0  3.186044e-01 -0.1480731413  0.0367666387 -0.2188310889
#> 38  038 0  2.500067e-01  0.3142244965  0.0354344328  0.3936473611
#> 39  039 0 -3.033400e-01 -0.2435009780 -0.0457275077 -1.0369674017
#> 40  040 0  7.809331e-02  0.2234033707  0.4192101701  0.1747102093
#> 41  041 0 -1.667727e-01  0.0438320880 -0.2831386118 -0.5270843592
#> 42  042 0  3.603200e-02 -0.1648343270 -0.0823726871 -0.5717549408
#> 43  043 0  1.600608e-01 -0.3882071138 -0.5217387737 -0.8830131092
#> 44  044 0  3.400212e-01 -0.1442711192  0.2474608520 -0.1373826920
#> 45  045 0  3.681884e-01  0.0274312562  0.0734612938  0.1028360637
#> 46  046 0  2.851582e-01 -0.2357794273 -0.2797680181 -0.4640553244
#> 47  047 0 -1.516776e-01  0.3516611938 -0.2171884485 -0.0382826969
#> 48  048 0  1.120619e-01  0.1220338825  0.3459734061  0.0417934139
#> 49  049 0 -1.699956e-01  0.5729670179  0.3481524704  0.4132139955
#> 50  050 0  1.582456e-01 -0.0396785525  0.2676190463 -0.1689670850
#> 51  051 0  4.566264e-01 -0.0328974380 -0.0597183948  0.0730407374
#> 52  052 0 -4.410141e-02  0.0895973926  0.0620923656 -0.2426782225
#> 53  053 0  3.267762e-01 -0.2809181134  0.3107701686 -0.3378754267
#> 54  054 0 -2.388060e-02  0.4211585190 -0.1344335830  0.2203905187
#> 55  055 0  1.592105e-01  0.2393398017  0.1019844943  0.2037280894
#> 56  056 0 -3.395783e-02  0.0862931781  0.0987320020 -0.2276220163
#> 57  057 0  3.185655e-01 -0.3586808978 -0.3776698988 -0.6357070981
#> 58  058 0  3.950048e-02 -0.0859849939  0.2404782919 -0.3696742873
#> 59  059 0 -2.830167e-01  0.3302817405  0.1990355686 -0.1033181037
#> 60  060 0  3.494163e-01 -0.0260564017 -0.4177447904 -0.1203276797
#> 61  061 0 -1.867871e-01 -0.1160328299 -0.1146871356 -0.7425635209
#> 62  062 0 -2.551186e-02 -0.1871337863  0.1185363978 -0.6189480100
#> 63  063 0 -4.746426e-01  0.1410029150 -0.5520106519 -0.7756587187
#> 64  064 0 -1.551986e-01 -0.1422508307 -0.1597636834 -0.7594968367
#> 65  065 0 -2.832940e-02  0.0063185813 -0.3717727570 -0.4590797972
#> 66  066 0 -1.598631e-02 -0.1547942782  0.5465682179 -0.4528380698
#> 67  067 0 -4.886321e-01 -0.0512980163  0.0574763804 -0.9214357017
#> 68  068 0  1.352984e-01  0.1820127262 -0.2545071012  0.0034874314
#> 69  069 0 -7.110285e-02 -0.1374796247 -0.4459624965 -0.7360469650
#> 70  070 0  3.687766e-01  0.2524703739 -0.5284499434  0.2849147474
#> 71  071 0  2.229915e-01 -0.5892374286 -0.3720029008 -1.0767612178
#> 72  072 0 -2.345664e-02 -0.0291719143 -0.1568073526 -0.4522326947
#> 73  073 0  4.560248e-02  0.0469825414 -0.0489495296 -0.2392927778
#> 74  074 0 -1.750897e-01  0.1913800795  0.1511846652 -0.2074545048
#> 75  075 0 -6.046945e-02  0.2343937181  0.1286664855 -0.0285228265
#> 76  076 0  1.532877e-01  0.1144176390  0.0517077841 -0.0003618504
#> 77  077 0  5.027956e-01  0.0122317412 -0.2635271150  0.1371231990
#> 78  078 0  2.311473e-01 -0.0046260207 -0.0745546681 -0.1265841890
#> 79  079 0  3.151882e-01  0.0882390223 -0.0487076385  0.1062020144
#> 80  080 0 -1.715705e-01 -0.3417018920  0.0219889169 -1.0264435959
#> 81  081 0  1.215828e-02 -0.1478999785  0.0671129063 -0.5340677252
#> 82  082 0  1.084441e-01 -0.0728044402 -0.2982024643 -0.4135254710
#> 83  083 0 -2.608939e-02  0.5108856009 -0.8974177487  0.1582861489
#> 84  084 0  1.858825e-01 -0.2360167027 -0.1606925668 -0.5390145641
#> 85  085 0  2.380004e-01 -0.0595294897  0.1463555371 -0.1449181231
#> 86  086 0  1.065975e-01  0.1392426745  0.1136952173  0.0028538836
#> 87  087 0  2.053730e-01 -0.0026062773  0.5356451054  0.0033867631
#> 88  088 0  9.336307e-02  0.0601516625 -0.1871444196 -0.2043226651
#> 89  089 0 -5.355954e-02  0.0584724514  0.3261026168 -0.2320831679
#> 90  090 0 -1.598064e-01 -0.0334721977 -0.2029422876 -0.6140475943
#> 91  091 0  2.292841e-01  0.0131866493  0.0976089829 -0.0586549528
#> 92  092 0  5.128390e-01 -0.0773149143  0.1217023812  0.1123229252
#> 93  093 0 -2.534063e-01 -0.1593083852 -0.2219162305 -0.9040430451
#> 94  094 0  6.503778e-02  0.2528983660 -0.1686916533  0.0561316694
#> 95  095 0  5.856940e-01 -0.0424997018  0.1491126016  0.2476745051
#> 96  096 0 -1.140605e-01  0.1296831870 -0.1372678500 -0.3074085439
#> 97  097 0 -2.189429e-01 -0.0873241903  0.4346239209 -0.5951501146
#> 98  098 0 -1.107661e-01 -0.1171484469 -0.1492620054 -0.6727845537
#> 99  099 0 -3.154295e-01  0.2147750334  0.1572260946 -0.3192493806
#> 100 100 0  2.525856e-01 -0.0795335783 -0.0642141950 -0.2123819338
#> 101 101 0  3.319414e-01  0.5619716075 -0.5990186928  0.6869425886
#> 102 102 0  5.782853e-02  0.3337762218  0.4262766819  0.3187049044
#> 103 103 0 -2.164195e-01 -0.1065885440 -0.2660845301 -0.7980633388
#> 104 104 0  2.548144e-01  0.1743086986  0.1521712707  0.2208450621
#> 105 105 0  2.828954e-01  0.1090260935 -0.1247269976  0.0837505901
#> 106 106 0  2.806722e-01 -0.2171988124 -0.1179211984 -0.4003602504
#> 107 107 0  9.550293e-02  0.1306846151  0.3615811146  0.0410952557
#> 108 108 0  1.575383e-01  0.1908374419  0.0501605426  0.1169886444
#> 109 109 0 -1.826104e-01  0.3717189190 -0.0251661397  0.0073356889
#> 110 110 0 -2.217286e-01  0.1717419181  0.0401390293 -0.3137960530
#> 111 111 0  2.468895e-01 -0.2090741308 -0.0100431423 -0.3966904644
#> 112 112 0 -4.518024e-01 -0.1845422636 -0.3627535512 -1.1862382427
#> 113 113 0 -1.801694e-01 -0.1190242639 -0.0831494308 -0.7320508744
#> 114 114 0 -1.157558e-01  0.2434156080 -0.3019972227 -0.1822540903
#> 115 115 0  4.268047e-01 -0.2651326884  0.0836539505 -0.2663738210
#> 116 116 0  3.525336e-01  0.2261782260 -0.1405670071  0.3268119262
#> 117 117 0 -5.660493e-01  0.1299591363  0.4942785717 -0.6240894682
#> 118 118 0 -1.174360e-01  0.3224051228  0.2233754991  0.0657688425
#> 119 119 0  9.891668e-02 -0.2452567860  0.2061445236 -0.5517421771
#> 120 120 0 -1.077275e-01  0.2998373609 -0.0552945061 -0.0278416503
#> 121 121 0  2.247801e-02  0.2660844960  0.2905284135  0.1468082621
#> 122 122 0 -2.449774e-01 -0.4911424537 -0.2228227566 -1.3871822572
#> 123 123 0 -6.515390e-02 -0.1794470823 -0.2122832959 -0.7329364823
#> 124 124 0 -1.583197e-01 -0.1782884356  0.1544318708 -0.7368214393
#> 125 125 0  1.085737e-01 -0.2436778907  0.2453061799 -0.5293248917
#> 126 126 0  6.339144e-02 -0.2346544690  0.2146276115 -0.5713485568
#> 127 127 0 -3.346311e-01 -0.0750939174 -0.1729534156 -0.8525170072
#> 128 128 0  3.883770e-01 -0.1480290612 -0.2737325098 -0.2236306049
#> 129 129 0 -1.478952e-01  0.4176277606 -0.2997590585  0.0426123542
#> 130 130 0 -2.263217e-01 -0.2857693530 -0.3229360646 -1.0884254045
#> 131 131 0 -1.408970e-01 -0.4877883096 -0.1849674516 -1.2628906718
#> 132 132 0 -3.753624e-01 -0.0709475266 -0.0685627722 -0.8629518956
#> 133 133 0 -2.111547e-01  0.0949786220 -0.8843203388 -0.6499639554
#> 134 134 0 -3.599358e-01 -0.3669690636  0.1542744851 -1.2291100019
#> 135 135 0 -2.012514e-01 -0.0685798796 -0.1525422345 -0.6970519593
#> 136 136 0  5.464643e-02 -0.3627415469 -0.0738567995 -0.8432847561
#> 137 137 0 -3.141652e-01 -0.7115336493  0.5052368220 -1.6028440189
#> 138 138 0  4.107890e-01  0.3592025266  0.2783546464  0.6912290229
#> 139 139 0 -1.246375e-01 -0.0824314409  0.0003600024 -0.5981599702
#> 140 140 0  2.961635e-02 -0.0372043401  0.4175865291 -0.2630578311
#> 141 141 0 -3.334606e-01  0.2279831880 -0.5586371594 -0.4995405349
#> 142 142 0  1.748206e-01 -0.0512482666 -0.2272245336 -0.2936496720
#> 143 143 0 -1.744681e-01  0.1667468554 -0.3686600650 -0.3746377833
#> 144 144 0 -6.567553e-01 -0.3461002301 -0.2390342133 -1.6105525708
#> 145 145 0  2.568974e-01  0.0528151649  0.7646565464  0.1977156716
#> 146 146 0 -2.182416e-01 -0.1787257611  0.0097119706 -0.8372208694
#> 147 147 0 -3.512469e-01 -0.0447862653 -0.3789341750 -0.8771591128
#> 148 148 0 -4.121290e-01 -0.2540388872 -0.4634546054 -1.2728399906
#> 149 149 0  2.182981e-02  0.4846720691 -0.3444012166  0.3096799310
#> 150 150 0  1.062008e-01 -0.4678889566 -0.0059423864 -0.9275961609
#> 151 151 0  1.067502e-01 -0.0790595484  0.4423820198 -0.2374849476
#> 152 152 0  3.746218e-01  0.0776233255  0.2879891849  0.2382052838
#> 153 153 0  4.231502e-01 -0.1982899557  0.1094907854 -0.1646351956
#> 154 154 0 -3.131021e-01 -0.1389022238 -0.2825669134 -0.9520739126
#> 155 155 0  1.896357e-01 -0.0031165475 -0.1095772935 -0.1769706396
#> 156 156 0 -1.106011e-01 -0.3564519135 -0.1238525529 -1.0208539735
#> 157 157 0  1.109618e-01  0.3578326728 -0.4678536747  0.1845008634
#> 158 158 0 -6.943343e-02  0.1442104152  0.0405562658 -0.1938927038
#> 159 159 0 -2.568602e-01  0.3445117512 -0.0033600983 -0.1057774504
#> 160 160 0 -6.070484e-02 -0.0330366285  0.0629565566 -0.4417201796
#> 161 161 0  5.146971e-01 -0.0141534174 -0.3703813605  0.0835742195
#> 162 162 0 -8.488751e-03  0.0892226560 -0.1339419372 -0.2552036866
#> 163 163 0 -3.786002e-01  0.1870130279  0.0032951556 -0.4658987497
#> 164 164 0  7.707851e-02  0.1944000117 -0.1009242615 -0.0007487301
#> 165 165 0 -2.955001e-01 -0.1562085902 -0.2157314608 -0.9422760150
#> 166 166 0 -1.840477e-01  0.0093966869 -0.0637306199 -0.5409068993
#> 167 167 0 -3.184030e-01  0.0966616202 -0.2478762943 -0.5997794788
#> 168 168 0  1.003101e-01  0.0729413323 -0.1090070127 -0.1583014031
#> 169 169 0  6.692390e-01  0.2796275529 -0.2715597626  0.7069124084
#> 170 170 0  4.101134e-01 -0.2233971493 -0.2599530350 -0.3089277888
#> 171 171 0  8.777064e-02  0.3121611360  0.1632985424  0.2518077582
#> 172 172 0 -1.764922e-01  0.0897258088 -0.1367143496 -0.4323247964
#> 173 173 0 -2.335280e-01  0.0034628074 -0.0999701722 -0.6110356993
#> 174 174 0  6.276451e-01 -0.2675750943 -0.1386860078 -0.1143700309
#> 175 175 0  2.514094e-01  0.4893893540  0.4211655529  0.7521808989
#> 176 176 0  3.703101e-02  0.3165573986 -0.1047697618  0.1370929594
#> 177 177 0  1.122411e-03  0.2527873984 -0.1560829318 -0.0082487239
#> 178 178 0  8.457809e-02 -0.2600070526 -0.1355516131 -0.6750479086
#> 179 179 0 -2.370469e-01 -0.0538219279 -0.2831879836 -0.7459331840
#> 180 180 0 -3.217669e-02 -0.1749010139  0.7807299597 -0.4405535444
#> 181 181 0  2.260988e-01  0.0883467142 -0.2765523374 -0.0451481912
#> 182 182 0  9.690334e-02 -0.0618382500  0.0629053562 -0.3182147549
#> 183 183 0  6.490896e-02  0.0201583950  0.6692001987 -0.0772588495
#> 184 184 0 -1.486228e-01  0.0885153270 -0.1061806286 -0.3970155686
#> 185 185 0  6.106299e-02 -0.1532883205 -0.6893863286 -0.6815995420
#> 186 186 0 -3.094551e-01  0.1840390314  0.6253925353 -0.2402268053
#> 187 187 0  1.958911e-01 -0.0266581167  0.1930777659 -0.1288029583
#> 188 188 0 -5.352711e-01 -0.3553887566 -0.0220718988 -1.4413965398
#> 189 189 0  5.997639e-01  0.4025776835  0.1193142228  0.9146162882
#> 190 190 0 -3.955868e-01 -0.4194137222  0.0437913131 -1.3723435077
#> 191 191 0  1.656952e-01  0.5483592528 -0.4463242812  0.5300308701
#> 192 192 0 -1.440724e-01  0.1851030914  0.1142136575 -0.1933885781
#> 193 193 0  2.733745e-01 -0.0394802495 -0.2361553505 -0.1745359943
#> 194 194 0  8.534068e-02 -0.0102992868  0.4440713178 -0.1577283282
#> 195 195 0 -6.465421e-01 -0.3509960519 -0.0134054068 -1.5500365601
#> 196 196 0 -7.312981e-02 -0.3928007194 -0.0279858939 -1.0109909331
#> 197 197 0  5.082956e-02  0.4780628140 -0.3395372416  0.3316948166
#> 198 198 0  3.198369e-01  0.1082771306  0.1738339406  0.1970237021
#> 199 199 0 -2.083076e-01 -0.2249019801  0.2741917575 -0.8283643311
#> 200 200 0  3.692386e-01  0.0065855277  0.0745807561  0.0733316559
#> 201 201 0  1.140718e-01  0.2452928272 -0.5598841098 -0.0022606214
#> 202 202 0  1.520264e-01 -0.1877577653 -0.1569894641 -0.5022589081
#> 203 203 0  7.811287e-02 -0.2389335384  0.5452528689 -0.4786386547
#> 204 204 0  2.721109e-01  0.1020953405 -0.1891960036  0.0458189584
#> 205 205 0 -1.597886e-01  0.3753409600  0.4453509458  0.1556392402
#> 206 206 0  1.480402e-01 -0.2293350646  0.2650489421 -0.4614611468
#> 207 207 0 -6.987853e-01  0.0178453225  0.0948861757 -1.0311249638
#> 208 208 0 -8.301910e-02  0.5417714034  0.0564108256  0.3849963522
#> 209 209 0 -3.606072e-01 -0.0960641062  0.0163239715 -0.8631709403
#> 210 210 0  2.487984e-01  0.4039463271  0.4380346305  0.6270572602
#> 211 211 0  2.681345e-01 -0.5254412451 -0.0373331294 -0.8500570143
#> 212 212 0 -1.773965e-01 -0.4072119823  0.2159245295 -1.0806825065
#> 213 213 0 -9.061345e-02  0.2093740217 -0.0043163786 -0.1309877313
#> 214 214 0 -5.349753e-01  0.3547420542  0.1279665010 -0.3507213590
#> 215 215 0 -6.366348e-02 -0.0823520324 -0.7250561802 -0.7170080444
#> 216 216 0 -1.590590e-01  0.2629249456 -0.3157214219 -0.2024722442
#> 217 217 0  2.458792e-02 -0.0426061728  0.0755235630 -0.3627833371
#> 218 218 0 -3.440651e-01  0.1722987536  0.3036505131 -0.3754074717
#> 219 219 0 -2.447128e-01  0.0915771407  0.4525781920 -0.3526466023
#> 220 220 0  6.278865e-02  0.1862994611  0.0468877931  0.0095189624
#> 221 221 0  4.427749e-01  0.2167078494  0.0411091763  0.4538370020
#> 222 222 0  1.640796e-01 -0.1616433043 -0.5033990229 -0.5383601856
#> 223 223 0  1.172533e-01  0.0095401644 -0.1172798372 -0.2364886181
#> 224 224 0 -2.115644e-01 -0.1981889827  0.3605625598 -0.7703881384
#> 225 225 0  5.455380e-02 -0.1131207548 -0.3019191317 -0.5310452509
#> 226 226 0 -6.526888e-02  0.0064365530 -0.2857957903 -0.4761383504
#> 227 227 0  7.585417e-02 -0.4278219934 -0.7156333585 -1.0795086584
#> 228 228 0 -1.849223e-02  0.3496962469  0.2194683502  0.2095691115
#> 229 229 0 -3.499658e-01 -0.2416992267  0.2078182560 -1.0194114262
#> 230 230 0 -1.424110e-01 -0.2796108040  0.2876071663 -0.8365664198
#> 231 231 0 -1.578678e-01  0.2345762920 -0.4255968592 -0.2709889045
#> 232 232 0 -1.277550e-01  0.2107670841  0.3792773423 -0.0711812873
#> 233 233 0  2.823405e-01  0.2660425128  0.1167144618  0.3768700536
#> 234 234 0  1.283049e-01 -0.0629266186 -0.2298193436 -0.3606659505
#> 235 235 0  2.662292e-01 -0.0967859236 -0.1051978477 -0.2339172030
#> 236 236 0  1.871283e-01  0.1998030090 -0.1639783302  0.1073811680
#> 237 237 0  5.463572e-02 -0.1152990484 -0.4801714395 -0.5792199478
#> 238 238 0 -3.039054e-02  0.1459251160  0.0793817067 -0.1403704554
#> 239 239 0 -1.953297e-01 -0.3320588087  0.1694545284 -0.9999520161
#> 240 240 0 -4.216691e-02  0.1403690094  0.4620092479 -0.0643586259
#> 241 241 0  2.674594e-01  0.5220589831 -0.1845436061  0.6645014934
#> 242 242 0  4.294530e-01  0.1399219885 -0.0283309971  0.3084440967
#> 243 243 0 -6.595482e-02  0.1951921761  0.1134187937 -0.0962586079
#> 244 244 0  1.940829e-01  0.3145746245 -0.1652262232  0.2844985210
#> 245 245 0 -2.163137e-02 -0.0199800091 -0.0290177328 -0.4044008045
#> 246 246 0 -2.945440e-01  0.2561217897 -0.1093839389 -0.3033012574
#> 247 247 0 -2.017928e-01 -0.2819461297 -0.0165905747 -0.9794989505
#> 248 248 0  9.020296e-02  0.2106410510  0.2311630217  0.1210585289
#> 249 249 0  3.080886e-01 -0.0333895058  0.3179045263  0.0110730817
#> 250 250 0  5.108000e-03  0.0211890761 -0.2560302678 -0.3725391373
#> 251 251 1 -4.458333e-02  0.2613496640  0.2129804450  0.0494807561
#> 252 252 1  4.297242e-01  0.3152407683 -0.2398845436  0.5151177043
#> 253 253 1  1.487701e-01 -0.1602595629  0.0222486348 -0.4196570293
#> 254 254 1  1.707739e-01  0.4739560202 -0.2206886582  0.4821197098
#> 255 255 1  4.245229e-01  0.1616642773 -0.1899691497  0.2946331210
#> 256 256 1  3.756468e-01  0.3618816578  0.2185775291  0.6430390577
#> 257 257 1  1.429400e-01  0.0491183499  0.1451025306 -0.0844567292
#> 258 258 1 -5.112458e-02  0.4270976072  0.1032007798  0.2604974211
#> 259 259 1  4.220458e-01  0.6528789418  0.0878174818  1.0902103711
#> 260 260 1  2.845616e-02  0.3108745851  0.2713671146  0.2146533636
#> 261 261 1  9.493573e-02 -0.0501570044  0.2242301497 -0.2622211069
#> 262 262 1  2.996157e-01  0.6768908276 -0.0805113016  0.9541643010
#> 263 263 1 -3.906050e-01  0.4122565196  0.6444429672  0.0172427590
#> 264 264 1  5.662678e-01 -0.1099662399  0.2123653283  0.1431790081
#> 265 265 1  2.437460e-01  0.1159043635  0.3653813042  0.1764783584
#> 266 266 1 -1.055057e-01  0.8120588229 -0.2531804533  0.6836526806
#> 267 267 1  6.642247e-02  0.5313415289  0.2551675166  0.5773436068
#> 268 268 1  3.143521e-01  0.4759824046  0.0692068398  0.7097696069
#> 269 269 1 -4.566895e-02  0.0006358371  0.6195293698 -0.2353499289
#> 270 270 1 -8.574234e-02  0.2772513812  0.2857983340  0.0480405897
#> 271 271 1  3.435270e-01  0.3140489813  0.6941056912  0.6584112883
#> 272 272 1  2.368386e-01  0.6895431358 -0.1701821036  0.8840609065
#> 273 273 1  3.388738e-02  0.1093153655  0.4261304970 -0.0392439807
#> 274 274 1  1.001400e-01  0.3974781758 -0.0334658705  0.3415874194
#> 275 275 1 -1.932959e-01  0.2110271997  0.1381129911 -0.2008377875
#> 276 276 1  1.335598e-01  0.4378237943  0.0345200099  0.4538002923
#> 277 277 1  1.747952e-01  0.4741097292  0.1327804843  0.5758866518
#> 278 278 1  2.221422e-01  0.4269355581  0.3008121150  0.5983515751
#> 279 279 1  3.245331e-02 -0.0002185824  0.0503066713 -0.2980385678
#> 280 280 1  9.015759e-02  0.7869924890  0.8255830994  1.1253707292
#> 281 281 1 -8.709402e-03 -0.3632550176  0.2944796072 -0.8178024756
#> 282 282 1  2.160030e-01  0.2494933227 -0.3165435549  0.1729315278
#> 283 283 1  1.100396e-01  0.6391658550  0.0226041939  0.7243884741
#> 284 284 1  6.031872e-01  0.1341334996  0.0543862578  0.5039721021
#> 285 285 1  4.474330e-01  0.1218616073 -0.2258500588  0.2507378862
#> 286 286 1 -7.348782e-02 -0.4876265930  0.0152307930 -1.1409883160
#> 287 287 1  2.219771e-01  0.3576314981 -0.0110452185  0.4166785360
#> 288 288 1  4.675911e-01  0.4701520794  0.3529605967  0.9344113443
#> 289 289 1 -1.439543e-02  0.1135148110 -0.1703472678 -0.2346272738
#> 290 290 1  1.744594e-01  0.1008150483 -0.0771178328 -0.0307411613
#> 291 291 1  1.510436e-02 -0.2414569328 -0.3302699810 -0.7700109292
#> 292 292 1  3.242650e-01  0.3750217740  0.4654965393  0.6707096570
#> 293 293 1  5.388837e-01  0.4313548986  0.6457913730  1.0260716986
#> 294 294 1  3.600823e-01  0.6498316058 -0.4233888132  0.8912026029
#> 295 295 1  2.743864e-01 -0.2755287127  0.4283963239 -0.3554187276
#> 296 296 1  1.190235e-01  0.5495191283  0.0321211393  0.6034045615
#> 297 297 1  3.991981e-01  0.2036374059 -0.3343290027  0.2936636177
#> 298 298 1 -2.030256e-01  0.4747122302 -0.3021785292  0.0684667188
#> 299 299 1 -2.074211e-01  0.1307280934 -0.6880925579 -0.5434699127
#> 300 300 1  2.616362e-01  0.3643796722 -0.1435250385  0.4350328205
#> 301 301 1  5.879013e-02  0.6279999226 -0.2324033204  0.5893714856
#> 302 302 1  6.256273e-01  0.5635261175  0.5055206284  1.2779950833
#> 303 303 1 -8.102813e-02  0.2854771390 -0.2267242236 -0.0642779747
#> 304 304 1  2.807759e-02  0.3191478904 -0.0179107689  0.1534340799
#> 305 305 1  3.841675e-01  0.3442520398 -0.2733463036  0.5016192631
#> 306 306 1  1.470545e-01  0.4037494039 -0.2072879058  0.3564415489
#> 307 307 1  4.580996e-01  0.2633629775  0.0190942144  0.5335815675
#> 308 308 1  2.079510e-01  0.6603189385 -0.6225214694  0.6960090772
#> 309 309 1  3.670519e-01  0.0529788909  0.1957411163  0.1703928559
#> 310 310 1  3.697195e-01  0.4381630987 -0.1768647949  0.6499399276
#> 311 311 1  3.928879e-01  0.4980242112  0.2933230645  0.8818756177
#> 312 312 1  8.204651e-01  0.0283535206 -0.0580299978  0.5479276638
#> 313 313 1  3.423720e-01 -0.1113495697  0.1385591291 -0.1136242563
#> 314 314 1  1.686966e-01  0.6291240191  0.0378932756  0.7752245689
#> 315 315 1 -7.380947e-02  0.2310469570 -0.1105356116 -0.1079812729
#> 316 316 1  1.149272e-01 -0.1029326885 -0.2255796259 -0.4329936612
#> 317 317 1 -4.358445e-02  0.2189914840  0.1575338749 -0.0262510843
#> 318 318 1  4.935949e-01 -0.0373665893  0.3579001190  0.2109068522
#> 319 319 1  2.597061e-01  0.4503163724  0.3506304530  0.6852018937
#> 320 320 1 -1.139689e-01  0.4906434981  0.3796822907  0.3582534005
#> 321 321 1  5.220337e-01  0.4308099666  0.1727032234  0.8879765272
#> 322 322 1  2.141882e-01  0.2132452111 -0.0222348494  0.1916484430
#> 323 323 1 -1.740960e-01  0.4305567953 -0.1999111524  0.0593693025
#> 324 324 1 -4.821077e-01  0.2082607733 -0.1401819108 -0.5798083685
#> 325 325 1  1.694866e-01  0.7099152316  0.4214833074  0.9927034085
#> 326 326 1 -1.027638e-01  0.1586391922  0.1625002185 -0.1768492690
#> 327 327 1 -5.655258e-01  0.4557609419  0.6358028082 -0.1049247693
#> 328 328 1  9.790665e-02  0.1782892778 -0.1018859408 -0.0029046892
#> 329 329 1  4.279486e-04  0.2742177382 -0.3350259334 -0.0224271649
#> 330 330 1  4.106453e-01 -0.0341811310  0.6321125577  0.1974298124
#> 331 331 1  3.730994e-01  0.0233569512  0.2012135716  0.1342511931
#> 332 332 1  1.150959e-01  0.4143813148  0.0794816252  0.4109449622
#> 333 333 1  6.873407e-01  0.8323438359  0.1766323415  1.6583904050
#> 334 334 1  2.106039e-02  0.1752847832 -0.0774712788 -0.0822270212
#> 335 335 1  5.455894e-02  0.2952347840  0.5863807075  0.2985839905
#> 336 336 1  2.932836e-01  0.2697537485  0.4515554970  0.4785027323
#> 337 337 1  4.111194e-01  0.2029850607 -0.0518648348  0.3766285164
#> 338 338 1  3.653577e-02  0.5406769685  0.4139460064  0.5997755378
#> 339 339 1 -1.342176e-01  0.0097181504  0.5772546685 -0.3259475689
#> 340 340 1  1.862406e-01  0.5770358581  0.3797566644  0.8028940560
#> 341 341 1 -2.983725e-01  0.1986554971 -0.1174127534 -0.3945356032
#> 342 342 1  1.156769e-01  0.0928799452  0.1895992561 -0.0371079802
#> 343 343 1  6.642279e-02  0.2863682521  0.0956294865  0.1739733657
#> 344 344 1  2.905732e-01  0.7278269263  0.2798273738  1.1111531649
#> 345 345 1  1.198107e-01  0.4907415215  0.6542888929  0.6743037570
#> 346 346 1  7.365965e-01  0.4399130973 -0.0159454399  1.0800738550
#> 347 347 1  1.847986e-01  0.2899067660  0.3644151232  0.3719543285
#> 348 348 1  2.388492e-01  0.5851274968  0.1805777512  0.8200447551
#> 349 349 1  5.156732e-01  0.1799340079  0.4737388130  0.5855072841
#> 350 350 1  7.849427e-02  0.1532584765  0.6289284390  0.1241560221
#> 351 351 1 -1.251492e-01  0.6785719285  0.5281154948  0.6624843563
#> 352 352 1  2.747521e-01  0.4938268961  0.0312567869  0.6848689793
#> 353 353 1  4.917481e-02  0.2514880809 -0.1721472996  0.0364403764
#> 354 354 1  4.395689e-01  0.4741457381 -0.1553625746  0.7823595950
#> 355 355 1 -4.097488e-01  0.5430517444  0.0745569953  0.0469286805
#> 356 356 1 -2.003791e-01  0.4693210105  0.0705673059  0.1574359331
#> 357 357 1 -7.362283e-02  0.0517448428  0.1563751067 -0.3060905039
#> 358 358 1 -4.771330e-02  0.4131729959 -0.2585385488  0.1520697057
#> 359 359 1  3.054496e-01  0.5436167502 -0.1368407870  0.7485651216
#> 360 360 1  5.173940e-01  0.7742242299 -0.0589559961  1.3335200520
#> 361 361 1  2.578024e-01  0.1955780673  0.0520807645  0.2302321967
#> 362 362 1  6.961762e-01  0.7299318762  0.0407364253  1.4815948337
#> 363 363 1  1.255748e-01  0.6272006969  0.4804283828  0.8387000939
#> 364 364 1 -8.837952e-02  0.0718201517 -0.0031146685 -0.3321918127
#> 365 365 1  1.545675e-01  0.4682752903 -0.2017296965  0.4613998739
#> 366 366 1  4.204037e-01  0.4576212101 -0.0934169617  0.7533083795
#> 367 367 1  4.528129e-02  0.1291462220  0.6094376724  0.0484716761
#> 368 368 1  3.472242e-01  0.4404627600 -0.3677303531  0.5814066840
#> 369 369 1  2.067300e-01  0.8401975826  0.1366458507  1.1531048832
#> 370 370 1  8.467653e-02  0.2634891411  0.1827955285  0.1813355649
#> 371 371 1  3.648243e-01  0.3270650923 -0.1426011884  0.4887796810
#> 372 372 1  1.915172e-01 -0.0078349666  0.3058591765 -0.0770258737
#> 373 373 1  3.288709e-01  0.2324727531  0.1016237083  0.3723733518
#> 374 374 1 -2.793889e-01 -0.1129510508  0.2581659003 -0.7414528400
#> 375 375 1  4.435550e-01  0.2835568841  0.3284150523  0.6263175524
#> 376 376 1  1.886249e-01  0.0815063987 -0.1158706096 -0.0542099534
#> 377 377 1  1.379989e-01  0.0467073775  0.1400337193 -0.0945211635
#> 378 378 1 -2.398485e-01  0.5812333247  0.8433426258  0.4769047026
#> 379 379 1  3.039409e-01  0.2416069611  0.2773770064  0.4040224780
#> 380 380 1  1.125011e-01  0.3855414524  0.1392838718  0.3805743339
#> 381 381 1 -1.223074e-02  0.2900936836  0.2594932837  0.1379490536
#> 382 382 1  2.869468e-01  0.1432225327 -0.1865485929  0.1230859604
#> 383 383 1  1.671283e-01  0.1175510907  0.5162997391  0.1362496922
#> 384 384 1  3.594933e-01  0.9414190792  0.1987071611  1.4798953395
#> 385 385 1 -2.557190e-01  0.6613944726 -0.3078774918  0.2881343261
#> 386 386 1  5.942237e-01  0.1123004490  0.0780763855  0.4681467719
#> 387 387 1  1.297438e-01  0.3815647811  0.0497358714  0.3702407828
#> 388 388 1  1.830746e-01  0.4361927162 -0.1448454710  0.4582842856
#> 389 389 1  5.875569e-02  0.0555460900  0.4033977864 -0.0984518440
#> 390 390 1  9.622396e-02  0.2595521081 -0.3734653815  0.0471470645
#> 391 391 1 -1.459943e-01  0.1740313044 -0.1514577846 -0.2789422042
#> 392 392 1  3.303317e-01  0.5264092250  0.3274612075  0.8666004558
#> 393 393 1  2.175536e-01  0.5306469265 -0.3498390415  0.5828423267
#> 394 394 1  3.294659e-01  0.0935281515 -0.0894847289  0.1187955629
#> 395 395 1  2.424612e-01  0.0118708723 -0.0304258191 -0.0790553086
#> 396 396 1  3.114731e-01  0.1303025445  0.4645338295  0.2942877049
#> 397 397 1  5.611355e-01  0.1311594633  0.5118050267  0.5707790737
#> 398 398 1  1.178129e-01  0.3565796411  0.1949593346  0.3573181536
#> 399 399 1  9.327660e-02 -0.0595457203  0.4005230044 -0.2333476987
#> 400 400 1  4.390381e-01  0.3666724786  0.2302326331  0.7199324847
#> 401 401 1  7.130624e-01 -0.1516725826  0.3986534009  0.2832312412
#> 402 402 1  4.466609e-01  0.6413420772  0.3935367123  1.1763049302
#> 403 403 1 -1.269192e-02  0.6034871455 -0.0567479339  0.5220379551
#> 404 404 1  2.752621e-01 -0.2534000688 -0.1350092035 -0.4640350039
#> 405 405 1 -1.365232e-01  0.1783045580  0.2208597806 -0.1685609459
#> 406 406 1  3.492618e-01  0.4633454456  0.1741394907  0.7543640794
#> 407 407 1  3.065712e-01 -0.1954227482 -0.0453239815 -0.3224346668
#> 408 408 1  1.903053e-01  0.4175411697  0.4979334210  0.6006548408
#> 409 409 1  5.304455e-01 -0.1311934740  0.5281325670  0.1537165900
#> 410 410 1  1.872045e-01  0.2319953765  0.1961412605  0.2461513065
#> 411 411 1  4.980695e-02  0.0735321109  0.2209998472 -0.1273122891
#> 412 412 1  3.528165e-01  0.0919747777  0.0417016456  0.1742596785
#> 413 413 1  4.279175e-01  0.1381617289  0.1189098135  0.3414141310
#> 414 414 1  7.317643e-01  0.2744485279  0.0455778135  0.8452922230
#> 415 415 1 -1.756326e-01  0.8830036813  0.2203471780  0.8344753280
#> 416 416 1  1.166801e-01  0.4227622154  0.1484725463  0.4424661080
#> 417 417 1  1.957361e-01  0.5796510364  0.6150685596  0.8762312552
#> 418 418 1  1.011917e-01  0.0663491029  0.1912615917 -0.0912836223
#> 419 419 1 -1.977695e-01 -0.2529125506  0.5611415051 -0.7862714762
#> 420 420 1  9.746826e-02  0.0959329794  0.4015160213  0.0017526162
#> 421 421 1 -3.214024e-01  0.1930353485  0.2261306628 -0.3403599644
#> 422 422 1  1.736611e-01  0.9080417291 -0.1114669652  1.1560997418
#> 423 423 1  3.851199e-01  0.6454753282  0.0886385151  1.0405052883
#> 424 424 1 -4.359669e-01  0.3102468924  0.1274401816 -0.3123906534
#> 425 425 1  5.186057e-01  0.0827401360 -0.3746224952  0.2302269016
#> 426 426 1 -4.666322e-02  0.5490938162 -0.1237683170  0.3886678631
#> 427 427 1  2.578821e-02  0.2657570276 -0.6781056448 -0.0948968878
#> 428 428 1  2.206462e-02  0.5399258443 -0.2372165437  0.4188956724
#> 429 429 1  1.219450e-02  0.2761542964 -0.1314554101  0.0442802136
#> 430 430 1  1.236173e-01  0.4491541885 -0.0809540238  0.4309350976
#> 431 431 1  1.687802e-01  0.1948013950  0.3136905194  0.2012953360
#> 432 432 1  2.416619e-01 -0.3807195040 -0.2523906012 -0.7178177388
#> 433 433 1  7.803396e-01  0.4502255980 -0.3865251075  1.0478651900
#> 434 434 1 -1.298738e-01 -0.0063024266  0.3726929047 -0.3967897301
#> 435 435 1 -3.117900e-01  0.1167783594  0.1775155578 -0.4555228405
#> 436 436 1  1.766243e-01  0.2866183455  0.8232851649  0.4743869807
#> 437 437 1  3.309975e-01  0.3180386479  0.3522521516  0.5647473292
#> 438 438 1  1.680394e-01  0.5721510676 -0.3462356959  0.5930496887
#> 439 439 1  4.193224e-01  0.1405071641 -0.4698835404  0.1870767280
#> 440 440 1  1.213618e-02  0.2849162213  0.5683360721  0.2339960313
#> 441 441 1  4.093425e-01  0.0574828283  0.6917418019  0.3469720804
#> 442 442 1 -4.474079e-02  0.4711263534 -0.0658557745  0.2897732014
#> 443 443 1 -1.371197e-01  0.5205562720  0.3889963430  0.3805256132
#> 444 444 1  6.125653e-01  0.7325688275 -0.2239145105  1.3304719051
#> 445 445 1 -2.079514e-01  0.3251371817  0.0748429655 -0.0631586398
#> 446 446 1  9.264127e-02  0.0310530718  0.4224993364 -0.0941926115
#> 447 447 1  3.746687e-01  0.2492157896 -0.1538592282  0.3809392592
#> 448 448 1 -3.639989e-01  0.1671420946 -0.0422567644 -0.4914591041
#> 449 449 1  1.824022e-01 -0.2314133360  0.0680262530 -0.4780799914
#> 450 450 1 -3.995208e-02  0.1487176698 -0.3937762879 -0.2658508633
#> 451 451 1  4.527307e-01  0.0177212644 -0.0867334311  0.1371277586
#> 452 452 1 -2.141433e-01  0.3198116661  0.1447105319 -0.0599300836
#> 453 453 1  3.728422e-01  0.3887310416  0.2626524173  0.6910088525
#> 454 454 1  2.136245e-01  0.7653291366  0.0488765639  1.0272420695
#> 455 455 1  3.258931e-02  0.5225543584  0.4550340111  0.5791352449
#> 456 456 1 -1.054588e-01 -0.0504315444 -0.3473132187 -0.6183437353
#> 457 457 1  8.699577e-02  0.6805730189  0.6673148590  0.9243321023
#> 458 458 1 -1.335989e-01  0.4425799261 -0.0712832764  0.1523899981
#> 459 459 1  4.357849e-01  0.7204230025 -0.2745388881  1.1132595557
#> 460 460 1 -2.178262e-01 -0.0820594075  0.4577890144 -0.5803173891
#> 461 461 1 -1.701505e-01  0.1296571662  0.8145923380 -0.1261233740
#> 462 462 1  3.310799e-01  0.1852463081 -0.5323670633  0.1445422163
#> 463 463 1  2.508644e-01  0.2988992939 -0.0269165648  0.3560869058
#> 464 464 1 -9.329388e-02  0.5381943013  0.3635799654  0.4464612748
#> 465 465 1  1.909951e-01  0.0495809121  0.8028429998  0.1330729052
#> 466 466 1  1.716781e-01  0.3270710404 -0.0458280000  0.3095563551
#> 467 467 1  7.358717e-01 -0.0197406389  0.5082965300  0.5305156474
#> 468 468 1  8.566323e-01  0.2698454578  0.0027455290  0.9593275670
#> 469 469 1  2.864741e-01 -0.0339972060  0.5124597638  0.0365304038
#> 470 470 1  1.247599e-01  0.1153613636  0.4289623640  0.0662607044
#> 471 471 1  1.726196e-02  0.1991000530  0.4309344334  0.0775040224
#> 472 472 1  5.682435e-01  0.0008066331  0.4929635299  0.3803236459
#> 473 473 1  1.531874e-01  0.3596825599  0.0270554566  0.3568024016
#> 474 474 1 -4.730882e-05  0.3436277214  0.1865207742  0.2117023846
#> 475 475 1 -1.753777e-02 -0.1017986997  0.2117207520 -0.4605254668
#> 476 476 1  1.771616e-01  0.7882828663  0.1650005357  1.0521463093
#> 477 477 1 -9.880234e-02  0.1422557489 -0.6433395853 -0.4005361499
#> 478 478 1  1.942740e-01  0.4084490099  0.4497776053  0.5791989976
#> 479 479 1  1.976402e-02  0.4727201419  0.1221652737  0.4076588417
#> 480 480 1  3.694169e-01  0.2337804904 -0.1170710850  0.3618189210
#> 481 481 1  4.138108e-01  0.1714212159 -0.0468103779  0.3339639961
#> 482 482 1 -4.309998e-01  0.6539868975 -0.0630945452  0.1541559215
#> 483 483 1  2.624444e-01  0.3649091190  0.0286403057  0.4801646892
#> 484 484 1  9.074253e-02  0.4447470385 -0.3498419642  0.3218052490
#> 485 485 1  3.181938e-01  0.4020215656  0.0624662203  0.6025032604
#> 486 486 1  1.399420e-01  0.1969176939 -0.0667939867  0.0778971830
#> 487 487 1  6.138032e-02  0.1327954717  0.3369002405  0.0020047662
#> 488 488 1  3.283314e-01  0.0831273999  0.2268571581  0.1821035700
#> 489 489 1 -1.231343e-01  0.5523028955  0.1801123948  0.3895529033
#> 490 490 1 -2.079663e-01  0.7483108278  0.1641375620  0.5865544106
#> 491 491 1  1.919162e-01  0.2616394250  0.4219997696  0.3521140158
#> 492 492 1 -1.807571e-01 -0.0397304250  0.1903807411 -0.5460490018
#> 493 493 1  6.868113e-02  0.2820648200 -0.2610509877  0.0798670645
#> 494 494 1  3.176367e-01  0.6128492194 -0.5418361459  0.7617078305
#> 495 495 1  4.306172e-01  0.0297871678  0.4793735416  0.2747088353
#> 496 496 1  4.416725e-01  0.3748481236 -0.0638774257  0.6605250062
#> 497 497 1  7.631777e-01  0.3743219888 -0.1787372487  0.9697678921
#> 498 498 1  2.959246e-02  0.4124494531 -0.1948026666  0.2486212990
#> 499 499 1  3.357627e-01  0.3738424842  0.7105916726  0.7430061987
#> 500 500 1  4.965347e-01  0.4876239512  0.0775310435  0.9212449621