Get estimates, inquiries, assignment vectors, or samples from a design given data

Get estimates, inquiries, assignment vectors, or samples from a design given data

Usage

get_estimates(design, data = NULL, start = 1, end = length(design))

Arguments

design

A design object, typically created using the + operator

data

A data.frame object with sufficient information to get the data, estimates, inquiries, an assignment vector, or a sample.

start

(Defaults to 1) a scalar indicating which step in the design to begin with. By default all data steps are drawn, from step 1 to the last step of the design.

end

(Defaults to length(design)) a scalar indicating which step in the design to finish with.

Examples

design <- 
  declare_model(
    N = 100, 
    U = rnorm(N),
    potential_outcomes(Y ~ Z + U)
  ) +
  declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(S = complete_rs(N, n = 75)) +
  declare_assignment(Z = complete_ra(N, m = 50)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z)) +
  declare_estimator(Y ~ Z, inquiry = "ATE")

dat <- draw_data(design)

draw_data(design, data = dat, start = 2)
#>     ID           U       Y_Z_0        Y_Z_1 S Z            Y
#> 1  001 -0.57704855 -0.57704855  0.422951452 1 1  0.422951452
#> 2  002 -2.27391544 -2.27391544 -1.273915435 1 0 -2.273915435
#> 3  003  0.26557439  0.26557439  1.265574392 1 0  0.265574392
#> 4  004  0.56876808  0.56876808  1.568768076 1 0  0.568768076
#> 5  005 -0.02777070 -0.02777070  0.972229303 1 0 -0.027770697
#> 6  006 -0.93002894 -0.93002894  0.069971056 1 1  0.069971056
#> 7  007 -0.72258585 -0.72258585  0.277414151 1 0 -0.722585849
#> 8  010  0.43583452  0.43583452  1.435834519 1 1  1.435834519
#> 9  011 -2.21110858 -2.21110858 -1.211108578 1 1 -1.211108578
#> 10 012  0.08618844  0.08618844  1.086188444 1 0  0.086188444
#> 11 013 -0.92277875 -0.92277875  0.077221252 1 0 -0.922778748
#> 12 014  0.43912543  0.43912543  1.439125430 1 0  0.439125430
#> 13 015 -1.10465660 -1.10465660 -0.104656602 1 1 -0.104656602
#> 14 016  0.21963259  0.21963259  1.219632595 1 1  1.219632595
#> 15 017 -0.20348822 -0.20348822  0.796511783 1 0 -0.203488217
#> 16 018  1.49113595  1.49113595  2.491135951 1 1  2.491135951
#> 17 021  0.56836162  0.56836162  1.568361619 1 1  1.568361619
#> 18 022  0.81955231  0.81955231  1.819552307 1 1  1.819552307
#> 19 023 -0.92702736 -0.92702736  0.072972638 1 1  0.072972638
#> 20 024  0.85399758  0.85399758  1.853997576 1 1  1.853997576
#> 21 026 -1.49716863 -1.49716863 -0.497168632 1 1 -0.497168632
#> 22 028  1.11978975  1.11978975  2.119789754 1 1  2.119789754
#> 23 030 -1.15565907 -1.15565907 -0.155659068 1 0 -1.155659068
#> 24 031 -1.25209514 -1.25209514 -0.252095135 1 1 -0.252095135
#> 25 033 -0.42200577 -0.42200577  0.577994232 1 1  0.577994232
#> 26 034  0.14822676  0.14822676  1.148226763 1 1  1.148226763
#> 27 035 -0.25601631 -0.25601631  0.743983686 1 1  0.743983686
#> 28 037 -0.19213392 -0.19213392  0.807866081 1 1  0.807866081
#> 29 038 -0.39260506 -0.39260506  0.607394936 1 1  0.607394936
#> 30 040 -1.18633492 -1.18633492 -0.186334921 1 1 -0.186334921
#> 31 041  0.05440519  0.05440519  1.054405185 1 0  0.054405185
#> 32 046  1.09896282  1.09896282  2.098962820 1 0  1.098962820
#> 33 048  0.38331225  0.38331225  1.383312248 1 1  1.383312248
#> 34 049 -0.30556671 -0.30556671  0.694433290 1 1  0.694433290
#> 35 050 -1.91905120 -1.91905120 -0.919051200 1 0 -1.919051200
#> 36 051 -0.82708076 -0.82708076  0.172919237 1 1  0.172919237
#> 37 052 -2.83774792 -2.83774792 -1.837747923 1 0 -2.837747923
#> 38 053 -0.45812541 -0.45812541  0.541874591 1 1  0.541874591
#> 39 054 -0.23656473 -0.23656473  0.763435267 1 1  0.763435267
#> 40 055 -0.47489619 -0.47489619  0.525103812 1 1  0.525103812
#> 41 056 -1.59925207 -1.59925207 -0.599252065 1 0 -1.599252065
#> 42 057  1.23301800  1.23301800  2.233017996 1 0  1.233017996
#> 43 058 -1.31145503 -1.31145503 -0.311455031 1 1 -0.311455031
#> 44 059 -0.28150401 -0.28150401  0.718495986 1 1  0.718495986
#> 45 061 -1.32801407 -1.32801407 -0.328014072 1 1 -0.328014072
#> 46 065  1.91169874  1.91169874  2.911698737 1 1  2.911698737
#> 47 066 -0.46272244 -0.46272244  0.537277562 1 1  0.537277562
#> 48 067  1.59170374  1.59170374  2.591703743 1 0  1.591703743
#> 49 068 -0.03085492 -0.03085492  0.969145085 1 0 -0.030854915
#> 50 069  0.99085973  0.99085973  1.990859734 1 0  0.990859734
#> 51 070 -0.45462432 -0.45462432  0.545375682 1 1  0.545375682
#> 52 072  0.15029111  0.15029111  1.150291108 1 1  1.150291108
#> 53 073  0.44919300  0.44919300  1.449193005 1 1  1.449193005
#> 54 074 -0.59970476 -0.59970476  0.400295237 1 0 -0.599704763
#> 55 075  0.15699357  0.15699357  1.156993569 1 1  1.156993569
#> 56 076  0.55373718  0.55373718  1.553737181 1 1  1.553737181
#> 57 077 -0.83331435 -0.83331435  0.166685648 1 1  0.166685648
#> 58 078  0.22747511  0.22747511  1.227475111 1 0  0.227475111
#> 59 080  0.48052695  0.48052695  1.480526948 1 1  1.480526948
#> 60 081  0.40552456  0.40552456  1.405524558 1 1  1.405524558
#> 61 082 -1.75901020 -1.75901020 -0.759010196 1 1 -0.759010196
#> 62 085  1.33588827  1.33588827  2.335888268 1 1  2.335888268
#> 63 086 -0.16192815 -0.16192815  0.838071848 1 0 -0.161928152
#> 64 087 -1.00415356 -1.00415356 -0.004153563 1 1 -0.004153563
#> 65 088  0.15128386  0.15128386  1.151283862 1 1  1.151283862
#> 66 089 -1.13842252 -1.13842252 -0.138422523 1 0 -1.138422523
#> 67 090  0.41514191  0.41514191  1.415141909 1 1  1.415141909
#> 68 091 -0.50245349 -0.50245349  0.497546508 1 0 -0.502453492
#> 69 092 -0.55498346 -0.55498346  0.445016542 1 1  0.445016542
#> 70 093  0.25865149  0.25865149  1.258651492 1 1  1.258651492
#> 71 094  0.50782350  0.50782350  1.507823498 1 0  0.507823498
#> 72 095  0.51095102  0.51095102  1.510951025 1 1  1.510951025
#> 73 096  0.58407222  0.58407222  1.584072223 1 1  1.584072223
#> 74 099  0.59539915  0.59539915  1.595399146 1 1  1.595399146
#> 75 100  1.25383101  1.25383101  2.253831006 1 1  2.253831006

get_estimates(design, data = dat)
#>   estimator term  estimate std.error statistic     p.value  conf.low conf.high
#> 1 estimator    Z 0.7286061 0.2196781  3.316698 0.001421732 0.2907882  1.166424
#>   df outcome inquiry
#> 1 73       Y     ATE