Modify a design after the fact

Insert, delete and replace steps in an (already declared) design object.

Usage

insert_step(design, new_step, before, after)

delete_step(design, step)

replace_step(design, step, new_step)

Arguments

design

A design object, usually created using the + operator, expand_design, or the design library.

new_step

The new step; Either a function or a partial call.

before

The step before which to add steps.

after

The step after which to add steps.

step

The quoted label of the step to be deleted or replaced.

Value

A new design object.

Details

See modify_design for details.

Examples

 my_model <- 
   declare_model(
     N = 100, 
     U = rnorm(N),
     Y_Z_0 = U,
     Y_Z_1 = U + rnorm(N, mean = 2, sd = 2)
   )

 my_assignment <- declare_assignment(Z = complete_ra(N, m = 50))
 my_assignment_2 <- declare_assignment(Z = complete_ra(N, m = 25))

 design <- my_model + my_assignment

 draw_data(design)
#>      ID            U        Y_Z_0       Y_Z_1 Z
#> 1   001  1.037868824  1.037868824  3.01722927 0
#> 2   002  1.015994429  1.015994429  0.58184948 1
#> 3   003  0.685955048  0.685955048  0.51420749 0
#> 4   004  0.112720713  0.112720713  1.49766649 1
#> 5   005  1.175147894  1.175147894 -0.04675659 1
#> 6   006 -0.712876606 -0.712876606  0.94958759 1
#> 7   007 -2.092857638 -2.092857638 -0.50633244 0
#> 8   008 -0.140807201 -0.140807201 -4.47407250 0
#> 9   009  0.188072446  0.188072446  1.16876408 1
#> 10  010  0.534262285  0.534262285  2.43728123 1
#> 11  011  0.480013842  0.480013842  0.20517257 0
#> 12  012 -0.183191184 -0.183191184  1.40243401 1
#> 13  013 -0.749407096 -0.749407096  2.00060875 1
#> 14  014 -0.127507552 -0.127507552  2.90913533 0
#> 15  015 -1.266666546 -1.266666546  1.48678518 1
#> 16  016 -0.167466940 -0.167466940  0.62192933 0
#> 17  017  0.271912762  0.271912762  1.55365763 0
#> 18  018 -1.018993416 -1.018993416  1.17854062 1
#> 19  019 -1.973392010 -1.973392010 -1.29578967 0
#> 20  020  0.732424783  0.732424783  3.62734432 0
#> 21  021 -2.635011391 -2.635011391  0.12865392 1
#> 22  022 -1.453193158 -1.453193158  2.74888001 0
#> 23  023 -0.740666123 -0.740666123  0.80306687 0
#> 24  024  0.453568269  0.453568269  5.14690635 1
#> 25  025  0.738249848  0.738249848  0.86761331 0
#> 26  026  1.638179412  1.638179412  2.33896960 1
#> 27  027 -0.095068884 -0.095068884  4.41762246 0
#> 28  028  1.768563623  1.768563623  4.38837484 1
#> 29  029 -0.483929843 -0.483929843  1.90931722 1
#> 30  030 -0.602616297 -0.602616297  0.46815390 1
#> 31  031 -1.190753215 -1.190753215  0.97584730 0
#> 32  032 -0.041666609 -0.041666609 -0.98168773 1
#> 33  033  0.257347918  0.257347918  2.30147770 0
#> 34  034  0.552815410  0.552815410  3.26959318 1
#> 35  035 -0.145450650 -0.145450650  0.37711813 0
#> 36  036 -0.060706933 -0.060706933  3.05673230 1
#> 37  037  0.908305990  0.908305990  3.99043028 1
#> 38  038 -1.461837797 -1.461837797 -3.14480573 0
#> 39  039 -1.636829991 -1.636829991  0.07413458 1
#> 40  040  0.064375487  0.064375487  1.04795485 0
#> 41  041  0.079359593  0.079359593  5.82109706 1
#> 42  042 -0.037837731 -0.037837731  1.01914085 1
#> 43  043 -0.690664706 -0.690664706 -0.04390635 0
#> 44  044 -0.946703933 -0.946703933  1.02503443 0
#> 45  045  1.901153942  1.901153942  4.18406342 0
#> 46  046  2.256395357  2.256395357  5.18963046 1
#> 47  047  0.063915415  0.063915415  3.58184313 1
#> 48  048 -0.389367580 -0.389367580  2.51285688 1
#> 49  049 -1.394987376 -1.394987376  1.61624501 1
#> 50  050 -0.136713753 -0.136713753  0.60750080 0
#> 51  051  0.697342843  0.697342843  4.80154628 0
#> 52  052  0.022669013  0.022669013  1.72373271 0
#> 53  053  0.925237255  0.925237255  4.01715881 0
#> 54  054 -1.673420084 -1.673420084  2.21167707 1
#> 55  055 -0.672937567 -0.672937567  0.73545523 1
#> 56  056  0.860711442  0.860711442  1.65658543 0
#> 57  057  2.471983313  2.471983313  3.36499520 1
#> 58  058 -0.679209006 -0.679209006  1.00059890 0
#> 59  059 -0.000276982 -0.000276982  2.98993754 1
#> 60  060 -0.423235054 -0.423235054  1.37668680 1
#> 61  061 -1.922007841 -1.922007841  1.84794244 0
#> 62  062 -0.708213279 -0.708213279  0.47875341 0
#> 63  063 -0.340557312 -0.340557312  1.70472188 0
#> 64  064  0.021225363  0.021225363  2.28508079 1
#> 65  065  1.066551676  1.066551676  5.51065270 1
#> 66  066  0.293139221  0.293139221  1.94184002 0
#> 67  067 -0.578957621 -0.578957621  5.26485577 0
#> 68  068  0.067486646  0.067486646  0.37607680 0
#> 69  069  0.311464430  0.311464430  0.35840704 1
#> 70  070 -0.390659732 -0.390659732  1.02690469 0
#> 71  071 -0.953954057 -0.953954057  1.59519471 0
#> 72  072 -0.725694296 -0.725694296  0.49341980 1
#> 73  073 -1.911462361 -1.911462361  2.56703031 0
#> 74  074 -0.933655455 -0.933655455  1.98023167 0
#> 75  075 -0.925410850 -0.925410850  1.96059111 0
#> 76  076  1.067547104  1.067547104  3.75371550 1
#> 77  077  0.259159560  0.259159560  1.91640392 1
#> 78  078  2.387389812  2.387389812  3.80050080 0
#> 79  079 -0.362677473 -0.362677473 -1.41597663 0
#> 80  080 -0.828077981 -0.828077981  5.67425691 0
#> 81  081 -1.786271304 -1.786271304 -1.11169913 0
#> 82  082  0.865890109  0.865890109  3.22110424 0
#> 83  083 -0.883899502 -0.883899502  0.88851412 0
#> 84  084 -1.104627213 -1.104627213  0.48913583 1
#> 85  085  0.197847125  0.197847125  3.46061174 1
#> 86  086  0.217717990  0.217717990  5.32645716 1
#> 87  087  1.020448559  1.020448559  4.75008315 1
#> 88  088  0.148978091  0.148978091  1.99054649 1
#> 89  089  0.582329946  0.582329946  2.25613157 1
#> 90  090  0.448799009  0.448799009  3.60117213 1
#> 91  091  0.558902695  0.558902695  4.85762317 1
#> 92  092 -1.197371925 -1.197371925 -0.09819455 0
#> 93  093 -0.718111718 -0.718111718  0.69726146 0
#> 94  094 -1.929727034 -1.929727034  0.38076985 0
#> 95  095 -0.564578202 -0.564578202  1.86170837 1
#> 96  096 -1.593819278 -1.593819278  4.34053858 1
#> 97  097 -1.175673443 -1.175673443  2.30866225 0
#> 98  098  0.463448044  0.463448044  5.06878049 0
#> 99  099  0.545787888  0.545787888  0.52243534 1
#> 100 100  1.867782053  1.867782053  2.86035268 1
 
 design_modified <- replace_step(design, 2, my_assignment_2)
 
 draw_data(design)
#>      ID            U        Y_Z_0       Y_Z_1 Z
#> 1   001  0.277268441  0.277268441  2.01031511 0
#> 2   002  3.015891824  3.015891824  3.67059008 0
#> 3   003  2.241913683  2.241913683  4.13535473 0
#> 4   004  0.108142562  0.108142562  4.21220760 0
#> 5   005 -0.409574286 -0.409574286 -2.44982188 0
#> 6   006 -0.244110913 -0.244110913 -3.46008034 1
#> 7   007  0.923593236  0.923593236  1.64462990 1
#> 8   008  0.799109003  0.799109003  3.58846042 0
#> 9   009  0.853937812  0.853937812  0.79356917 0
#> 10  010  0.583052619  0.583052619  0.74434198 1
#> 11  011  1.578879463  1.578879463  2.60693964 1
#> 12  012  1.167401122  1.167401122  4.95617841 1
#> 13  013 -1.349895464 -1.349895464 -1.54691036 1
#> 14  014  0.802900871  0.802900871  2.99137366 1
#> 15  015  0.367451082  0.367451082  2.82121464 0
#> 16  016 -0.167827729 -0.167827729 -1.55537110 1
#> 17  017  0.973131047  0.973131047 -1.29026518 1
#> 18  018 -0.790943535 -0.790943535  2.68685113 1
#> 19  019 -0.744488200 -0.744488200  1.29485806 0
#> 20  020 -0.413587424 -0.413587424  1.88076532 1
#> 21  021 -0.445790353 -0.445790353  0.54408286 0
#> 22  022  0.816685815  0.816685815  2.40259424 0
#> 23  023 -0.235805270 -0.235805270  3.27132665 0
#> 24  024  0.127008590  0.127008590  3.33266272 1
#> 25  025  0.474081966  0.474081966  0.16245483 1
#> 26  026 -0.018332187 -0.018332187 -1.22766196 0
#> 27  027 -0.557966088 -0.557966088 -0.42443821 1
#> 28  028 -0.494542899 -0.494542899  1.15877686 1
#> 29  029 -0.488132952 -0.488132952 -3.41451305 1
#> 30  030 -1.182635200 -1.182635200 -0.01538937 0
#> 31  031 -3.363727629 -3.363727629 -3.58917494 1
#> 32  032 -0.287739673 -0.287739673  0.08755827 1
#> 33  033 -0.781989048 -0.781989048 -0.10460866 0
#> 34  034 -0.984059918 -0.984059918 -0.47333211 1
#> 35  035  0.470621723  0.470621723  3.21320205 0
#> 36  036  0.295226710  0.295226710  4.83718719 1
#> 37  037  0.518554626  0.518554626  5.30738150 0
#> 38  038 -0.430681901 -0.430681901  0.16413457 0
#> 39  039 -0.583756476 -0.583756476  3.80417666 0
#> 40  040 -1.283392555 -1.283392555  1.66968870 1
#> 41  041  1.388428253  1.388428253  6.59570645 0
#> 42  042 -0.467248671 -0.467248671 -1.95115363 0
#> 43  043 -1.535462041 -1.535462041 -0.52517667 0
#> 44  044  2.090159632  2.090159632  2.88539097 0
#> 45  045 -1.082768757 -1.082768757  0.74115762 0
#> 46  046 -0.036225682 -0.036225682  4.43676374 1
#> 47  047  2.250625558  2.250625558  2.87241921 0
#> 48  048 -0.218111704 -0.218111704  6.52501635 1
#> 49  049 -0.087869721 -0.087869721  0.42252686 0
#> 50  050  0.149421132  0.149421132  2.64727412 0
#> 51  051 -0.229611393 -0.229611393  4.37043516 1
#> 52  052  0.438474330  0.438474330  4.82289807 0
#> 53  053  1.028289212  1.028289212  3.57481826 0
#> 54  054  0.226235415  0.226235415  4.14579863 0
#> 55  055  0.736949166  0.736949166  4.92498817 0
#> 56  056 -0.909808086 -0.909808086  0.29225696 1
#> 57  057  0.900972265  0.900972265  4.96792423 1
#> 58  058 -0.372749840 -0.372749840  1.84793518 0
#> 59  059 -0.324720129 -0.324720129  1.11656553 0
#> 60  060  0.345174539  0.345174539  5.78982454 0
#> 61  061  0.429134426  0.429134426  2.49544507 1
#> 62  062 -1.237367106 -1.237367106 -1.72922590 0
#> 63  063  0.005268394  0.005268394  2.80045126 0
#> 64  064  0.565053560  0.565053560  2.62407425 1
#> 65  065  1.049768899  1.049768899  5.70114541 0
#> 66  066  1.019773758  1.019773758  3.10536432 0
#> 67  067  0.470169767  0.470169767  2.76781381 1
#> 68  068  1.226629516  1.226629516  3.73208428 0
#> 69  069 -0.213977259 -0.213977259  5.08821471 1
#> 70  070 -0.012738742 -0.012738742  3.94727740 1
#> 71  071  0.205011410  0.205011410  0.40427544 1
#> 72  072  1.092461453  1.092461453  5.12009828 1
#> 73  073 -2.060152303 -2.060152303  1.11676249 0
#> 74  074 -1.588494613 -1.588494613 -0.45597125 1
#> 75  075  0.387298712  0.387298712  3.04808232 0
#> 76  076 -0.287286731 -0.287286731  0.71840115 1
#> 77  077 -1.238301018 -1.238301018  1.60366905 1
#> 78  078 -0.958731661 -0.958731661  1.73496463 1
#> 79  079  1.052176289  1.052176289  1.12328371 1
#> 80  080  1.681373180  1.681373180  4.49961040 1
#> 81  081 -0.969598332 -0.969598332  1.45790690 0
#> 82  082 -0.831289176 -0.831289176  0.17274257 0
#> 83  083  0.072526594  0.072526594  2.00177928 0
#> 84  084  0.656158303  0.656158303  5.44436932 0
#> 85  085 -0.703636666 -0.703636666  2.46111040 0
#> 86  086 -0.481089588 -0.481089588  2.52574743 1
#> 87  087 -0.737538712 -0.737538712 -0.87010163 0
#> 88  088 -0.374721161 -0.374721161  4.49188414 1
#> 89  089 -0.445887087 -0.445887087  1.56847629 0
#> 90  090 -1.782968374 -1.782968374 -0.01548514 1
#> 91  091 -0.502382948 -0.502382948  1.37553905 0
#> 92  092 -0.710519852 -0.710519852  1.58645682 1
#> 93  093 -1.763942162 -1.763942162 -4.30061087 1
#> 94  094  1.104806593  1.104806593  4.38095574 1
#> 95  095  1.040650810  1.040650810  5.15169836 1
#> 96  096 -0.895746932 -0.895746932  0.94105347 1
#> 97  097 -0.956000908 -0.956000908  0.39043785 1
#> 98  098 -0.929110975 -0.929110975  2.72054299 1
#> 99  099 -0.083551962 -0.083551962  2.33239638 0
#> 100 100  0.324355931  0.324355931  4.54022663 1
 
 if (FALSE) { # \dontrun{
 
 design <- 
   declare_model(
     N = 100,
     U = rnorm(N),
     potential_outcomes(Y ~ 0.20 * Z + U)
   ) + 
     declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) + 
     declare_assignment(Z = complete_ra(N, m = N/2)) + 
     declare_measurement(Y = reveal_outcomes(Y ~ Z)) + 
     declare_estimator(Y ~ Z, inquiry = "ATE")
 
 insert_step(design, declare_sampling(S = complete_rs(N, n = 50)),
             after = 1)

 # If you are using a design created by a designer, for example from
 # the DesignLibrary package, you will not have access to the step
 # objects. Instead, you can always use the label of the step.
 
 design <- DesignLibrary::two_arm_designer()
 
 # get the labels for the steps
 names(design)
 
 insert_step(design, 
   declare_sampling(S = complete_rs(N, n = 50)),
   after = "potential_outcomes")
   
 } # }

design <- 
  declare_model(
    N = 100,
    U = rnorm(N),
    potential_outcomes(Y ~ 0.20 * Z + U)
  ) + 
    declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) + 
    declare_assignment(Z = complete_ra(N, m = N/2)) + 
    declare_measurement(Y = reveal_outcomes(Y ~ Z)) + 
    declare_estimator(Y ~ Z, inquiry = "ATE")
delete_step(design, step = 5)
#> 
#> Research design declaration summary
#> 
#> Step 1 (model): declare_model(N = 100, U = rnorm(N), potential_outcomes(Y ~ 0.2 * Z + U)) 
#> 
#> Step 2 (inquiry): declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) -------------------
#> 
#> Step 3 (assignment): declare_assignment(Z = complete_ra(N, m = N/2)) -----------
#> 
#> Step 4 (measurement): declare_measurement(Y = reveal_outcomes(Y ~ Z)) ----------
#> 
#> Run of the design:
#> 
#>  inquiry estimand
#>      ATE      0.2
#> 
#> No modifiable parameters saved in design 

design <- 
  declare_model(
    N = 100,
    U = rnorm(N),
    potential_outcomes(Y ~ 0.20 * Z + U)
  ) + 
    declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) + 
    declare_assignment(Z = complete_ra(N, m = N/2)) + 
    declare_measurement(Y = reveal_outcomes(Y ~ Z)) + 
    declare_estimator(Y ~ Z, inquiry = "ATE")

replace_step(
  design, 
  step = 3, 
  new_step = declare_assignment(Z = simple_ra(N, prob = 0.5)))
#> 
#> Research design declaration summary
#> 
#> Step 1 (model): declare_model(N = 100, U = rnorm(N), potential_outcomes(Y ~ 0.2 * Z + U)) 
#> 
#> Step 2 (inquiry): declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) -------------------
#> 
#> Step 3 (assignment): declare_assignment(Z = simple_ra(N, prob = 0.5)) ----------
#> 
#> Step 4 (measurement): declare_measurement(Y = reveal_outcomes(Y ~ Z)) ----------
#> 
#> Step 5 (estimator): declare_estimator(Y ~ Z, inquiry = "ATE") ------------------
#> 
#> Run of the design:
#> 
#>  inquiry estimand estimator term estimate std.error statistic p.value conf.low
#>      ATE      0.2 estimator    Z    0.355     0.193      1.84  0.0684  -0.0274
#>  conf.high df outcome
#>      0.737 98       Y
#> 
#> No modifiable parameters saved in design