DeclareDesign package

The four main types of functions are to declare a step, to combine steps into designs, and to manipulate designs and designers (functions that return designs).

Design Steps

declare_model

Model step

declare_inquiry

Inquiry step

declare_sampling

Data strategy step (sampling)

declare_assignment

Data strategy step (assignment)

declare_measurement

Data strategy step (measurement)

declare_estimator

Answer strategy step (Estimator)

declare_test

Answer strategy step (Testing function)

Design Objects

+

Add steps to create a design

redesign

Change design parameters

draw_data

Draw a simulated dataset

run_design

Draw one set of inquiry values and estimates

diagnose_design

Diagnose a design

cite_design

Cite a design

Design Editing

modify_design

Add, delete or replace a step

redesign

Modify local variables within a design (advanced)

Designers

expand_design

Generate designs from a designer

designs

See also the DesignLibrary package for designers to use

See also

Author

Maintainer: Graeme Blair graeme.blair@gmail.com (ORCID)

Authors:

Jasper Cooper jaspercooper@gmail.com (ORCID)

Alexander Coppock acoppock@gmail.com (ORCID)

Macartan Humphreys macartan@gmail.com (ORCID)

Neal Fultz nfultz@gmail.com