Declare a design

Declare a design

# S3 method for dd
+(lhs, rhs)

Arguments

lhs
A step in a research design, beginning with a function that draws the population. Steps are evaluated sequentially. With the exception of the first step, all steps must be functions that take a data.frame as an argument and return a data.frame. Typically, many steps are declared using the declare_ functions, i.e., declare_population, declare_population, declare_sampling, declare_potential_outcomes, declare_estimand, declare_assignment, and declare_estimator.

rhs
A second step in a research design

Value

a list of two functions, the design_function and the data_function. The design_function runs the design once, i.e. draws the data and calculates any estimates and estimands defined in ..., returned separately as two data.frame's. The data_function runs the design once also, but only returns the final data.

Details

Users can supply three kinds of functions to create a design:

1. Data generating functions. These include population, assignment, and sampling functions.

2. Estimand functions.

3. Estimator functions.

The location of the estimand and estimator functions in the pipeline of functions determine *when* the values of the estimand and estimator are calculated. This allows users to, for example, differentiate between a population average treatment effect and a sample average treatment effect by placing the estimand function before or after sampling.

Design objects declared with the + operator can be investigated with a series of post-declaration commands, such as draw_data, draw_estimands, draw_estimates, and diagnose_design.

The print and summary methods for a design object return some helpful descriptions of the steps in your research design. If randomizr functions are used for any assignment or sampling steps, additional details about those steps are provided.

Examples

my_population <- declare_population(N = 500, noise = rnorm(N))

my_potential_outcomes <- declare_potential_outcomes(Y ~ Z + noise)

my_sampling <- declare_sampling(n = 250)

my_assignment <- declare_assignment(m = 25)

my_estimand <- declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0))

my_estimator <- declare_estimator(Y ~ Z, estimand = my_estimand)

my_mutate <- declare_step(dplyr::mutate, noise_sq = noise^2)

my_reveal <- reveal_outcomes()

design <- my_population + my_potential_outcomes + my_sampling +
         my_estimand + my_mutate +
         my_assignment + my_reveal + my_estimator

design

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 500, noise = rnorm(N)) -------------
#> 
#> N = 500 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      500 character
#> 
#> Added variable: noise 
#>    min median  mean  max sd N_missing N_unique
#>  -2.87  -0.06 -0.03 3.13  1         0      500
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y ~ Z + noise) ---------
#> 
#> Formula: Y ~ Z + noise 
#> 
#> Added variable: Y_Z_0 
#>    min median  mean  max sd N_missing N_unique
#>  -2.87  -0.06 -0.03 3.13  1         0      500
#> 
#> Added variable: Y_Z_1 
#>    min median mean  max sd N_missing N_unique
#>  -1.87   0.94 0.97 4.13  1         0      500
#> 
#> Step 3 (sampling): declare_sampling(n = 250) -----------------------------------
#> 
#> N = 250 (250 subtracted) 
#> 
#> Added variable: S_inclusion_prob 
#>   0.5
#>   250
#>  1.00
#> 
#> Altered variable: ID 
#>   Before: 
#>  N_missing N_unique     class
#>          0      500 character
#> 
#>   After:
#>  N_missing N_unique     class
#>          0      250 character
#> 
#> Altered variable: noise 
#>   Before: 
#>    min median  mean  max sd N_missing N_unique
#>  -2.87  -0.06 -0.03 3.13  1         0      500
#> 
#>   After:
#>    min median mean  max   sd N_missing N_unique
#>  -2.87  -0.04    0 3.13 1.03         0      250
#> 
#> Altered variable: Y_Z_0 
#>   Before: 
#>    min median  mean  max sd N_missing N_unique
#>  -2.87  -0.06 -0.03 3.13  1         0      500
#> 
#>   After:
#>    min median mean  max   sd N_missing N_unique
#>  -2.87  -0.04    0 3.13 1.03         0      250
#> 
#> Altered variable: Y_Z_1 
#>   Before: 
#>    min median mean  max sd N_missing N_unique
#>  -1.87   0.94 0.97 4.13  1         0      500
#> 
#>   After:
#>    min median mean  max   sd N_missing N_unique
#>  -1.87   0.96    1 4.13 1.03         0      250
#> 
#> Step 4 (estimand): declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0)) -----------------
#> 
#> A single draw of the estimand:
#>  estimand_label estimand
#>             ATE        1
#> 
#> Step 5 (custom): declare_step(dplyr::mutate, noise_sq = noise^2) ---------------
#> 
#> Added variable: noise_sq 
#>  min median mean max   sd N_missing N_unique
#>    0   0.58 1.06 9.8 1.37         0      250
#> 
#> Step 6 (assignment): declare_assignment(m = 25) --------------------------------
#> 
#> Added variable: Z 
#>     0    1
#>   225   25
#>  0.90 0.10
#> 
#> Added variable: Z_cond_prob 
#>   0.1  0.9
#>    25  225
#>  0.10 0.90
#> 
#> Step 7 (reveal): reveal_outcomes() ---------------------------------------------
#> 
#> Added variable: Y 
#>    min median mean  max   sd N_missing N_unique
#>  -2.87   0.09  0.1 3.13 1.08         0      250
#> 
#> Step 8 (estimator): declare_estimator(Y ~ Z, estimand = my_estimand) -----------
#> 
#> Formula: Y ~ Z 
#> 
#> A single draw of the estimator:
#>  estimator_label term estimate std.error statistic      p.value  conf.low
#>        estimator    Z 1.080911 0.1998636  5.408245 6.629594e-06 0.6733124
#>  conf.high       df outcome estimand_label
#>    1.48851 31.04822       Y            ATE
#> 

df <- draw_data(design)

estimates <- draw_estimates(design)
estimands <- draw_estimands(design)

# You can add steps to a design

design <- my_population + my_potential_outcomes
design + my_sampling

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 500, noise = rnorm(N)) -------------
#> 
#> N = 500 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      500 character
#> 
#> Added variable: noise 
#>    min median  mean  max   sd N_missing N_unique
#>  -3.34   0.03 -0.03 3.14 1.06         0      500
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y ~ Z + noise) ---------
#> 
#> Formula: Y ~ Z + noise 
#> 
#> Added variable: Y_Z_0 
#>    min median  mean  max   sd N_missing N_unique
#>  -3.34   0.03 -0.03 3.14 1.06         0      500
#> 
#> Added variable: Y_Z_1 
#>    min median mean  max   sd N_missing N_unique
#>  -2.34   1.03 0.97 4.14 1.06         0      500
#> 
#> Step 3 (sampling): declare_sampling(n = 250) -----------------------------------
#> 
#> N = 250 (250 subtracted) 
#> 
#> Added variable: S_inclusion_prob 
#>   0.5
#>   250
#>  1.00
#> 
#> Altered variable: ID 
#>   Before: 
#>  N_missing N_unique     class
#>          0      500 character
#> 
#>   After:
#>  N_missing N_unique     class
#>          0      250 character
#> 
#> Altered variable: noise 
#>   Before: 
#>    min median  mean  max   sd N_missing N_unique
#>  -3.34   0.03 -0.03 3.14 1.06         0      500
#> 
#>   After:
#>    min median  mean  max   sd N_missing N_unique
#>  -3.34  -0.01 -0.08 3.14 1.13         0      250
#> 
#> Altered variable: Y_Z_0 
#>   Before: 
#>    min median  mean  max   sd N_missing N_unique
#>  -3.34   0.03 -0.03 3.14 1.06         0      500
#> 
#>   After:
#>    min median  mean  max   sd N_missing N_unique
#>  -3.34  -0.01 -0.08 3.14 1.13         0      250
#> 
#> Altered variable: Y_Z_1 
#>   Before: 
#>    min median mean  max   sd N_missing N_unique
#>  -2.34   1.03 0.97 4.14 1.06         0      500
#> 
#>   After:
#>    min median mean  max   sd N_missing N_unique
#>  -2.34   0.99 0.92 4.14 1.13         0      250
#> 

# Special Cases

# You may wish to have a design with only one step:

design <- my_population + NULL
design

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 500, noise = rnorm(N)) -------------
#> 
#> N = 500 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      500 character
#> 
#> Added variable: noise 
#>    min median mean max   sd N_missing N_unique
#>  -3.18   0.01 0.02 3.2 0.99         0      500
#> 

if (FALSE) {
diagnosis <- diagnose_design(design)

summary(diagnosis)
}