Declare estimator

Declares an estimator which generates estimates and associated statistics.

Use of declare_test is identical to use of declare_estimator. Use declare_test for hypothesis testing with no specific estimand in mind; use declare_estimator for hypothesis testing when you can link each estimate to an estimand. For example, declare_test could be used for a K-S test of distributional equality and declare_estimator for a difference-in-means estimate of an average treatment effect.

declare_estimator(
  ...,
  handler = label_estimator(model_handler),
  label = "estimator"
)

declare_estimators(
  ...,
  handler = label_estimator(model_handler),
  label = "estimator"
)

label_estimator(fn)

model_handler(
  data,
  ...,
  model = estimatr::difference_in_means,
  model_summary = tidy_try,
  term = FALSE
)

Arguments

...
arguments to be captured, and later passed to the handler

handler
a tidy-in, tidy-out function

label
a string describing the step

fn
A function that takes a data.frame as an argument and returns a data.frame with the estimates, summary statistics (i.e., standard error, p-value, and confidence interval), and a term column for labeling coefficient estimates.

data
a data.frame

model
A model function, e.g. lm or glm. By default, the model is the difference_in_means function from the estimatr package.

model_summary
A model-in data-out function to extract coefficient estimates or model summary statistics, such as tidy or glance. By default, the DeclareDesign model summary function tidy_try is used, which first attempts to use the available tidy method for the model object sent to model, then if not attempts to summarize coefficients using the coef(summary()) and confint methods. If these do not exist for the model object, it fails.

term
Symbols or literal character vector of term that represent quantities of interest, i.e. Z. If FALSE, return the first non-intercept term; if TRUE return all term. To escape non-standard-evaluation use !!.

Value

A function that accepts a data.frame as an argument and returns a data.frame containing the value of the estimator and associated statistics.

Details

declare_estimator is designed to handle two main ways of generating parameter estimates from data.

In declare_estimator, you can optionally provide the name of an estimand or an objected created by declare_estimand to connect your estimate(s) to estimand(s).

The first is through label_estimator(model_handler), which is the default value of the handler argument. Users can use standard modeling functions like lm, glm, or iv_robust. The models are summarized using the function passed to the model_summary argument. This will usually be a "tidier" like broom::tidy. The default model_summary function is tidy_try, which applies a tidy method if available, and if not, tries to make one on the fly.

An example of this approach is:

declare_estimator(Y ~ Z + X, model = lm_robust, model_summary = tidy, term = "Z", estimand = "ATE")

The second approach is using a custom data-in, data-out function, usually first passed to label_estimator. The reason to pass the custom function to label_estimator first is to enable clean labeling and linking to estimands.

An example of this approach is:

my_fun <- function(data){ with(data, median(Y[Z == 1]) - median(Y[Z == 0])) }

declare_estimator(handler = label_estimator(my_fun), estimand = "ATE")

label_estimator takes a data-in-data out function to fn, and returns a data-in-data-out function that first runs the provided estimation function fn and then appends a label for the estimator and, if an estimand is provided, a label for the estimand.

Examples

# Declare estimand
my_estimand <- declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0))

# Declare estimator using the default handler using `difference_in_means`
# estimator from `estimatr` package. Returns the first non-intercept term
# as estimate

my_estimator_dim <- declare_estimator(Y ~ Z, estimand = "ATE", label = "DIM")

# Use lm function from base R
my_estimator_lm <- declare_estimator(Y ~ Z, estimand = "ATE",
  model = lm, label = "LM")
# Use lm_robust (linear regression with robust standard errors) from
# `estimatr` package

my_estimator_lm_rob <- declare_estimator(
  Y ~ Z,
  estimand = "ATE",
  model = lm_robust,
  label = "LM_Robust"
)

# Set `term` if estimate of interest is not the first non-intercept variable
my_estimator_lm_rob_x <- declare_estimator(
  Y ~ X + Z,
  estimand = my_estimand,
  term = "Z",
  model = lm_robust
)

# Use glm from base R
my_estimator_glm <- declare_estimator(
  Y ~ X + Z,
  family = "gaussian",
  estimand = my_estimand,
  term = "Z",
  model = glm
)

# A probit
estimator_probit <- declare_estimator(
  Y ~ Z,
  model = glm,
  family = binomial(link = "probit"),
  term = "Z"
)

# Declare estimator using a custom handler

# Define your own estimator and use the `label_estimator` function for labeling
# Must have `data` argument that is a data.frame
my_estimator_function <- function(data){
  data.frame(estimate = with(data, mean(Y)))
}

my_estimator_custom <- declare_estimator(
  handler = label_estimator(my_estimator_function),
  estimand = my_estimand
)

# Customize labeling

my_estimator_function <- function(data){
  data.frame(
    estimator_label = "foo",
    estimand_label = "bar",
    estimate = with(data, mean(Y)),
    n = nrow(data),
    stringsAsFactors = FALSE
  )
}

my_estimator_custom2 <- declare_estimator(handler = my_estimator_function)

# Examples

# First, set up the rest of a design
set.seed(42)

design_def <-
  declare_population(N = 100, X = rnorm(N), W = rexp(N, 1), noise = rnorm(N)) +
  declare_potential_outcomes(Y ~ .25 * Z + noise) +
  declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_assignment(m = 50) +
  reveal_outcomes() +
  my_estimator_dim

draw_estimates(design_def)

#>   estimator_label term estimate std.error statistic   p.value     conf.low
#> 1             DIM    Z 0.404084 0.2041834  1.979024 0.0506206 -0.001113517
#>   conf.high       df outcome estimand_label
#> 1 0.8092816 97.95838       Y            ATE

# Can also use declared estimator on a data.frame
dat <- draw_data(design_def)
my_estimator_dim(dat)

#>   estimator_label term   estimate std.error statistic  p.value   conf.low
#> 1             DIM    Z 0.07526085 0.2027897 0.3711277 0.711344 -0.3271717
#>   conf.high       df outcome estimand_label
#> 1 0.4776934 97.93985       Y            ATE

# ----------
# 2. Using existing estimators
# ----------

design <- replace_step(design_def, my_estimator_dim, my_estimator_lm_rob)
draw_estimates(design)

#>   estimator_label term    estimate std.error  statistic   p.value   conf.low
#> 1       LM_Robust    Z -0.04532556 0.1925646 -0.2353784 0.8144062 -0.4274638
#>   conf.high df outcome estimand_label
#> 1 0.3368127 98       Y            ATE

design <- replace_step(design_def, my_estimator_dim, my_estimator_lm)
draw_estimates(design)

#>   estimator_label term   estimate std.error  statistic  p.value  conf.low
#> 1              LM    Z -0.1204321 0.2203432 -0.5465663 0.585919 -0.557696
#>   conf.high estimand_label
#> 1 0.3168317            ATE

design <- replace_step(design_def, my_estimator_dim, my_estimator_glm)
draw_estimates(design)

#>   estimator_label term  estimate std.error statistic   p.value  conf.low
#> 1       estimator    Z 0.2783732 0.2012977  1.382893 0.1698728 -0.116163
#>   conf.high estimand_label
#> 1 0.6729095            ATE

# ----------
# 3. Using custom estimators
# ----------

design <- replace_step(design_def, my_estimator_dim, my_estimator_custom)

draw_estimates(design)

#>   estimator_label  estimate estimand_label
#> 1       estimator 0.1658533            ATE

# The names in your custom estimator return should match with
# your diagnosands when diagnosing a design
my_median <- function(data) data.frame(med = median(data$Y))

my_estimator_median <- declare_estimator(
  handler = label_estimator(my_median),
  estimand = my_estimand
)

design <- replace_step(design_def, my_estimator_dim, my_estimator_median)

draw_estimates(design)

#>   estimator_label       med estimand_label
#> 1       estimator 0.1933848            ATE

my_diagnosand <- declare_diagnosands(med_to_estimand = mean(med - estimand))

if (FALSE) {
diagnose_design(design, diagnosands = my_diagnosand, sims = 5,
  bootstrap_sims = FALSE)
}

# ----------
# 4. Multiple estimators per estimand
# ----------

design_two <- insert_step(design_def,  my_estimator_lm,
  after = my_estimator_dim)

draw_estimates(design_two)

#>   estimator_label term  estimate std.error statistic    p.value   conf.low
#> 1             DIM    Z 0.5016346 0.2104809  2.383279 0.01911112 0.08387755
#> 2              LM    Z 0.5016346 0.2104809  2.383279 0.01908682 0.08394223
#>   conf.high       df outcome estimand_label
#> 1 0.9193918 96.80092       Y            ATE
#> 2 0.9193271       NA    <NA>            ATE

if (FALSE) {
diagnose_design(design_two, sims = 5, bootstrap_sims = FALSE)
}