Declare measurement procedure

This function adds measured data columns that can be functions of unmeasured data columns.

Usage

declare_measurement(..., handler = measurement_handler, label = NULL)

measurement_handler(data, ...)

Arguments

...

arguments to be captured, and later passed to the handler

handler

a tidy-in, tidy-out function

label

a string describing the step

data

A data.frame.

Value

A function that returns a data.frame.

Details

It is also possible to include measured variables in your declare_model call or to add variables using declare_step. However, putting latent variables in declare_model and variables-as-measured in declare_measurement helps communicate which parts of your research design are in M and which parts are in D.

Examples

# declare_measurement in use
## Two-arm randomized experiment
design <-
  declare_model(
    N = 500,
    X = rep(c(0, 1), each = N / 2),
    U = rnorm(N, sd = 0.25),
    potential_outcomes(Y ~ 0.2 * Z + X + U)
  ) +
  declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(S = complete_rs(N = N, n = 200)) +
  declare_assignment(Z = complete_ra(N = N, m = 100)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z)) +
  declare_estimator(Y ~ Z, inquiry = "ATE")
  
run_design(design)
#>   inquiry estimand estimator term   estimate  std.error statistic   p.value
#> 1     ATE      0.2 estimator    Z 0.01267358 0.08170231  0.155119 0.8768855
#>     conf.low conf.high  df outcome
#> 1 -0.1484448  0.173792 198       Y

# Reveal potential outcomes according to treatment assignment
design <-
  declare_model(N = 100,
                potential_outcomes(Y ~ rbinom(
                  N, size = 1, prob = 0.1 * Z + 0.5
                ))) +
  declare_assignment(Z = complete_ra(N, m = 50)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z))
  
head(draw_data(design))
#>    ID Y_Z_0 Y_Z_1 Z Y
#> 1 001     0     0 1 0
#> 2 002     0     1 0 0
#> 3 003     1     0 1 0
#> 4 004     0     0 1 0
#> 5 005     0     1 0 0
#> 6 006     0     0 0 0

# Generate observed measurement from a latent value
design <- 
  declare_model(N = 100, latent = runif(N)) +
  declare_measurement(observed = rbinom(N, 1, prob = latent))
  
head(draw_data(design))
#>    ID    latent observed
#> 1 001 0.1830983        1
#> 2 002 0.1779913        0
#> 3 003 0.2305162        0
#> 4 004 0.6299517        0
#> 5 005 0.8878633        1
#> 6 006 0.7812405        1

# Index creation
library(psych)

design <-
  declare_model(
    N = 500,
    X = rep(c(0, 1), each = N / 2),
    Y_1 = 0.2 * X + rnorm(N, sd = 0.25),
    Y_2 = 0.3 * X + 0.5 * rnorm(N, sd = 0.50),
    Y_3 = 0.1 * X + 0.4 * rnorm(N, sd = 0.75)) +
  declare_measurement(
    index = fa(
      r = cbind(Y_1, Y_2, Y_3),
      nfactors = 1,
      rotate = "varimax"
    )$scores
  )

draw_data(design)
#>      ID X          Y_1          Y_2          Y_3          MR1
#> 1   001 0  0.217380106 -0.075675091 -0.273417273  0.412030020
#> 2   002 0  0.222766155  0.040045057  0.068188240  0.433643954
#> 3   003 0 -0.133464455 -0.627876517 -0.167340895 -0.882196184
#> 4   004 0  0.220755097 -0.152654970  0.016001543  0.426431205
#> 5   005 0 -0.161018274  0.079815279 -0.877468855 -0.990166503
#> 6   006 0 -0.218831231  0.241356980 -0.259024577 -1.200553515
#> 7   007 0 -0.246598450  0.078295894  0.395429537 -1.298817916
#> 8   008 0 -0.275677935  0.328818248  0.362149026 -1.407150961
#> 9   009 0  0.225131916 -0.436397205  0.132020471  0.444070594
#> 10  010 0 -0.256536159 -0.001023887 -0.039901078 -1.337906870
#> 11  011 0  0.110914583  0.087726041  0.008606862  0.019867905
#> 12  012 0  0.054162826  0.200558561  0.339314475 -0.188174078
#> 13  013 0 -0.496073414  0.073525768  0.015818842 -2.222869861
#> 14  014 0  0.200904045 -0.200135476  0.449023466  0.355784577
#> 15  015 0  0.104682414  0.451855916 -0.218988121 -0.005510272
#> 16  016 0  0.114106816  0.110146268  0.424594269  0.034073367
#> 17  017 0 -0.184295058 -0.041376035  0.113811745 -1.069952235
#> 18  018 0  0.392526576  0.016546420  0.190297611  1.061698328
#> 19  019 0 -0.219627064  0.039437366  0.411273110 -1.198958051
#> 20  020 0 -0.163798250  0.387829283 -0.086533612 -0.996583356
#> 21  021 0 -0.318549297  0.032387232  0.217773189 -1.565605891
#> 22  022 0  0.050107468  0.269951804  0.066571294 -0.204969342
#> 23  023 0 -0.154017736 -0.193478787  0.579486375 -0.954893846
#> 24  024 0  0.089035428  0.231729681 -0.010408261 -0.061483216
#> 25  025 0  0.022550626  0.006292297  0.206091924 -0.305240180
#> 26  026 0 -0.061213101 -0.228345385  0.302038848 -0.613533798
#> 27  027 0 -0.154837597  0.042941607 -0.069678647 -0.962428302
#> 28  028 0  0.113332665 -0.193294527  0.063290837  0.029897880
#> 29  029 0  0.619102099 -0.299293267 -0.197890834  1.897453717
#> 30  030 0  0.030779401 -0.453425077 -0.037799404 -0.275023915
#> 31  031 0 -0.036308492  0.629662060  0.040092629 -0.525419420
#> 32  032 0  0.062599698 -0.029765402  0.135857749 -0.157577401
#> 33  033 0 -0.357830314 -0.021351018 -0.094357015 -1.712456427
#> 34  034 0  0.110558219 -0.158027374  0.200699476  0.020366063
#> 35  035 0 -0.075878972  0.140828386  0.263367882 -0.668965807
#> 36  036 0 -0.279154508  0.392819373 -0.282952594 -1.424005232
#> 37  037 0 -0.211636340  0.314882464  0.377769705 -1.170386379
#> 38  038 0  0.248955856 -0.126190620 -0.485129499  0.527582915
#> 39  039 0  0.423391913  0.118834885  0.162078291  1.175297659
#> 40  040 0  0.327583665 -0.108834012  0.415644437  0.823416963
#> 41  041 0  0.058880723 -0.357692283  0.100707260 -0.170629034
#> 42  042 0  0.270850300 -0.455469525 -0.092393622  0.611718843
#> 43  043 0 -0.436239071 -0.338246923  0.186525048 -1.999639170
#> 44  044 0 -0.209821274 -0.117769769  0.478897072 -1.161892989
#> 45  045 0  0.067381024 -0.120215915  0.142841723 -0.139620998
#> 46  046 0  0.266371074 -0.315413743  0.148979316  0.596218362
#> 47  047 0 -0.235421529  0.189875519  0.649388469 -1.256316132
#> 48  048 0  0.395305310  0.360914973 -0.159108391  1.068945837
#> 49  049 0  0.063155320  0.324242468 -0.457313784 -0.160019060
#> 50  050 0  0.372018223 -0.004126818 -0.567700332  0.981472968
#> 51  051 0 -0.066546019 -0.241866450 -0.253920209 -0.636505127
#> 52  052 0 -0.472462616  0.454567372  0.327889920 -2.134818219
#> 53  053 0  0.289843541 -0.510039145 -0.334247945  0.680611413
#> 54  054 0 -0.291381415  0.322881968  0.261926629 -1.465754499
#> 55  055 0  0.028047709 -0.136286184 -0.281403929 -0.287434161
#> 56  056 0  0.022780494  0.175706977 -0.225286772 -0.307418236
#> 57  057 0  0.117396704 -0.038323685  0.046340949  0.044389022
#> 58  058 0 -0.044210198 -0.158327922 -0.018043402 -0.552801806
#> 59  059 0 -0.126188280 -0.198373080 -0.263779267 -0.857060964
#> 60  060 0  0.054221451 -0.097421553 -0.248201566 -0.190631485
#> 61  061 0 -0.182729973 -0.164122239  0.171251340 -1.063491530
#> 62  062 0  0.222471683 -0.070173225 -0.007422617  0.432408735
#> 63  063 0  0.072681423 -0.004118845  0.165861075 -0.120217474
#> 64  064 0 -0.012157550 -0.373496131  0.249608298 -0.432187298
#> 65  065 0 -0.218733880  0.534232367  0.172797301 -1.198430855
#> 66  066 0 -0.334055527 -0.084160784  0.118843226 -1.623169729
#> 67  067 0  0.370619362  0.238017230 -0.104770088  0.978391013
#> 68  068 0 -0.064956835 -0.058389923 -0.513624215 -0.632679041
#> 69  069 0  0.450020393 -0.412014845 -0.243087954  1.272737541
#> 70  070 0 -0.583504557  0.096180053  0.455508464 -2.543377421
#> 71  071 0  0.008763528  0.114373700  0.290409778 -0.355978744
#> 72  072 0 -0.220392336  0.378892630  0.084446114 -1.204658017
#> 73  073 0 -0.233414644 -0.134648320  0.770466753 -1.247291802
#> 74  074 0 -0.127586736 -0.022839337  0.135902667 -0.860334646
#> 75  075 0 -0.312795146 -0.017953238 -0.368280700 -1.547688411
#> 76  076 0 -0.045720226 -0.123568928 -0.018432035 -0.558478936
#> 77  077 0 -0.289342791  0.580239495  0.273018165 -1.458861285
#> 78  078 0 -0.026091030 -0.088914997 -0.310959315 -0.487782149
#> 79  079 0  0.061437202  0.462992707  0.366904148 -0.161850660
#> 80  080 0  0.385898651 -0.115483439  0.648583835  1.040292981
#> 81  081 0  0.188556714 -0.346126570  0.123659654  0.308628188
#> 82  082 0 -0.072572403 -0.041372528  0.083160542 -0.657319356
#> 83  083 0  0.136102535 -0.124483077 -0.446590827  0.110814402
#> 84  084 0  0.185967916 -0.478476347  0.237854245  0.300103840
#> 85  085 0 -0.211913045 -0.560607086  0.296735233 -1.169490621
#> 86  086 0 -0.258879193  0.095538542  0.056883024 -1.346253997
#> 87  087 0  0.313823441  0.031941636  0.107389940  0.770355589
#> 88  088 0  0.289086899 -0.428939183  0.077300218  0.680038572
#> 89  089 0  0.063311964  0.482700361  0.180355624 -0.156085826
#> 90  090 0 -0.274764656  0.205957999  0.147599110 -1.404714410
#> 91  091 0 -0.416247157 -0.161005047 -0.302496790 -1.929160448
#> 92  092 0 -0.199126699  0.436575023 -0.208655482 -1.127981113
#> 93  093 0 -0.061053640 -0.111389088 -0.304112948 -0.616866651
#> 94  094 0  0.156597952  0.559966215  0.242896829  0.188765542
#> 95  095 0  0.365168602 -0.329116167 -0.258595643  0.958890972
#> 96  096 0 -0.312437239  0.195257182 -0.171581578 -1.545781638
#> 97  097 0 -0.054028338 -0.279291703 -0.071253686 -0.589064413
#> 98  098 0  0.623101278  0.137941337  0.550332034  1.915477866
#> 99  099 0  0.375032830  0.085614596 -0.127495020  0.994981470
#> 100 100 0  0.143184036  0.365911755  0.209168229  0.139532636
#> 101 101 0 -0.277796872  0.221153853  0.125965332 -1.416088633
#> 102 102 0 -0.012250570  0.307711734  0.103212513 -0.435267970
#> 103 103 0 -0.074942017  0.313930571  0.046593716 -0.667266200
#> 104 104 0 -0.154497708  0.346919996 -0.308395784 -0.963423949
#> 105 105 0 -0.182455870 -0.119108972 -0.143312252 -1.064471099
#> 106 106 0  0.019148115 -0.325266858 -0.248436537 -0.319604283
#> 107 107 0 -0.217049372  0.167763011  0.094498007 -1.191667355
#> 108 108 0  0.219938245  0.005370203  0.454146621  0.425583071
#> 109 109 0 -0.118934998  0.073630856  0.085630710 -0.828929599
#> 110 110 0  0.239403437 -0.009721258 -0.480894909  0.491992670
#> 111 111 0  0.424566607  0.105745669 -0.115188678  1.178026629
#> 112 112 0 -0.031267056  0.392360006  0.081865257 -0.505892702
#> 113 113 0  0.333221796  0.063069060  0.573216711  0.844714902
#> 114 114 0  0.123319804  0.006169494  0.049860387  0.066173825
#> 115 115 0 -0.197948327 -0.122449027  0.043390390 -1.120597240
#> 116 116 0  0.002978290 -0.071108673 -0.713185005 -0.382809722
#> 117 117 0  0.038259722 -0.004177953  0.562301776 -0.245049955
#> 118 118 0  0.100735019 -0.094159654  0.622700729 -0.013598348
#> 119 119 0 -0.007396155  0.272948932 -0.359077180 -0.419982357
#> 120 120 0 -0.178393372  0.013068190  0.071749916 -1.048544658
#> 121 121 0 -0.101941087 -0.286197252  0.302524108 -0.763862180
#> 122 122 0  0.002098044 -0.290139601  0.173966047 -0.380190813
#> 123 123 0  0.289798277 -0.207721672  0.435321613  0.684187981
#> 124 124 0 -0.193862806 -0.473091748  0.492186857 -1.101873603
#> 125 125 0 -0.134627008 -0.179599653  0.312289084 -0.884870755
#> 126 126 0 -0.235485966  0.709076862  0.289813887 -1.260113755
#> 127 127 0  0.222196606  0.218973670  0.741717360  0.435050862
#> 128 128 0  0.375316554  0.002702836  0.049271316  0.997307357
#> 129 129 0 -0.009774418  0.076507846 -0.080864737 -0.426578585
#> 130 130 0 -0.046150934  0.055415413 -0.143512253 -0.561304159
#> 131 131 0 -0.430347472  0.426212670  0.162615527 -1.980107111
#> 132 132 0  0.384524967  0.216397040  0.304077310  1.032260652
#> 133 133 0 -0.109007421 -0.426312699 -0.135563106 -0.792191148
#> 134 134 0  0.039720643  0.009951222 -0.177262919 -0.244084756
#> 135 135 0  0.095989642  0.309193474 -0.298016943 -0.037708579
#> 136 136 0  0.115310584 -0.036676954  0.153835298  0.037314985
#> 137 137 0  0.111065292 -0.273437738 -0.057669193  0.021020915
#> 138 138 0 -0.258580573 -0.263876460  0.309688130 -1.342663407
#> 139 139 0 -0.099251132 -0.697564560 -0.021753932 -0.754721994
#> 140 140 0  0.060235214 -0.002955713 -0.217186086 -0.168485284
#> 141 141 0  0.102423148 -0.014350383  0.103307035 -0.010665481
#> 142 142 0  0.119665595  0.017233278  0.353906403  0.054447713
#> 143 143 0  0.064398874  0.047659297 -0.094866437 -0.152512528
#> 144 144 0 -0.288559429  0.003547435  0.373596360 -1.453788498
#> 145 145 0 -0.074648992  0.210268690 -0.043037073 -0.666431894
#> 146 146 0 -0.355366942 -0.258630634 -0.150619847 -1.703038195
#> 147 147 0  0.302815728  0.145414486  0.577898469  0.732166648
#> 148 148 0 -0.031031253 -0.294031162 -0.161066137 -0.504583455
#> 149 149 0 -0.426457179  0.079440064  0.128723824 -1.964983378
#> 150 150 0 -0.239140276 -0.217930249  0.406085414 -1.270384729
#> 151 151 0  0.412233878  0.223266065  0.181482874  1.133897797
#> 152 152 0  0.092371549  0.115066909 -0.081145853 -0.049256782
#> 153 153 0  0.400409729 -0.261279856  0.117208335  1.091153801
#> 154 154 0 -0.413909494  0.209471284 -0.392505474 -1.922073031
#> 155 155 0  0.007513882  0.031649294 -0.218254046 -0.363391703
#> 156 156 0 -0.218844938 -0.036859763  0.204999702 -1.197084558
#> 157 157 0  0.212350547 -0.416802720 -0.022561765  0.395871291
#> 158 158 0  0.137964131  0.013234240  0.236802674  0.121375952
#> 159 159 0  0.385413901 -0.116181707  0.041637543  1.034897533
#> 160 160 0  0.196430905  0.032043259 -0.130034244  0.335179435
#> 161 161 0  0.080466449 -0.004299488 -0.133103811 -0.093227643
#> 162 162 0  0.038898268  0.267051803 -0.204283867 -0.247988670
#> 163 163 0  0.004361478 -0.086406703 -0.403032541 -0.375814121
#> 164 164 0  0.162189077  0.211539631 -0.039229172  0.208703481
#> 165 165 0 -0.253070268 -0.380936986  0.221109301 -1.322508284
#> 166 166 0  0.143678624 -0.130585288 -0.280176953  0.139813523
#> 167 167 0  0.151350945 -0.195943406  0.141638448  0.170848121
#> 168 168 0 -0.666993453  0.207201325 -0.007805150 -2.854925848
#> 169 169 0 -0.316296167  0.060175064  0.444405227 -1.556010194
#> 170 170 0  0.266709068 -0.161686120  0.367438204  0.598343881
#> 171 171 0  0.242327912 -0.744259697  0.035690310  0.507881247
#> 172 172 0 -0.296297338  0.080623571 -0.189270292 -1.485935631
#> 173 173 0 -0.059964552  0.106023804 -0.099234812 -0.612221087
#> 174 174 0 -0.226739406  0.260652260 -0.415341711 -1.230755865
#> 175 175 0  0.039493785  0.009074836 -0.157197518 -0.244801375
#> 176 176 0 -0.183175602  0.100715549 -0.223781807 -1.068211146
#> 177 177 0 -0.277992325  0.161529719 -0.040577304 -1.417636929
#> 178 178 0  0.365503952  0.533543355 -0.025539264  0.959150323
#> 179 179 0 -0.395397535 -0.422020270  0.263439248 -1.848043161
#> 180 180 0  0.285521986 -0.288727774 -0.396675339  0.663665760
#> 181 181 0  0.098967699 -0.219579407 -0.315379897 -0.025358497
#> 182 182 0 -0.127216197  0.427853265 -0.686092801 -0.865084762
#> 183 183 0 -0.142636615  0.676111451 -0.241918037 -0.920104541
#> 184 184 0 -0.464345195  0.668235510  0.020420578 -2.107236864
#> 185 185 0  0.056583017 -0.097588461  0.418133586 -0.177945966
#> 186 186 0 -0.087640943 -0.232886794 -0.208675654 -0.714206625
#> 187 187 0  0.046331830 -0.261034255  0.188811002 -0.216738627
#> 188 188 0 -0.165286968 -0.028430497 -0.192350372 -1.001571945
#> 189 189 0  0.104825989  0.075526765  0.142413253 -0.001800984
#> 190 190 0 -0.189312623 -0.321138902  0.100488587 -1.087804481
#> 191 191 0  0.086418103  0.378180498 -0.206532517 -0.072720925
#> 192 192 0  0.143941045  0.335795937 -0.124178091  0.140431729
#> 193 193 0  0.258173947  0.526913316 -0.676614503  0.558715913
#> 194 194 0  0.619636242  0.053945499 -0.158380695  1.898693924
#> 195 195 0  0.089295357  0.057231547  0.046090097 -0.059708805
#> 196 196 0  0.304366304 -0.252649045 -0.154232413  0.734637083
#> 197 197 0  0.080188432 -0.038635845 -0.427319179 -0.095908904
#> 198 198 0 -0.039494795 -0.527334171  0.183222705 -0.533171145
#> 199 199 0  0.002717621  0.119382569  0.086926275 -0.379541101
#> 200 200 0 -0.024244494 -0.102678543  0.214172433 -0.477801371
#> 201 201 0  0.284748470  0.327291867 -0.156190552  0.660548010
#> 202 202 0  0.010076871 -0.113186874 -0.080454977 -0.352705736
#> 203 203 0 -0.103109328 -0.437533614 -0.198268465 -0.770739519
#> 204 204 0 -0.235469220  0.223499345 -0.048628363 -1.260731826
#> 205 205 0 -0.496045370  0.205822952  0.400633324 -2.220842452
#> 206 206 0 -0.141926976  0.171214029 -0.312784714 -0.916519596
#> 207 207 0 -0.307935887 -0.350313950  0.043341582 -1.526376800
#> 208 208 0 -0.353530017  0.247539396 -0.285913270 -1.698441990
#> 209 209 0 -0.388232317 -0.596369725  0.269526809 -1.821053645
#> 210 210 0 -0.053105724  0.008888002 -0.196202730 -0.587187629
#> 211 211 0 -0.556081683  0.421569064  0.273043772 -2.444025946
#> 212 212 0 -0.221068993 -0.551540446 -0.138214495 -1.205930965
#> 213 213 0 -0.001027118 -0.074549775 -0.070056394 -0.393779072
#> 214 214 0 -0.300169224  0.267650838  0.198747554 -1.498449437
#> 215 215 0  0.023184051 -0.173218735  0.510288151 -0.300600246
#> 216 216 0  0.043882319 -0.184581730 -0.189900933 -0.228249262
#> 217 217 0  0.187691751 -0.204301790 -0.323239341  0.302388138
#> 218 218 0 -0.157127969 -0.307679866  0.148951782 -0.968631185
#> 219 219 0 -0.109036021  0.355757527 -0.495134834 -0.796576846
#> 220 220 0  0.458170249 -0.175110212  0.033031850  1.303842493
#> 221 221 0  0.203672985 -0.071967226 -0.056470825  0.362661074
#> 222 222 0  0.214245480 -0.070785503 -0.045631709  0.401787585
#> 223 223 0 -0.005827609  0.131065748  0.033803747 -0.411463363
#> 224 224 0  0.297371273 -0.061962572 -0.292796367  0.707444510
#> 225 225 0  0.067969864 -0.080481620  0.218699757 -0.137103427
#> 226 226 0  0.020920914  0.256210973 -0.131880195 -0.313955056
#> 227 227 0  0.127836255  0.001833791 -0.001238004  0.082570382
#> 228 228 0 -0.163097012 -0.261511417 -0.469284354 -0.994486625
#> 229 229 0 -0.017541613 -0.365485954  0.160529018 -0.452632628
#> 230 230 0  0.189480502 -0.248313852  0.081154451  0.311520940
#> 231 231 0  0.447661744  0.122692668 -0.246121038  1.262538724
#> 232 232 0  0.218551039  0.309044260 -0.105218713  0.416301491
#> 233 233 0  0.143280010 -0.028213695 -0.581864095  0.136267555
#> 234 234 0  0.314662034  0.111391978 -0.315205654  0.770725542
#> 235 235 0 -0.280046711  0.251913423 -0.353640167 -1.427335701
#> 236 236 0  0.414278486 -0.031061870  0.109525508  1.141722173
#> 237 237 0 -0.249848483  0.137158119  0.064628596 -1.312953595
#> 238 238 0 -0.169987209  0.285802139  0.021324667 -1.018531051
#> 239 239 0 -0.323336126 -0.327771391 -0.290836962 -1.585327988
#> 240 240 0 -0.304273864  0.315613552  0.391061425 -1.512604889
#> 241 241 0 -0.078863339 -0.019426763  0.341783130 -0.679087880
#> 242 242 0  0.033858120  0.079723648  0.358442877 -0.262755076
#> 243 243 0 -0.381915568  0.137496366  0.253540394 -1.799819726
#> 244 244 0 -0.020965022  0.338835186  0.130518645 -0.467390919
#> 245 245 0  0.002200134  0.197329364 -0.134667126 -0.382983472
#> 246 246 0 -0.253209058  0.039905089  0.578894221 -1.322048805
#> 247 247 0  0.459145504 -0.018989306 -0.158740616  1.305878712
#> 248 248 0 -0.275736861 -0.124418214  0.266229282 -1.406696318
#> 249 249 0 -0.076387504  0.247516418  0.112147137 -0.672035747
#> 250 250 0  0.201510066 -0.258674226  0.318816521  0.357410639
#> 251 251 1 -0.370689281  0.706405774  0.248339114 -1.759928862
#> 252 252 1 -0.146264991  0.630941741 -0.027869025 -0.932115790
#> 253 253 1  0.159938816 -0.151872724  0.421121620  0.204120064
#> 254 254 1  0.235351367 -0.246086779 -0.375069402  0.478296912
#> 255 255 1  0.224989647  0.517458472  0.218053601  0.441441634
#> 256 256 1  0.596651397  0.340057424 -0.282598308  1.812242703
#> 257 257 1  0.036841797  0.309738339 -0.061415091 -0.254855465
#> 258 258 1  0.164411073  0.161847399  0.026058529  0.217437867
#> 259 259 1  0.294997828  0.274937253  0.438305363  0.702095110
#> 260 260 1  0.163741386 -0.028443846  0.487648618  0.218227519
#> 261 261 1 -0.059072571 -0.086735231  0.539483040 -0.604601894
#> 262 262 1  0.780940538  0.187173965  0.194686166  2.496445603
#> 263 263 1  0.327744292  0.915184306  0.208266965  0.819971573
#> 264 264 1  0.347803318  0.551815294 -0.440730449  0.891229544
#> 265 265 1  0.297200088  0.402430128 -0.020496917  0.707157000
#> 266 266 1  0.318330393  0.152669110  0.324950709  0.787970515
#> 267 267 1  0.053372501  0.357476406  0.490901046 -0.190623765
#> 268 268 1  0.287029578  0.466406099 -0.673419456  0.665522246
#> 269 269 1  0.038616724 -0.209551616  0.428742417 -0.243961470
#> 270 270 1  0.068290525  0.524813408  0.609537115 -0.135255467
#> 271 271 1  0.327532878  0.542093607  0.202404840  0.820178182
#> 272 272 1  0.145662988  0.436482689  0.127070006  0.148011136
#> 273 273 1  0.203326029  0.232315405  0.138558260  0.361703833
#> 274 274 1  0.233049336  0.047239800 -0.035505253  0.471004467
#> 275 275 1  0.156401787  0.442912909 -0.338272215  0.184908493
#> 276 276 1  0.255947095  0.306558711  0.377182637  0.557352884
#> 277 277 1  0.258007315 -0.177579944 -0.267174466  0.562463890
#> 278 278 1  0.217109121  0.243665504 -0.040525832  0.411537182
#> 279 279 1  0.287457583  0.445157925  0.362737380  0.673318360
#> 280 280 1  0.430619498  0.241583215  0.702953632  1.204880776
#> 281 281 1 -0.123611249  0.591838055  0.081055020 -0.847655894
#> 282 282 1  0.365081547  0.546254945  0.221847513  0.959024559
#> 283 283 1  0.379014911 -0.076491182 -0.311609872  1.009045644
#> 284 284 1  0.003835327  0.448149099  0.006304214 -0.376791326
#> 285 285 1  0.423399742  0.737606081 -0.332416580  1.170692505
#> 286 286 1  0.362764934  0.506088725  0.535378392  0.952437636
#> 287 287 1  0.337367690  0.471787914  0.423482810  0.858024049
#> 288 288 1  0.411209513  0.016655785  0.571725704  1.132997728
#> 289 289 1  0.140084234  0.081532076  0.385856966  0.129908147
#> 290 290 1  0.180002704  0.560830602 -0.051280236  0.273495814
#> 291 291 1  0.087743010  0.329969501  0.054294920 -0.066143530
#> 292 292 1 -0.117323916  0.045991462  0.174653813 -0.822371959
#> 293 293 1  0.315496179  0.315497891 -0.014021397  0.775037773
#> 294 294 1  0.358051443  0.070542406  0.021504285  0.933161879
#> 295 295 1  0.077450169  0.194905206  0.064170680 -0.103746677
#> 296 296 1 -0.017459560  0.779071290  0.128300258 -0.455658079
#> 297 297 1 -0.011502728  0.251284414  0.148416623 -0.432081447
#> 298 298 1  0.138179239 -0.049078152 -0.050256640  0.120635986
#> 299 299 1  0.084013097  0.919880371  0.207549571 -0.080631879
#> 300 300 1  0.504541765 -0.026169831 -0.144604824  1.473721483
#> 301 301 1  0.627260965  0.048056446 -0.218962064  1.926523454
#> 302 302 1  0.402737572 -0.355097053  0.038825788  1.099546608
#> 303 303 1  0.253281955  0.099497474  0.662900750  0.549770350
#> 304 304 1  0.026816967 -0.028967886  0.198721071 -0.289423208
#> 305 305 1  0.194234422 -0.088336894  0.069741364  0.328580369
#> 306 306 1  0.434155982  0.422927894  0.255717385  1.214793741
#> 307 307 1 -0.037257490  0.374226061 -0.204446440 -0.529678780
#> 308 308 1  0.218439274  0.323754576  0.015496085  0.416565435
#> 309 309 1  0.265107389  0.489074303 -0.093283655  0.587904582
#> 310 310 1  0.537457589  0.335305603 -0.416825233  1.592737198
#> 311 311 1 -0.411891405  0.053656164 -0.038063346 -1.912083176
#> 312 312 1  0.598396673  0.546557993  0.574647695  1.823218878
#> 313 313 1  0.226195290  0.320759137  0.329732886  0.447099170
#> 314 314 1  0.094048392  0.026277419  0.271871570 -0.040720011
#> 315 315 1  0.633709061  0.462345278  0.412820576  1.952967404
#> 316 316 1  0.061846105 -0.050534048 -0.029829224 -0.161289442
#> 317 317 1  0.015264997  0.049973875 -0.409447253 -0.335937561
#> 318 318 1  0.209782598  0.320042246  0.069179624  0.384908175
#> 319 319 1 -0.196142130  0.236565301  0.101469035 -1.114562298
#> 320 320 1  0.285407059  0.209287490  0.169174855  0.665238117
#> 321 321 1 -0.262470303  0.590857767 -0.203479512 -1.362427738
#> 322 322 1  0.387398485  0.287958158 -0.165768168  1.039890534
#> 323 323 1  0.027737928  0.890431431  0.135278706 -0.288917213
#> 324 324 1  0.112820857 -0.088615938  0.166274760  0.028331720
#> 325 325 1 -0.006011730  0.376475578 -0.106822439 -0.413651882
#> 326 326 1 -0.042008564  0.217421934 -0.378413058 -0.547837836
#> 327 327 1  0.635060888  0.754317026 -0.109560689  1.954058364
#> 328 328 1  0.239034131  0.052855749 -0.079522896  0.492841346
#> 329 329 1  0.422537023  0.174915147  0.241773743  1.172458648
#> 330 330 1  0.284671865  0.431889187  0.229234602  0.662268292
#> 331 331 1  0.359493463  0.214801389  0.317459716  0.939853175
#> 332 332 1  0.948817130  0.243055525  0.252055692  3.116937000
#> 333 333 1  0.115341158  0.256431441  0.032064265  0.035901056
#> 334 334 1  0.274945577  0.343076638 -0.109036899  0.624563240
#> 335 335 1  0.704788091 -0.032840472  0.054328127  2.214831547
#> 336 336 1  0.260198981  0.439392536  0.192862569  0.571604357
#> 337 337 1 -0.335101148  0.759863531  0.024384626 -1.629907939
#> 338 338 1  0.192878019  0.492076896 -0.229327505  0.320200648
#> 339 339 1  0.030205988  0.099016508  0.426378161 -0.275898937
#> 340 340 1 -0.117221371  0.293795608  0.266730899 -0.822125186
#> 341 341 1  0.372935992  0.644676278  0.150637020  0.987353851
#> 342 342 1 -0.147488182  0.448821678  0.378222126 -0.933723483
#> 343 343 1  0.183328961  0.361110976  0.150303964  0.287531473
#> 344 344 1  0.493090911  0.388244194  0.126842731  1.431887544
#> 345 345 1  0.039256843  0.425640024  0.219499523 -0.244580478
#> 346 346 1  0.485715233  0.427509312  0.501749326  1.406754336
#> 347 347 1  0.105806425  0.258805078  0.495630442  0.003418037
#> 348 348 1  0.475616113  0.387208091  0.473721451  1.369382042
#> 349 349 1  0.016548565  0.628092157  0.365574211 -0.328174518
#> 350 350 1 -0.188784274  0.194554909  0.668611882 -1.083890167
#> 351 351 1  0.276539168  0.008918170  0.189041496  0.633138515
#> 352 352 1  0.283927917  0.043228327 -0.050454272  0.658922907
#> 353 353 1  0.573944087  0.458151120 -0.044073324  1.729432761
#> 354 354 1  0.024333620  0.037439055 -0.208279467 -0.301199407
#> 355 355 1  0.581098783  0.267057437  0.383057996  1.758930953
#> 356 356 1  0.050562961  0.103049022  0.188391196 -0.202105025
#> 357 357 1  0.186284301  0.459461234 -0.512415525  0.294244285
#> 358 358 1 -0.024158859  0.357564798  0.151002502 -0.479121771
#> 359 359 1 -0.511738158  0.487751301  0.479303095 -2.279132598
#> 360 360 1 -0.160390728  0.108302224  0.402612853 -0.980320107
#> 361 361 1  0.047665212  0.290990600 -0.138946816 -0.215272239
#> 362 362 1 -0.181181047  0.574491504  0.182656407 -1.059724974
#> 363 363 1  0.103620657  0.261588766 -0.049830314 -0.007906894
#> 364 364 1  0.481023068  0.240397447  0.122244138  1.387674802
#> 365 365 1  0.084622720  0.300403221  0.226363176 -0.076569602
#> 366 366 1  0.239543233 -0.386850683  0.181515733  0.497478659
#> 367 367 1  0.524202796  0.099684916  0.751905413  1.551350676
#> 368 368 1  0.471735258  0.459144810  0.114890409  1.352713095
#> 369 369 1  0.138165506  0.251377569  0.221009116  0.121373462
#> 370 370 1  0.119796872  0.197909832  0.256714933  0.053860091
#> 371 371 1  0.239843262  0.301288821  0.441794628  0.498247593
#> 372 372 1  0.212269501  0.542161747  0.526204999  0.396203908
#> 373 373 1  0.202041021  0.220777628 -0.349741614  0.354086090
#> 374 374 1  0.192654410  0.064249238  0.406785657  0.324326571
#> 375 375 1  0.265661328  0.849469846  0.111361451  0.590179483
#> 376 376 1  0.434971573  0.168285166  0.331335588  1.218954595
#> 377 377 1  0.341826480  0.242059166  0.215614656  0.873893877
#> 378 378 1  0.415008248 -0.258625536  0.082996469  1.144884756
#> 379 379 1  0.167602625  0.468680187 -0.511384991  0.225196379
#> 380 380 1 -0.294988010  0.623316467 -0.184234900 -1.482555231
#> 381 381 1  0.146392523  0.749724751  0.324995302  0.151024179
#> 382 382 1  0.393108956  0.312343699 -0.369282842  1.059714677
#> 383 383 1  0.175601717  0.151230546 -0.121136928  0.257941802
#> 384 384 1  0.575030602  0.380555650 -0.332312759  1.731947520
#> 385 385 1 -0.238067070  0.360050662  0.288722631 -1.268700760
#> 386 386 1 -0.214140409  0.616846791 -0.104388867 -1.183331349
#> 387 387 1  0.325375570  0.781553034 -0.041010247  0.810104327
#> 388 388 1  0.322255474  0.482307532  0.127313576  0.800395893
#> 389 389 1  0.062011539  0.094474310 -0.024569651 -0.161044370
#> 390 390 1  0.355820035 -0.013995264  0.180092563  0.926090807
#> 391 391 1  0.512833875  0.106754209  0.110902267  1.505514591
#> 392 392 1  0.277631499  0.364813656  0.089742848  0.635609203
#> 393 393 1  0.332219410  0.033563705 -0.077576934  0.837225224
#> 394 394 1 -0.093532500  0.531568164  0.512283735 -0.733787599
#> 395 395 1  0.401359891  0.360425275  0.034890414  1.092471501
#> 396 396 1  0.411592670  0.231579824 -0.301318072  1.128637152
#> 397 397 1 -0.345055111  0.442713861  0.688672907 -1.661871617
#> 398 398 1  0.406604851  0.269065574  0.033174366  1.112091832
#> 399 399 1  0.009684093  0.168060062  0.008859430 -0.354397259
#> 400 400 1  0.274099402  0.245197339  0.188911135  0.623475185
#> 401 401 1  0.394597488  0.252582467 -0.319648844  1.065673503
#> 402 402 1 -0.032039515  0.355972338  0.539112969 -0.505930430
#> 403 403 1  0.418134409  0.519733628 -0.062413961  1.153438522
#> 404 404 1 -0.123388921  0.567887205 -0.237058785 -0.848658839
#> 405 405 1  0.719243494  0.097950308  0.474436152  2.270381847
#> 406 406 1  0.211027293  0.441919321 -0.731166871  0.384417948
#> 407 407 1  0.186079713  0.107542680 -0.073007827  0.297063687
#> 408 408 1  0.211821026  0.496254695  0.202438070  0.392748940
#> 409 409 1  0.153258566  0.453767065  0.123546806  0.176008474
#> 410 410 1  0.041910919  0.284771189 -0.293953659 -0.237438279
#> 411 411 1  0.076873770  0.408328170  0.071745326 -0.106416445
#> 412 412 1 -0.215094478  0.115452667  0.578790512 -1.181423188
#> 413 413 1  0.538258400  0.708540309  0.093396973  1.597704702
#> 414 414 1  0.144371436  0.272651133 -0.060317907  0.142574526
#> 415 415 1  0.438064776 -0.159395383 -0.068232617  1.228908063
#> 416 416 1  0.288709322  0.349647959 -0.101777101  0.675445377
#> 417 417 1  0.529714727 -0.025879698 -0.078850369  1.567125419
#> 418 418 1  0.209059898  0.033342483 -0.177398669  0.381558560
#> 419 419 1  0.371337287  0.042577348  0.304112636  0.984008836
#> 420 420 1  0.475510491  0.123457337  0.524328124  1.370015373
#> 421 421 1  0.363725147  0.325695281 -0.269345197  0.951698746
#> 422 422 1 -0.006648481  0.366241246 -0.039158555 -0.415574593
#> 423 423 1  0.473796470  0.456345321 -0.309852352  1.357813305
#> 424 424 1 -0.154030209  0.544280689 -0.157605702 -0.961341564
#> 425 425 1  0.155598151  0.018718225  0.208599936  0.186351024
#> 426 426 1  0.429148034 -0.161064471  0.360589447  1.198513166
#> 427 427 1  0.263932757  0.245679699  0.007963092  0.584833011
#> 428 428 1  0.304720070  0.044483394  0.114920092  0.736729014
#> 429 429 1 -0.133495153  0.737723560  0.262553620 -0.883498377
#> 430 430 1  0.350627584 -0.113090107  0.237602970  0.907518021
#> 431 431 1  0.293564708  0.361661842 -0.251522969  0.692463365
#> 432 432 1  0.390496602  0.342244703 -0.583695459  1.048706130
#> 433 433 1  0.469335968  0.462003784 -0.230239220  1.341789276
#> 434 434 1 -0.075999218  0.020105004 -0.334579901 -0.672631970
#> 435 435 1 -0.304599754  0.281046667  0.381301013 -1.513772301
#> 436 436 1  0.024201436  0.256567702  0.794543267 -0.296330101
#> 437 437 1  0.062940884  0.190792990  0.126114928 -0.156979144
#> 438 438 1  0.479578829  0.802273473 -0.475856343  1.377244624
#> 439 439 1  0.206193084  0.442025800  0.468547458  0.373683449
#> 440 440 1  0.553544706  0.632412747 -0.342551275  1.651806022
#> 441 441 1  0.280829473  0.385806524  0.126967822  0.647589342
#> 442 442 1  0.014790294  0.319638024  0.756838296 -0.331501153
#> 443 443 1  0.352518250  0.308389902 -0.101664856  0.911333000
#> 444 444 1  0.335581614  0.404324249  0.031923781  0.849282939
#> 445 445 1  0.278463223  0.424008984  0.293579391  0.639731280
#> 446 446 1  0.242438434  0.274235388  0.183607110  0.506376861
#> 447 447 1 -0.393847639  0.493552407 -0.191142886 -1.847526750
#> 448 448 1  0.115927644  0.560279300 -0.273465042  0.035419970
#> 449 449 1  0.176406430  0.328173809 -0.240804710  0.259719214
#> 450 450 1  0.062892285 -0.433501078 -0.011059390 -0.156262602
#> 451 451 1  0.403778700  0.037065128  0.265922846  1.103667997
#> 452 452 1  0.235249553  0.152192773 -0.406647175  0.476641431
#> 453 453 1  0.264838604  0.621256737 -0.192074378  0.585962150
#> 454 454 1  0.165624459  0.282731321  0.368935904  0.223627222
#> 455 455 1  0.592530840  0.105088986  0.339010267  1.801354614
#> 456 456 1  0.632295763  0.130563338 -0.297423977  1.944434691
#> 457 457 1  0.132566962  0.331665713 -0.195066459  0.097994647
#> 458 458 1  0.214809485  0.110957139  0.359148863  0.405778470
#> 459 459 1  0.508636953  0.134884822 -0.352036495  1.487179285
#> 460 460 1  0.672536574  0.352983296  0.319787218  2.096181981
#> 461 461 1  0.246096890 -0.213272654  0.297578662  0.521908242
#> 462 462 1  0.361513686  0.544187799  0.345384752  0.946580990
#> 463 463 1  0.724906724  0.438527224  0.012495512  2.287629323
#> 464 464 1  0.042222992  0.164276891  0.058077311 -0.233863286
#> 465 465 1  0.171005672  0.052406722  0.046969756  0.242229111
#> 466 466 1  0.242710566  0.341620079 -0.566924919  0.502738352
#> 467 467 1  0.550641780  0.338488122  0.384299151  1.646203974
#> 468 468 1  0.279122275 -0.007100794 -0.051802692  0.641296004
#> 469 469 1  0.572513612  0.103068453 -0.618714537  1.721706146
#> 470 470 1  0.113508330  0.690465654 -0.074957680  0.027303226
#> 471 471 1 -0.504212686  0.447001093  0.535641863 -2.250879558
#> 472 472 1  0.136867983  0.804204459 -0.026297229  0.113594468
#> 473 473 1  0.525701500  0.183512612  0.089431491  1.552722484
#> 474 474 1  0.016314914  0.338455635 -0.018235546 -0.330524416
#> 475 475 1  0.636733530  0.200440373 -0.150352194  1.961514542
#> 476 476 1  0.621667071  0.389034737  0.143602031  1.907073630
#> 477 477 1  0.124420041  0.523859128 -0.249419857  0.067042054
#> 478 478 1  0.127689472  0.546244808 -0.368568814  0.078353305
#> 479 479 1  0.227806761  0.159170529  0.231137107  0.452910642
#> 480 480 1  0.188323618  0.409877049  0.169920945  0.305969611
#> 481 481 1 -0.070744137  0.523724581  0.867571404 -0.647452176
#> 482 482 1  0.213946965  0.307527951  0.380822182  0.402181448
#> 483 483 1  0.111696129  0.455776310  0.709214566  0.025909626
#> 484 484 1  0.559785873  0.455480783  0.144728186  1.678247283
#> 485 485 1  0.431526468  0.383201108 -0.283958320  1.201980061
#> 486 486 1 -0.011772904  1.013954007  0.152849289 -0.435143833
#> 487 487 1  0.232782383  0.128294791  0.572672333  0.473409439
#> 488 488 1  0.725883954  0.205667410 -0.266652563  2.290219862
#> 489 489 1  0.091094547  0.333029917  0.231458382 -0.052715364
#> 490 490 1 -0.119800858  0.767453538 -0.012440133 -0.834613363
#> 491 491 1 -0.091042006  0.224075848  0.214458832 -0.725511963
#> 492 492 1  0.153442378  0.313998606  0.010343507  0.176398150
#> 493 493 1  0.058433986  0.066710933  0.124644771 -0.173300758
#> 494 494 1  0.618834530 -0.034485961  0.162034667  1.897877753
#> 495 495 1  0.087529520  0.465641239  0.431231089 -0.065064649
#> 496 496 1  0.193460308 -0.100862776  0.111113904  0.326000170
#> 497 497 1  0.267725429  0.540655402  0.325004548  0.599922144
#> 498 498 1  0.432941533  0.215426945  0.166717790  1.210346303
#> 499 499 1  0.245263332  0.544315810  0.020264479  0.515104070
#> 500 500 1 -0.162453328  0.284172938 -0.053952611 -0.991136176