Declare measurement procedure

This function adds measured data columns that can be functions of unmeasured data columns.

Usage

declare_measurement(..., handler = measurement_handler, label = NULL)

measurement_handler(data, ...)

Arguments

...

arguments to be captured, and later passed to the handler

handler

a tidy-in, tidy-out function

label

a string describing the step

data

A data.frame.

Value

A function that returns a data.frame.

Details

It is also possible to include measured variables in your declare_model call or to add variables using declare_step. However, putting latent variables in declare_model and variables-as-measured in declare_measurement helps communicate which parts of your research design are in M and which parts are in D.

Examples

# declare_measurement in use
## Two-arm randomized experiment
design <-
  declare_model(
    N = 500,
    X = rep(c(0, 1), each = N / 2),
    U = rnorm(N, sd = 0.25),
    potential_outcomes(Y ~ 0.2 * Z + X + U)
  ) +
  declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(S = complete_rs(N = N, n = 200)) +
  declare_assignment(Z = complete_ra(N = N, m = 100)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z)) +
  declare_estimator(Y ~ Z, inquiry = "ATE")
  
run_design(design)
#>   inquiry estimand estimator term   estimate  std.error statistic   p.value
#> 1     ATE      0.2 estimator    Z 0.02291571 0.08087335 0.2833531 0.7772022
#>     conf.low conf.high  df outcome
#> 1 -0.1365679 0.1823994 198       Y

# Reveal potential outcomes according to treatment assignment
design <-
  declare_model(N = 100,
                potential_outcomes(Y ~ rbinom(
                  N, size = 1, prob = 0.1 * Z + 0.5
                ))) +
  declare_assignment(Z = complete_ra(N, m = 50)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z))
  
head(draw_data(design))
#>    ID Y_Z_0 Y_Z_1 Z Y
#> 1 001     0     1 1 1
#> 2 002     0     0 0 0
#> 3 003     1     1 1 1
#> 4 004     1     0 1 0
#> 5 005     0     0 0 0
#> 6 006     0     1 1 1

# Generate observed measurement from a latent value
design <- 
  declare_model(N = 100, latent = runif(N)) +
  declare_measurement(observed = rbinom(N, 1, prob = latent))
  
head(draw_data(design))
#>    ID    latent observed
#> 1 001 0.8554179        1
#> 2 002 0.6752466        1
#> 3 003 0.5269439        1
#> 4 004 0.3955966        1
#> 5 005 0.1413085        0
#> 6 006 0.1830983        1

if(require("psych")) {

  # Index creation
  
  library(psych)
  
  design <-
    declare_model(
      N = 500,
      X = rep(c(0, 1), each = N / 2),
      Y_1 = 0.2 * X + rnorm(N, sd = 0.25),
      Y_2 = 0.3 * X + 0.5 * rnorm(N, sd = 0.50),
      Y_3 = 0.1 * X + 0.4 * rnorm(N, sd = 0.75)) +
    declare_measurement(
      index = fa(
        r = cbind(Y_1, Y_2, Y_3),
        nfactors = 1,
        rotate = "varimax"
      )$scores
    )
  
  draw_data(design)
}
#> Loading required package: psych
#>      ID X           Y_1          Y_2           Y_3           MR1
#> 1   001 0 -0.0711851459 -0.342287729  0.1845648072 -0.6836113830
#> 2   002 0  0.0303726277  0.256863266 -0.3870754270 -0.2532872743
#> 3   003 0 -0.3552338225 -0.136984191  0.4922903784 -0.6059223568
#> 4   004 0  0.0841552269  0.631810865 -0.3324421071  0.2713721634
#> 5   005 0 -0.0565980652 -0.188426429  0.6191836361 -0.2102122730
#> 6   006 0 -0.5463059462 -0.318410203 -0.2113643938 -1.5043508970
#> 7   007 0  0.0655355159  0.342037724  0.2486451797  0.2987888102
#> 8   008 0 -0.1973674653  0.184853512  0.2409842843 -0.2101371482
#> 9   009 0  0.0573134367 -0.081006975 -0.1630925049 -0.4550607956
#> 10  010 0 -0.2870543501 -0.380363010  0.4041825763 -0.8524462016
#> 11  011 0 -0.0023161131  0.012661109 -0.1356342723 -0.4058854641
#> 12  012 0 -0.1224885224  0.153224509  0.6695776482  0.1253855747
#> 13  013 0  0.6023252515 -0.437681126 -0.1508108962 -0.1711635589
#> 14  014 0  0.4424084292 -0.410382864 -0.0935728242 -0.3021291563
#> 15  015 0  0.0462677317 -0.055825728  0.2665124265 -0.1617326041
#> 16  016 0 -0.0254969821  0.306185068  0.0277778794  0.0019272965
#> 17  017 0  0.1338660915  0.138852806  0.0219950519  0.0078504490
#> 18  018 0  0.0954012372 -0.229986399  0.4677568939 -0.1662592747
#> 19  019 0 -0.4880787443 -0.195264315 -0.2885412477 -1.3433043499
#> 20  020 0 -0.2020671072 -0.138949793 -0.3802740906 -0.9837439744
#> 21  021 0  0.1008427230 -0.610803227 -0.2963350642 -1.0841371591
#> 22  022 0 -0.0501069098  0.140938506  0.7164125578  0.2319319445
#> 23  023 0 -0.1395411537 -0.115462241  0.2241528911 -0.4874290657
#> 24  024 0 -0.0709685739 -0.020906659  0.1933744652 -0.3156227006
#> 25  025 0  0.3782846176  0.219666478  0.1969915739  0.5163242492
#> 26  026 0 -0.4299557490 -0.149558121  0.2873378623 -0.8459570503
#> 27  027 0  0.0793679079 -0.098636933 -0.3644412241 -0.5781172894
#> 28  028 0  0.1366698489 -0.437788523  0.0686944817 -0.6078977647
#> 29  029 0 -0.0174783608  0.004652866  0.0167272458 -0.3349174253
#> 30  030 0 -0.2723749495 -0.575302728 -0.0079493814 -1.3211371158
#> 31  031 0  0.0263624401 -0.042085339  0.3881976494 -0.0920652474
#> 32  032 0 -0.5018631450  0.071513607  0.2726423457 -0.6958872394
#> 33  033 0 -0.3007015960 -0.189726799  0.2821510390 -0.7338485039
#> 34  034 0  0.5808254462 -0.046296346  0.6041266324  0.7327117916
#> 35  035 0 -0.2141574884 -0.225159281 -0.4321311010 -1.1295254723
#> 36  036 0 -0.2659503190 -0.342280295  0.0456378630 -1.0159077746
#> 37  037 0  0.1649542375  0.238628636  0.1002235321  0.2096436638
#> 38  038 0 -0.1099155676 -0.257851218  0.0223714973 -0.7418808472
#> 39  039 0  0.1380971115 -0.415083741 -0.1004439744 -0.6902739346
#> 40  040 0  0.3573634745  0.314575296  0.6173572991  0.8699204519
#> 41  041 0 -0.1180036914  0.128147537  0.3398451944 -0.1111926893
#> 42  042 0  0.4315495388 -0.302997337 -0.0478118594 -0.1649856184
#> 43  043 0 -0.0294006704  0.030288929  0.3813271664 -0.0843377863
#> 44  044 0  0.2862012714  0.098949394 -0.1043835505  0.0703365790
#> 45  045 0 -0.0080050091  0.134351978  0.2453814225 -0.0287091941
#> 46  046 0 -0.1697414571 -0.151450507 -0.2551991865 -0.8764897315
#> 47  047 0 -0.1823218250  0.189280715 -0.0436395091 -0.3710866929
#> 48  048 0  0.1668963233 -0.354687345  0.2314651070 -0.3711131190
#> 49  049 0  0.2725762947 -0.007523067  0.2338727938  0.1529018206
#> 50  050 0  0.2822248727 -0.110582116  0.0522616180 -0.0690037422
#> 51  051 0 -0.0302115473 -0.032078080  0.5790590380 -0.0273213154
#> 52  052 0 -0.0457538934  0.003391563 -0.1481670279 -0.4784692769
#> 53  053 0  0.2520334650 -0.192484679  0.3147622913 -0.0285073998
#> 54  054 0 -0.2481922974 -0.463415261  0.0743199951 -1.1116664203
#> 55  055 0 -0.1431178295  0.450940179  0.1919586562  0.1252349346
#> 56  056 0 -0.1181823984  0.496530428 -0.2823837236 -0.1001277263
#> 57  057 0 -0.0137684404 -0.136123720 -0.1415800094 -0.5915744405
#> 58  058 0 -0.1671253270 -0.246963030 -0.5534196003 -1.1742874732
#> 59  059 0 -0.2799181853  0.041226719 -0.0565490393 -0.6675848337
#> 60  060 0 -0.0089071650 -0.059127787 -0.1195350740 -0.4845001110
#> 61  061 0 -0.5031008097 -0.073886739 -0.0954357223 -1.0999896008
#> 62  062 0  0.2796582469 -0.148428913  0.0175312381 -0.1373566117
#> 63  063 0 -0.1379548659  0.122291238  0.0970235129 -0.3001235671
#> 64  064 0  0.1815171627 -0.108152423 -0.1846640500 -0.3451920473
#> 65  065 0  0.3082471342 -0.264977377  0.1297581572 -0.1602832972
#> 66  066 0 -0.0560541320  0.052112918 -0.4480379844 -0.6309331509
#> 67  067 0  0.1989341781  0.133916275  0.6869310968  0.5145635874
#> 68  068 0  0.2870642271  0.209833410 -0.0322568104  0.2431005387
#> 69  069 0 -0.1076486800  0.022722794  0.3621124112 -0.2026220543
#> 70  070 0 -0.0216898873  0.066078256  0.1218826452 -0.2027468005
#> 71  071 0  0.3777274453  0.434829756  0.0356092181  0.6532978250
#> 72  072 0  0.0048825335 -0.189628372  0.2178075719 -0.3955045010
#> 73  073 0  0.0517204851  0.082645426 -0.2929076884 -0.3618922055
#> 74  074 0  0.0543250246 -0.140577560 -0.4186102649 -0.6916397717
#> 75  075 0 -0.2274247544 -0.333874821 -0.0715953293 -1.0345886610
#> 76  076 0 -0.1155179090 -0.198739145 -0.7139163714 -1.1599168374
#> 77  077 0 -0.1847329578 -0.227334820  0.0161349243 -0.8045837455
#> 78  078 0 -0.1672260987  0.333350756 -0.1977519521 -0.2900130497
#> 79  079 0 -0.2048123568  0.093528716  0.3102595667 -0.2773217196
#> 80  080 0  0.2043944961  0.170687429 -0.0319485944  0.0964117182
#> 81  081 0  0.1553918779  0.005829733 -0.3394262947 -0.3496867640
#> 82  082 0  0.3994464801  0.079077321 -0.7187407495 -0.2097874130
#> 83  083 0 -0.0321254059  0.128763100 -0.2095884225 -0.3601504010
#> 84  084 0  0.0521757479 -0.207854702  0.1965356648 -0.3710282841
#> 85  085 0  0.2776987254 -0.157268451  0.7401841171  0.3190582967
#> 86  086 0 -0.1207415238 -0.066379789  0.4011232977 -0.2939598918
#> 87  087 0 -0.3375291216  0.216196423  0.1370398043 -0.4165468680
#> 88  088 0 -0.2743256100 -0.072822538 -0.5431106988 -1.1047444300
#> 89  089 0  0.0059689482  0.076417656 -0.0292951835 -0.2547823942
#> 90  090 0  0.0828229701 -0.332249236 -0.0572298308 -0.6376645198
#> 91  091 0 -0.0793200875 -0.048950049  0.3443336814 -0.2596632853
#> 92  092 0 -0.1144552078  0.435865366  0.4280178795  0.2970340816
#> 93  093 0 -0.2050161911 -0.139798743  0.0745246360 -0.6933245742
#> 94  094 0 -0.0955320684  0.031006362 -0.1828176181 -0.5317389318
#> 95  095 0  0.2453200779 -0.095937226 -0.2192585327 -0.2745404367
#> 96  096 0  0.3103369225  0.263955910  0.5616995019  0.7183210723
#> 97  097 0 -0.3225670798  0.221692631 -0.1280872345 -0.5637480572
#> 98  098 0  0.0370412044 -0.403564019 -0.3313362705 -0.9527421310
#> 99  099 0  0.2862848189 -0.031507409 -0.2910757158 -0.1976195915
#> 100 100 0  0.1146529054 -0.214170146  0.3388275821 -0.2081461862
#> 101 101 0 -0.0004401573 -0.166288810  0.1419378224 -0.4250523695
#> 102 102 0  0.2015556837  0.203762048 -0.2468927234 -0.0093040692
#> 103 103 0 -0.2381711680  0.209474873 -0.2106398406 -0.5261236553
#> 104 104 0  0.2372464442 -0.095202445  0.5834863919  0.2370147904
#> 105 105 0 -0.0730998146 -0.075490567  1.1617989767  0.2484921537
#> 106 106 0  0.0772960386 -0.019410887 -0.4018039135 -0.5156915397
#> 107 107 0 -0.5490274932 -0.011484512 -0.1410370599 -1.1163896411
#> 108 108 0 -0.3841564281  0.009027583 -0.0878932872 -0.8538017937
#> 109 109 0  0.1560720318 -0.035904536  0.1450251997 -0.0815712260
#> 110 110 0 -0.2302694140  0.199456661 -0.3737314181 -0.6333856374
#> 111 111 0  0.1620721762 -0.166358087  0.1888261318 -0.1926384771
#> 112 112 0 -0.4900328282  0.138198558 -0.0604844171 -0.8221730347
#> 113 113 0 -0.3336095175  0.051881039 -0.1387215035 -0.7756535511
#> 114 114 0  0.0756732358  0.004571091 -0.0607322149 -0.2694317909
#> 115 115 0  0.2092757342  0.103880560  0.4265865504  0.3246964402
#> 116 116 0 -0.1163875807  0.132409490 -0.1985765500 -0.4536741060
#> 117 117 0  0.2223069425 -0.244279350  0.1893152691 -0.2051936452
#> 118 118 0  0.2446646052  0.335990362  0.6110834515  0.7498383548
#> 119 119 0  0.0512367800 -0.077571182  0.0385365152 -0.3279433805
#> 120 120 0  0.0980853073  0.042169471 -0.1025675520 -0.2263528233
#> 121 121 0  0.1106054010  0.073901333 -0.0324886904 -0.1295796077
#> 122 122 0  0.3646613877  0.159432188  0.4473331724  0.5939412892
#> 123 123 0  0.4074772406 -0.094627309  0.4050082417  0.3335494570
#> 124 124 0  0.4973836956 -0.224115079 -0.1413288945 -0.0549393429
#> 125 125 0  0.2392339384  0.066619110 -0.1749059262 -0.0702314186
#> 126 126 0 -0.2477883595 -0.439919428  0.0165496096 -1.1221759155
#> 127 127 0  0.0554139495  0.610332204  0.0307036617  0.4470249890
#> 128 128 0  0.0924315419  0.348962161 -0.5719201280 -0.1922952467
#> 129 129 0  0.0844291047 -0.140414275  0.3632639720 -0.1467961614
#> 130 130 0  0.0258745769  0.206436645  0.2120329623  0.0729800465
#> 131 131 0  0.1373997818 -0.044958924 -0.1131226905 -0.2824567819
#> 132 132 0  0.0206050352  0.144753605  0.5261815930  0.2007461899
#> 133 133 0 -0.4122474741 -0.212001137  0.3285303528 -0.8675509486
#> 134 134 0  0.0697428292 -0.195928140  0.0458956251 -0.4334756827
#> 135 135 0  0.4219089933 -0.225162191  0.0944654560  0.0030005294
#> 136 136 0  0.1469610821 -0.105721217 -0.1952756405 -0.3923059484
#> 137 137 0  0.0847004475  0.543484612  0.0329230866  0.4095834772
#> 138 138 0  0.2667150380 -0.119504983  0.2351468991  0.0203036128
#> 139 139 0 -0.1457100070  0.188822872 -0.0079511309 -0.3029256872
#> 140 140 0  0.0299858483 -0.003105637 -0.0719355738 -0.3421562740
#> 141 141 0 -0.1389600320 -0.198854575 -0.0934169713 -0.7866576219
#> 142 142 0  0.1843027517  0.097309117  0.0473174557  0.0401978624
#> 143 143 0  0.0182861703  0.104178440  0.0057605680 -0.1854550926
#> 144 144 0 -0.2665027323  0.156595591 -0.0699339913 -0.5296350777
#> 145 145 0  0.0600914532 -0.274581219 -0.3486625408 -0.7900342119
#> 146 146 0  0.3381920251 -0.431928473 -0.4163711928 -0.6653945337
#> 147 147 0  0.0968917610 -0.560191507 -0.2918928148 -1.0291589607
#> 148 148 0  0.2346150139  0.022395619  0.0983860235  0.0515051458
#> 149 149 0 -0.0939349358  0.195291485  0.4143002616  0.0426677984
#> 150 150 0  0.5135308616 -0.100660860 -0.0276307570  0.1779573939
#> 151 151 0  0.1524074919  0.024468199  0.0174726540 -0.1008713504
#> 152 152 0  0.1219361364  0.214428418 -0.0328380651  0.0425727937
#> 153 153 0 -0.0053352464  0.086170010 -0.0092508862 -0.2448500006
#> 154 154 0 -0.1987521765  0.263983293 -0.1734840104 -0.3916060112
#> 155 155 0 -0.1341837233  0.263257332 -0.0516616289 -0.2331067630
#> 156 156 0 -0.5143661418 -0.132044148 -0.1088298287 -1.1881951506
#> 157 157 0  0.1463114091  0.131141658 -0.3328648923 -0.2155695640
#> 158 158 0  0.2498771262 -0.043203180 -0.0698408864 -0.1125422311
#> 159 159 0 -0.2196678172  0.347358400 -0.2448773310 -0.3700145805
#> 160 160 0 -0.2105492847  0.134165624 -0.1912863124 -0.5640507918
#> 161 161 0  0.2774456318  0.404983038 -0.4842168183  0.1577562523
#> 162 162 0 -0.2638734482  0.153371969 -0.1632059357 -0.5905053522
#> 163 163 0 -0.1413005377 -0.245594447  0.0510318036 -0.7485071822
#> 164 164 0  0.1755503299  0.507936202 -0.2411669963  0.3046967324
#> 165 165 0 -0.4035397310  0.223842768  0.5910388413 -0.1954902117
#> 166 166 0 -0.1543720770  0.130169303  0.3620937465 -0.1397038335
#> 167 167 0  0.1662326605 -0.118492879 -0.3223264918 -0.4651509677
#> 168 168 0 -0.0881969715 -0.483044361 -0.2211140838 -1.1264901128
#> 169 169 0 -0.2227794805 -0.060198110  0.2164620608 -0.5336704017
#> 170 170 0 -0.1738777260  0.161438384  0.0120456735 -0.3558238581
#> 171 171 0 -0.2872950783  0.040179139  0.1847404198 -0.5214052505
#> 172 172 0  0.3495145725  0.202803615 -0.2484665509  0.1725730170
#> 173 173 0  0.2612488557  0.039429690 -0.0070715305  0.0353961494
#> 174 174 0  0.0546835606 -0.024814183  0.3488159238 -0.0629436342
#> 175 175 0 -0.2857253270 -0.166894873 -0.5050740415 -1.2002069227
#> 176 176 0  0.3112809312 -0.363481592 -0.1774751360 -0.4667784908
#> 177 177 0 -0.4040996963 -0.032503550 -0.0054765432 -0.8719143446
#> 178 178 0  0.4909626920 -0.307135147 -0.3572910887 -0.2965391000
#> 179 179 0  0.0827506322  0.074686122 -0.4008206400 -0.4022801526
#> 180 180 0  0.2087533857  0.013442567  0.2397803190  0.1009901449
#> 181 181 0 -0.0403887774  0.196338055 -0.2153849248 -0.2980692030
#> 182 182 0  0.1252777573 -0.119430829  0.5762696786  0.0658160883
#> 183 183 0  0.1784075350 -0.640600080 -0.3565588158 -1.0603224800
#> 184 184 0 -0.2357382846 -0.086443944 -0.0901791393 -0.7782753041
#> 185 185 0 -0.0465656585  0.208873340 -0.0243771994 -0.1677171026
#> 186 186 0 -0.1868916817  0.126970956  0.2642259937 -0.2472325405
#> 187 187 0 -0.1083378624  0.033574221  0.2659166305 -0.2536612590
#> 188 188 0 -0.1233745345  0.015817799 -0.4785551652 -0.7753222550
#> 189 189 0  0.2399623504 -0.308282306 -0.1109352843 -0.4501155600
#> 190 190 0 -0.2119320300  0.169209330 -0.0749525998 -0.4508274025
#> 191 191 0  0.3397791388  0.008417233 -0.1275126663  0.0199771049
#> 192 192 0  0.4827431053 -0.051632900 -0.2320669251  0.0622755676
#> 193 193 0  0.5593447821  0.211467407 -0.0221107501  0.5900870601
#> 194 194 0 -0.0537742751 -0.348888141 -0.2675720390 -0.9627125807
#> 195 195 0 -0.4594839759  0.237510284 -0.0718607823 -0.6797027308
#> 196 196 0  0.0392008815  0.073989331 -0.2229544188 -0.3418288440
#> 197 197 0 -0.0974908450  0.296221136 -0.7109131183 -0.5780287506
#> 198 198 0  0.0446460143 -0.171569343 -0.1349009596 -0.5545328096
#> 199 199 0 -0.1803109530 -0.101908152 -0.2900299633 -0.8564234868
#> 200 200 0  0.0665561702 -0.273871703 -0.6267448445 -0.9615975347
#> 201 201 0 -0.1177876961  0.056497199 -0.1469902670 -0.5074571262
#> 202 202 0  0.3517464750  0.181188686  0.1464717689  0.4072104276
#> 203 203 0  0.2273076371  0.751332467 -0.0079363015  0.7945200617
#> 204 204 0 -0.0862197999  0.117476252  0.3367019967 -0.0857306517
#> 205 205 0 -0.3911270781  0.050647007 -0.3279669210 -0.9713327012
#> 206 206 0 -0.1545759686 -0.210216795  0.0856382202 -0.7027145921
#> 207 207 0 -0.2397026547  0.039954020  0.1662225349 -0.4744937629
#> 208 208 0 -0.1382327381  0.147423673 -0.1889884038 -0.4577050509
#> 209 209 0  0.0941213458 -0.020970192 -0.5408641792 -0.5867396206
#> 210 210 0 -0.1139270932  0.290576659 -0.5145548188 -0.4774401633
#> 211 211 0  0.3086607624  0.010309728  0.2241731939  0.2115650595
#> 212 212 0  0.2507240743 -0.181266244 -0.0584523603 -0.2596155093
#> 213 213 0 -0.4535788086  0.275958852  0.3674305928 -0.3440689377
#> 214 214 0  0.1188103378 -0.201479876 -0.4297993578 -0.6873155200
#> 215 215 0  0.0793993292 -0.414757656  0.0437339654 -0.6693607684
#> 216 216 0  0.5179564829 -0.666127880  0.0991906838 -0.3712096244
#> 217 217 0  0.1645220921 -0.369353360  0.1094238160 -0.4697572934
#> 218 218 0  0.0848407137 -0.102674133 -0.0538059617 -0.3743405522
#> 219 219 0 -0.3535542577 -0.226234993 -0.3329653966 -1.2397363493
#> 220 220 0 -0.0278313308 -0.025134013 -0.3267014990 -0.6041360462
#> 221 221 0  0.1738949793  0.276238576 -0.1166577056  0.1224269479
#> 222 222 0  0.1157767807 -0.323026945 -0.1520274364 -0.6477987788
#> 223 223 0  0.1873296622  0.030916143  0.1171511140  0.0144798139
#> 224 224 0  0.0670282067  0.086759566 -0.4478551981 -0.4387432992
#> 225 225 0  0.2834249940 -0.108337631 -0.4395711709 -0.3840511320
#> 226 226 0  0.2246396314  0.151438956  0.1946542176  0.2469083462
#> 227 227 0  0.2783265557 -0.235596336 -0.2919177992 -0.4379476581
#> 228 228 0  0.0066231119 -0.306648051  0.0139422707 -0.6574054222
#> 229 229 0  0.0609906349  0.245804904  0.0926711801  0.0835553080
#> 230 230 0 -0.1097827709  0.204086819 -0.4393908124 -0.5209479604
#> 231 231 0 -0.4691354306 -0.281333804  0.2088817112 -1.0940039820
#> 232 232 0 -0.5024999590  0.040250968  0.3244967029 -0.6982537764
#> 233 233 0  0.0778867648  0.411498303 -0.5363545333 -0.1168673819
#> 234 234 0 -0.1406264423  0.184624969 -0.5161151264 -0.6309955577
#> 235 235 0  0.0340383059 -0.011114398 -0.3912597857 -0.5532943371
#> 236 236 0 -0.0012900769 -0.230078634  0.2538147545 -0.4253839880
#> 237 237 0  0.2956741108  0.139808936 -0.0033194616  0.1937032811
#> 238 238 0 -0.0589908952 -0.274216415  0.9813528753 -0.0748707049
#> 239 239 0  0.0713279509  0.046449481 -0.1993032970 -0.3175585273
#> 240 240 0  0.2725319959  0.130001381  0.1937419815  0.2817203929
#> 241 241 0 -0.0192122320  0.306524592  0.2671955325  0.1654424172
#> 242 242 0 -0.2132439918  0.395487019  0.2579614629  0.0183930282
#> 243 243 0  0.1829252210 -0.259218104 -0.1640155144 -0.5002065252
#> 244 244 0 -0.2599345250  0.228536212  0.0683524401 -0.3507229760
#> 245 245 0  0.0520092141 -0.144177310 -0.9931168460 -1.0712751268
#> 246 246 0 -0.0481296162 -0.175138620  0.1409444256 -0.4949641324
#> 247 247 0 -0.0092007686  0.283411946  0.1715571611  0.0898132096
#> 248 248 0  0.1770977907  0.328452403 -0.2374589308  0.1068555363
#> 249 249 0 -0.5179826776  0.052945567 -0.3407756151 -1.1347900472
#> 250 250 0  0.0537534559 -0.093260667 -0.2652978737 -0.5395939262
#> 251 251 1 -0.2927007191  0.467349691 -0.1691387728 -0.2765339578
#> 252 252 1  0.4068218611  0.733597379  0.6743854082  1.4404021662
#> 253 253 1  0.4334642059  0.454841393  0.0233797377  0.7372105023
#> 254 254 1  0.5149467013  0.212706140 -0.4023788570  0.2895838884
#> 255 255 1 -0.3141806232  0.491224787  0.2016054943 -0.0358361124
#> 256 256 1  0.4464848291  0.668983205  0.3103320019  1.1807658265
#> 257 257 1  0.0927607380  0.289328977 -0.0313711800  0.0916147593
#> 258 258 1  0.2129650696  0.862133501  0.2151939047  1.0462442691
#> 259 259 1  0.2660878337 -0.091254019 -0.0458138523 -0.1309226718
#> 260 260 1 -0.1020678190  0.643350991 -0.0877260356  0.2115694702
#> 261 261 1  0.3072494753  0.157577038 -0.3033439583  0.0334749437
#> 262 262 1 -0.0570729512  0.012289961  0.2484633394 -0.2252136835
#> 263 263 1 -0.4101885891  0.417973667 -0.6430852097 -0.7857044586
#> 264 264 1  0.5433418996 -0.223828093  0.0373865945  0.1184686069
#> 265 265 1  0.0176274798  0.286977293  0.2841086255  0.2002042586
#> 266 266 1  0.5217780298 -0.339727181  0.1267681619  0.0190910325
#> 267 267 1  0.3606811583  0.330939485  0.1690823940  0.6016679923
#> 268 268 1  0.0775171080  0.458681105  0.0812846291  0.3365023016
#> 269 269 1  0.0858974598  0.672123911  0.0077275638  0.5396262990
#> 270 270 1  0.6606097489  0.095280402  0.0471203002  0.6300428873
#> 271 271 1  0.3885387299  0.469583845  0.0696803581  0.7279911115
#> 272 272 1  0.3267264579  0.255477238 -0.4186664136  0.0931549274
#> 273 273 1  0.5090586975  0.700798328  0.1366818879  1.1817566992
#> 274 274 1  0.0499898659  0.369762342  0.3263578212  0.3611025350
#> 275 275 1 -0.0059414353  0.131772009  0.4121235676  0.0791216781
#> 276 276 1  0.2161970153  0.377348639 -0.1488373695  0.2680414811
#> 277 277 1  0.4778286718  0.392538982 -0.1026471296  0.6404396811
#> 278 278 1  0.2489316475  0.673465622  0.0780299420  0.7894678887
#> 279 279 1  0.0693306031  0.798912635  0.5512100356  1.0144161952
#> 280 280 1  0.2558488114  0.700084420  0.3945093426  1.0333628532
#> 281 281 1  0.3057717357  0.711455626  0.2238389726  0.9975280527
#> 282 282 1  0.3010248686 -0.047945602  0.0753939387  0.0399334604
#> 283 283 1  0.2228610841  0.304800045  0.0383059672  0.3160148967
#> 284 284 1  0.5433124056 -0.239254020  0.4214059259  0.3501819147
#> 285 285 1  0.6378312529  0.271090830 -0.4213616757  0.4958330347
#> 286 286 1  0.2341256177  0.418602363  0.2344173209  0.5854325861
#> 287 287 1  0.4505630403  0.342757311  0.0831276031  0.6709804790
#> 288 288 1  0.2823766210  0.511738160  0.2371302150  0.7520979924
#> 289 289 1 -0.2134799835  0.450615077  0.0298544764 -0.0677844470
#> 290 290 1  0.1501102885  0.156707014  0.0652587569  0.0762266769
#> 291 291 1  0.1381080418  0.936412465  0.0419160442  0.9244212327
#> 292 292 1  0.4413763864  0.150006325  0.0513750406  0.4218434364
#> 293 293 1  0.4886726710  0.282647601 -0.0072686941  0.5920167284
#> 294 294 1  0.2401518325  0.630147217 -0.0009251878  0.6785360477
#> 295 295 1 -0.0203834357 -0.069207532  0.0291850900 -0.4136443976
#> 296 296 1  0.3425844448  0.201143347 -0.1434179960  0.2302341050
#> 297 297 1  0.2778839140  0.416076910  0.2844942730  0.6694851673
#> 298 298 1  0.6667028559 -0.049840517  0.1790978495  0.5597729032
#> 299 299 1  0.1654502238  0.302924513  0.2817409037  0.4004395930
#> 300 300 1  0.2520428095  0.236918210 -0.2322051581  0.1003489922
#> 301 301 1  0.3186049156  0.554220052 -0.0248628300  0.6750339350
#> 302 302 1 -0.1473804298  0.514798004 -0.0996865986  0.0026645159
#> 303 303 1  0.3612665268  0.296123185  0.3855259846  0.7035926791
#> 304 304 1  0.0830738097  0.474699564  0.0387630418  0.3338698065
#> 305 305 1 -0.1127316274  0.148354735  0.4668991177  0.0005482163
#> 306 306 1  0.1700682805  0.515188739  0.3700325816  0.7025543770
#> 307 307 1  0.1863349269  0.586502284  0.1551360952  0.6636983474
#> 308 308 1 -0.0736620587  0.484270557 -0.4434020664 -0.1630368140
#> 309 309 1 -0.2073900725 -0.062703089  0.0928703729 -0.5975341759
#> 310 310 1  0.2154884752 -0.062194453  0.1235116543 -0.0512601289
#> 311 311 1 -0.1446357430  0.397302108 -0.7493259265 -0.5477123181
#> 312 312 1  0.8223344382  0.280191519  0.0468943681  1.0392751434
#> 313 313 1  0.0465300632  0.742509760  0.2337174965  0.7165651417
#> 314 314 1  0.3949441390  0.063532479  0.2440388300  0.3916907571
#> 315 315 1  0.0581197621  0.156016380  0.3130678816  0.1218260622
#> 316 316 1  0.4783204102  0.131091781  0.1130296769  0.4864730082
#> 317 317 1  0.3625156763  0.312531431 -0.1693783753  0.3636437713
#> 318 318 1  0.4099303488  0.245474460  0.1812441228  0.5745280254
#> 319 319 1  0.1612407699  0.257691148 -0.0042023157  0.1587538206
#> 320 320 1  0.3909094992  0.006939977  0.0809994840  0.2171591233
#> 321 321 1  0.0499434037  0.351683523  0.2736590766  0.3064933617
#> 322 322 1  0.2580757660  0.163164261  0.3072252995  0.3747200096
#> 323 323 1  0.6024978186  0.369683328  0.3261651172  1.0478977345
#> 324 324 1  0.1021883884  0.175251411  0.0988362452  0.0593100300
#> 325 325 1 -0.1201548587  0.064528610  0.4119429119 -0.1387558938
#> 326 326 1  0.2721326679  0.256589757  0.0013487177  0.2990015124
#> 327 327 1  0.5477584281  0.470222295  0.3353200956  1.0990191373
#> 328 328 1  0.5450202857  0.344628184  0.0248653959  0.7527429531
#> 329 329 1  0.0019440701  0.722030757  0.0894403450  0.5444603869
#> 330 330 1  0.3130359049  0.269324797  0.3736107506  0.6057053494
#> 331 331 1  0.3605735924  0.275442123 -0.6406176766  0.0137432451
#> 332 332 1  0.3557549892  0.309309575  0.0645656713  0.5033752516
#> 333 333 1 -0.0040610075  0.709558586 -0.2729837027  0.2878285532
#> 334 334 1  0.3016109534  0.522407477  0.3657226023  0.8714546726
#> 335 335 1  0.0350653098  0.330646331  0.3488346171  0.3130658522
#> 336 336 1  0.3230888932  0.812995706  0.5453990926  1.3420422712
#> 337 337 1  0.1267551201  0.076857029  0.7508999424  0.4020404390
#> 338 338 1  0.3200233397  0.126144125  0.5204464316  0.5483718561
#> 339 339 1 -0.3303166245  0.681480326  0.1922808322  0.1523547836
#> 340 340 1  0.0877316273  0.212382564  0.0962987857  0.0815124217
#> 341 341 1  0.3408379508  0.305039853 -0.0832914458  0.3840977210
#> 342 342 1  0.3716914855  0.325654388  0.3003650161  0.6945727929
#> 343 343 1  0.3611444403  0.406240598  0.2553208598  0.7430090640
#> 344 344 1 -0.2101247145  0.717207619 -0.1115504442  0.1449440402
#> 345 345 1  0.1199218924 -0.426465461  0.4332998511 -0.3794371746
#> 346 346 1  0.6007089821  0.058335188  0.0224682842  0.4979522915
#> 347 347 1  0.0162399011  0.562815089  0.0573906289  0.3621038274
#> 348 348 1  0.0704001555  0.129499878  0.0138907608 -0.0868581614
#> 349 349 1  0.1371505551  0.339671833  0.6297026478  0.6323770244
#> 350 350 1  0.5468654944  0.649272819  0.3764530152  1.3262759542
#> 351 351 1  0.2421428791  0.187851707 -0.0964483868  0.1208402534
#> 352 352 1  0.4206124645  0.610622185 -0.0457160113  0.8518773069
#> 353 353 1  0.2113082807  0.201807771  0.0717184727  0.2073147778
#> 354 354 1  0.0684767105  0.373341098  0.0505231087  0.2091779813
#> 355 355 1 -0.0347408891  0.110466425  0.1427421311 -0.1554437627
#> 356 356 1  0.0522584994 -0.012691757  0.3662243391 -0.0410109191
#> 357 357 1  0.8103827475  0.057728201  0.0369148978  0.7673568302
#> 358 358 1  0.0228261712  0.228682915  0.2682412171  0.1307119013
#> 359 359 1  0.0886814271  0.258754715 -0.0045823909  0.0694822353
#> 360 360 1  0.3301717806  0.398639906  0.0909666903  0.5893131850
#> 361 361 1  0.2729591237  0.280976210  0.0904872094  0.3853251337
#> 362 362 1 -0.2003797945  0.426090917  0.4470475403  0.1915274094
#> 363 363 1 -0.0663750509  0.530327479 -0.1837755255  0.0663308799
#> 364 364 1 -0.0127490259  0.490209252  0.1347900365  0.2944860454
#> 365 365 1 -0.0691071239 -0.021812687 -0.0175695006 -0.4511748364
#> 366 366 1 -0.0910590595  0.407121053  0.2792162803  0.1972156664
#> 367 367 1 -0.0047946642  0.331083698  0.0286974761  0.0563117774
#> 368 368 1  0.6130276757  0.684268131 -0.3894741855  0.9510815152
#> 369 369 1 -0.0058283273  0.241455320 -0.4939958082 -0.3850191614
#> 370 370 1  0.4104079395  0.797729104  0.4403435053  1.3652687721
#> 371 371 1  0.1521337221 -0.109194365 -0.0819004645 -0.3162361053
#> 372 372 1 -0.1600503660  0.765128992 -0.2182156494  0.1919899300
#> 373 373 1  0.2684607977  0.274755776  0.2517733917  0.4773566058
#> 374 374 1  0.4325512743  0.096875375 -0.3588619174  0.0848891522
#> 375 375 1  0.0408215839  0.053212880 -0.0984136754 -0.2824226234
#> 376 376 1 -0.1007045984  0.144742117  0.2545389925 -0.1263310725
#> 377 377 1 -0.0222157127  0.014196002  0.0181632578 -0.3291270315
#> 378 378 1  0.6381725224  0.208624092  0.3979708939  0.9574222947
#> 379 379 1  0.2086261130  0.512287093  0.4442449220  0.7953741680
#> 380 380 1  0.2840666526  0.363286845  0.1167994831  0.5089211413
#> 381 381 1  0.2955448543 -0.098049966 -0.2901727693 -0.2604732131
#> 382 382 1  0.0093527860  0.329658671  0.1604097231  0.1577440972
#> 383 383 1  0.0579568664  0.164872961  0.0211649789 -0.0577672840
#> 384 384 1 -0.1305364219  0.672702289  0.2376593375  0.4203199249
#> 385 385 1  0.2037662997  0.519219867 -0.1919170215  0.3844415294
#> 386 386 1  0.3312007159  0.582523858  0.2786929070  0.9195039280
#> 387 387 1  0.5168367058  0.191351368  0.3817512266  0.7765705124
#> 388 388 1  0.2226732681  0.259165143 -0.0120961396  0.2316806360
#> 389 389 1  0.9547623897  0.338323172  0.5738915467  1.6113007960
#> 390 390 1  0.4218597940  0.051789969  0.0758421383  0.3028172697
#> 391 391 1  0.4078628467  0.582229882 -0.6168005697  0.4335617582
#> 392 392 1  0.0384554919  0.544968041  0.0676286227  0.3762662980
#> 393 393 1  0.3435308907  0.046702172  0.1424556926  0.2429036239
#> 394 394 1  0.2121468098  0.196342889 -0.0258537567  0.1389035166
#> 395 395 1  0.4025759046  0.290762591  0.2015560842  0.6295729001
#> 396 396 1  0.4211323852  0.342283489 -0.2784884772  0.3992595173
#> 397 397 1  0.7294990657  0.513816641 -0.1613062457  1.0519294644
#> 398 398 1  0.4945939019  0.352764642 -0.6578154369  0.2563286455
#> 399 399 1 -0.2342979893  0.642189478  0.2766945928  0.2822526245
#> 400 400 1  0.3087780306 -0.160718805  0.1104982371 -0.0546806309
#> 401 401 1 -0.1857126120  0.363607122  0.4522950782  0.1427875142
#> 402 402 1  0.0917411538  0.383134539  0.2859816988  0.4018868311
#> 403 403 1 -0.1862864276  0.108600870 -0.0347943001 -0.4611561941
#> 404 404 1  0.2414126267  0.255805464  0.0884926075  0.3164526896
#> 405 405 1  0.3599608868 -0.002731350 -0.1019940861  0.0490687950
#> 406 406 1  0.2565641437  0.265010596  0.7045162243  0.7452958055
#> 407 407 1  0.2200968682  0.749705075 -0.0229594071  0.7739748472
#> 408 408 1 -0.1216827962  0.408513910  0.7242775314  0.4494315207
#> 409 409 1  0.1895104947  0.571385229  0.0765879119  0.5996623372
#> 410 410 1  0.4077535913 -0.100996921 -0.6135160288 -0.3340313389
#> 411 411 1  0.0048781722  0.122417346  0.4311121479  0.0943565814
#> 412 412 1  0.1336606535  0.359813916 -0.2237343836  0.0970733919
#> 413 413 1  0.2283374197  0.379612464  0.1164349359  0.4577780287
#> 414 414 1  0.2531506432  0.356801515 -0.0685448615  0.3429350429
#> 415 415 1  0.1647853134 -0.097673359  0.2636632642 -0.0633489744
#> 416 416 1  0.6045485002  0.239382611 -0.2222329024  0.5479130874
#> 417 417 1 -0.0827586759  0.090505673 -0.2018924758 -0.4612103497
#> 418 418 1  0.1351281595  0.440227248  0.6300943924  0.7433822584
#> 419 419 1  0.4393508374  0.277059891 -0.1295697065  0.4450534655
#> 420 420 1  0.3260130716  0.277459799  0.1849664471  0.5086252620
#> 421 421 1  0.4844372333  0.418717251  0.0619506111  0.7849241545
#> 422 422 1  0.3464503877  0.329689141  0.1100033066  0.5442380089
#> 423 423 1  0.3610941631  0.645617381 -0.1643599514  0.7403205011
#> 424 424 1  0.2936075135 -0.189226010  0.0043799457 -0.1744971686
#> 425 425 1  0.6947476417  0.145893129  0.1886647679  0.8213259865
#> 426 426 1  0.2857894655  0.102568611 -0.1380821313  0.0520398565
#> 427 427 1  0.6412874570  0.379171274  0.7204671750  1.3626139883
#> 428 428 1  0.3579045878  0.720665284  0.0626411425  0.9681515312
#> 429 429 1  0.3349290964  0.684433585  0.0840974361  0.9126908735
#> 430 430 1  0.1121451080  0.617803510  0.1728230328  0.6181797822
#> 431 431 1  0.1813232455 -0.145534363  0.1837752080 -0.1485217860
#> 432 432 1  0.0183509488 -0.006833109 -0.0146500936 -0.3236592315
#> 433 433 1  0.4523161106  0.218048075  0.0131613240  0.4872981901
#> 434 434 1  0.2434352880  0.087734607  0.2745399621  0.2503473373
#> 435 435 1  0.1697276033  0.354077429  0.0124586055  0.2886848466
#> 436 436 1  0.4200882691  0.084362579  1.0252727014  0.9532306935
#> 437 437 1  0.0881761868  0.230723955  0.2086194104  0.1755910089
#> 438 438 1  0.1016160860  0.528998716 -0.4670405486  0.0899574759
#> 439 439 1 -0.0211145563 -0.191912222 -0.2220734104 -0.7157678333
#> 440 440 1  0.6250113493  0.474764827  0.0026683867  0.9843931125
#> 441 441 1  0.1617386395  0.444094025 -0.0358721517  0.3487922143
#> 442 442 1 -0.0432990452  0.235996481 -0.2445640337 -0.2759460418
#> 443 443 1  0.2907193877  0.323907172 -0.1689163964  0.2874829177
#> 444 444 1 -0.2821542774  0.407971946  0.1255401873 -0.1391355417
#> 445 445 1  0.0823254764  0.178787527  0.6008252692  0.3642513611
#> 446 446 1  0.3776387411  0.128950520  0.1835844374  0.4046408628
#> 447 447 1  0.4170545432  0.167410527  0.7168069666  0.8428921073
#> 448 448 1  0.7132143876  0.495492750  0.4449500956  1.4043385359
#> 449 449 1 -0.1313649206 -0.010443431 -0.6397150866 -0.9193881641
#> 450 450 1  0.2457828805  0.485759936  0.0430486793  0.5514270554
#> 451 451 1  0.3601383806  0.269765962 -0.0021528378  0.4210012637
#> 452 452 1  0.2047253439  0.044507936 -0.0197197595 -0.0373724849
#> 453 453 1  0.5390959874  0.226508812 -0.0180439311  0.5844898912
#> 454 454 1  0.1429712409  0.124308861 -0.2531584673 -0.1757111925
#> 455 455 1 -0.5534937358  0.188077076 -0.1811544553 -0.9231820185
#> 456 456 1  0.1255928595  0.381699105 -0.5148273644 -0.0771483301
#> 457 457 1 -0.0547861613  0.517478192 -0.1911913023  0.0614566251
#> 458 458 1  0.5004713659  0.048712989  0.0171648819  0.3590341081
#> 459 459 1 -0.1450554101  0.915394544 -0.6664951771  0.0890811947
#> 460 460 1  0.2707568337  0.017904105  0.4315711338  0.3075339670
#> 461 461 1  0.4015823028  0.666125037  0.1587880565  1.0234014219
#> 462 462 1 -0.0298674497 -0.028230998 -0.0431081029 -0.4261800638
#> 463 463 1  0.5573184336  1.067155307  0.3706851085  1.8062353725
#> 464 464 1 -0.2429852387  0.247990170 -0.2461298786 -0.5117494816
#> 465 465 1  0.1080622083  0.126157331 -0.6744664513 -0.4903524119
#> 466 466 1  0.2751980205  0.072712121  0.4120132428  0.3621043278
#> 467 467 1  0.2352249965  0.186141893  0.5739283468  0.5452074556
#> 468 468 1  0.2974850375  0.491985471  0.2543561762  0.7598099377
#> 469 469 1  0.0441999202  0.597359032  0.2624334342  0.5687986440
#> 470 470 1 -0.1107830373  0.366780519  0.2076616746  0.0808306816
#> 471 471 1  0.1873622605  0.186292530  0.3745086102  0.3564927618
#> 472 472 1 -0.3170421471 -0.171343461 -0.0598642081 -0.9553362984
#> 473 473 1  0.2633295016  0.112686791  0.3669089434  0.3631134926
#> 474 474 1  0.0265817310 -0.119120487 -0.0871572803 -0.4869432543
#> 475 475 1  0.4263193980  0.469986502  0.3208238746  0.9383475036
#> 476 476 1  0.0016190140 -0.056331339  0.4014477110 -0.1302831352
#> 477 477 1  0.2512374828  0.237711457 -0.0970156784  0.1879429273
#> 478 478 1  0.5319711283  0.099999816  0.3288504260  0.6581721905
#> 479 479 1  0.1676586725  0.366479657  0.0195753859  0.3046990413
#> 480 480 1 -0.0718772696 -0.003607032 -0.0461871286 -0.4526776368
#> 481 481 1  0.0397293277  0.460093482 -0.2216431623  0.0945873870
#> 482 482 1  0.4121970371  0.328432208 -0.0371146502  0.5291338764
#> 483 483 1  0.1528279718  0.411953911  0.4542321064  0.6194561690
#> 484 484 1 -0.0834311819 -0.118446539 -0.4788562978 -0.8770861572
#> 485 485 1 -0.0166471386  0.511384521  0.4129578182  0.4939472991
#> 486 486 1  0.6508246903  0.296819568 -0.3287369851  0.6010591016
#> 487 487 1 -0.2998416179  0.401840137  0.0121916514 -0.2415683952
#> 488 488 1 -0.0784465126  0.384483816  0.2003970109  0.1362673416
#> 489 489 1  0.3924913966  0.711043892  0.3467529809  1.1846330676
#> 490 490 1 -0.0295462039  0.030063151 -0.1781350664 -0.4477143322
#> 491 491 1  0.7084306401  0.524017535  0.6083254641  1.5365075113
#> 492 492 1  0.2281885203  0.623101669  0.1975115803  0.7844566535
#> 493 493 1  0.0812353836  0.342650049 -0.2306837744  0.0080441267
#> 494 494 1  0.2585399560  0.290922794  0.0508029945  0.3528552143
#> 495 495 1  0.6275251215  0.244102455  0.0735861701  0.7737069386
#> 496 496 1 -0.0785142581  0.205413005  0.3213127625  0.0129193496
#> 497 497 1 -0.0338927607  0.605359973 -0.3537614043  0.0809619903
#> 498 498 1 -0.1791019830  0.604331684  0.0359402851  0.1520572086
#> 499 499 1  0.0615829635  0.422404152  0.2305310190  0.3726477190
#> 500 500 1 -0.0462786527  0.085321415  0.0435432071 -0.2624671964