Declare measurement procedure

This function adds measured data columns that can be functions of unmeasured data columns.

Usage

declare_measurement(..., handler = measurement_handler, label = NULL)

measurement_handler(data, ...)

Arguments

...

arguments to be captured, and later passed to the handler

handler

a tidy-in, tidy-out function

label

a string describing the step

data

A data.frame.

Value

A function that returns a data.frame.

Details

It is also possible to include measured variables in your declare_model call or to add variables using declare_step. However, putting latent variables in declare_model and variables-as-measured in declare_measurement helps communicate which parts of your research design are in M and which parts are in D.

Examples

# declare_measurement in use
## Two-arm randomized experiment
design <-
  declare_model(
    N = 500,
    X = rep(c(0, 1), each = N / 2),
    U = rnorm(N, sd = 0.25),
    potential_outcomes(Y ~ 0.2 * Z + X + U)
  ) +
  declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(S = complete_rs(N = N, n = 200)) +
  declare_assignment(Z = complete_ra(N = N, m = 100)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z)) +
  declare_estimator(Y ~ Z, inquiry = "ATE")
  
run_design(design)
#>   inquiry estimand estimator term  estimate  std.error statistic    p.value
#> 1     ATE      0.2 estimator    Z 0.1550252 0.08031084  1.930315 0.05499649
#>       conf.low conf.high  df outcome
#> 1 -0.003349185 0.3133996 198       Y

# Reveal potential outcomes according to treatment assignment
design <-
  declare_model(N = 100,
                potential_outcomes(Y ~ rbinom(
                  N, size = 1, prob = 0.1 * Z + 0.5
                ))) +
  declare_assignment(Z = complete_ra(N, m = 50)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z))
  
head(draw_data(design))
#>    ID Y_Z_0 Y_Z_1 Z Y
#> 1 001     0     1 1 1
#> 2 002     0     1 0 0
#> 3 003     1     0 1 0
#> 4 004     0     1 0 0
#> 5 005     1     0 1 0
#> 6 006     0     1 0 0

# Generate observed measurement from a latent value
design <- 
  declare_model(N = 100, latent = runif(N)) +
  declare_measurement(observed = rbinom(N, 1, prob = latent))
  
head(draw_data(design))
#>    ID    latent observed
#> 1 001 0.6470987        1
#> 2 002 0.4379810        0
#> 3 003 0.8303115        1
#> 4 004 0.9474058        1
#> 5 005 0.2534557        1
#> 6 006 0.1299259        0

# Index creation
library(psych)

design <-
  declare_model(
    N = 500,
    X = rep(c(0, 1), each = N / 2),
    Y_1 = 0.2 * X + rnorm(N, sd = 0.25),
    Y_2 = 0.3 * X + 0.5 * rnorm(N, sd = 0.50),
    Y_3 = 0.1 * X + 0.4 * rnorm(N, sd = 0.75)) +
  declare_measurement(
    index = fa(
      r = cbind(Y_1, Y_2, Y_3),
      nfactors = 1,
      rotate = "varimax"
    )$scores
  )

draw_data(design)
#>      ID X          Y_1          Y_2           Y_3           MR1
#> 1   001 0  0.234786179  0.255737035 -0.8315109123  0.2564617289
#> 2   002 0 -0.063783207  0.233236373  0.2451738586  0.2128395022
#> 3   003 0 -0.030403562 -0.171517786 -0.1613010331 -0.9149528202
#> 4   004 0 -0.063443733  0.440488941 -0.3768919101  0.7118722538
#> 5   005 0  0.317284997 -0.035044964  0.0361154701 -0.4188462672
#> 6   006 0  0.573104833  0.179916354 -0.0147183601  0.2368407837
#> 7   007 0  0.031938087 -0.261252743  0.2707672929 -1.0954925458
#> 8   008 0  0.349121342 -0.209030119  0.0843677090 -0.8757836379
#> 9   009 0 -0.039382002 -0.038312328  0.0782640058 -0.5325043744
#> 10  010 0  0.086751447  0.451452916 -0.2544417595  0.7998684452
#> 11  011 0  0.046005841 -0.866176488 -0.1154666581 -2.7696542941
#> 12  012 0 -0.136899441 -0.184823617 -0.2345016683 -0.9909643648
#> 13  013 0 -0.043393179  0.242023679 -0.1875246389  0.1992246892
#> 14  014 0 -0.035697960  0.029954495 -0.1725706064 -0.3716711788
#> 15  015 0  0.166550090 -0.020642783  0.0943805085 -0.4200226077
#> 16  016 0  0.206800261 -0.167635352  0.3298125221 -0.7823420762
#> 17  017 0  0.199102227 -0.016270780  0.4924094405 -0.3580537263
#> 18  018 0 -0.042389446 -0.345709235 -0.2566849327 -1.4003433796
#> 19  019 0 -0.206357599  0.167674859 -0.0460996167 -0.0378353409
#> 20  020 0 -0.250375278 -0.186858168  0.2919218433 -0.9780522202
#> 21  021 0  0.086616534 -0.435940217 -0.3336991111 -1.6132170748
#> 22  022 0  0.012377097  0.194854147 -0.2008372273  0.0870932128
#> 23  023 0  0.184314494  0.176894543  0.4255896345  0.1542056000
#> 24  024 0 -0.324754481  0.100771880 -0.5624530015 -0.3074668113
#> 25  025 0  0.026422221 -0.274198445  0.2779775606 -1.1315374154
#> 26  026 0 -0.052006411 -0.003580823  0.2762200849 -0.4222574374
#> 27  027 0  0.326410530  0.142941200 -0.2096255438  0.0415302295
#> 28  028 0  0.433344453 -0.261780779 -0.2674188204 -1.0284979622
#> 29  029 0  0.780871822  0.043255646  0.4423812170 -0.0238869716
#> 30  030 0 -0.472338091 -0.026343649  0.4937806743 -0.5905178652
#> 31  031 0 -0.007138264 -0.165314259 -0.3004417194 -0.9050672331
#> 32  032 0 -0.172035631 -0.024958867 -0.1231566011 -0.5572001264
#> 33  033 0  0.134014518  0.025602550 -0.2744736222 -0.3418572942
#> 34  034 0 -0.069522194 -0.038413249 -0.2093569066 -0.5710184812
#> 35  035 0 -0.009516512 -0.411093273  0.3272938328 -1.5085670326
#> 36  036 0  0.092863144  0.071289941  0.0638297397 -0.1964769271
#> 37  037 0 -0.017518246 -0.323165838 -0.7002919335 -1.3764065496
#> 38  038 0  0.148629105 -0.214740304 -0.1388822993 -0.9750919167
#> 39  039 0  0.371911710 -0.152697214 -0.0948890216 -0.7342289934
#> 40  040 0  0.005688023  0.221117061  0.8537349295  0.2626470526
#> 41  041 0  0.190749204  0.212024069 -0.0967752305  0.1986688889
#> 42  042 0 -0.281978380  0.380677231 -0.3856931591  0.4820600042
#> 43  043 0  0.353554082  0.121606974 -0.1245198472  0.0005972448
#> 44  044 0 -0.068156271  0.292585782 -0.4399565164  0.3032145801
#> 45  045 0 -0.311726581  0.514016198 -0.4678477910  0.8260538296
#> 46  046 0  0.283931064 -0.413398109  0.1524523908 -1.4427316516
#> 47  047 0  0.390128867  0.245697056  0.4986945182  0.4109847473
#> 48  048 0  0.057986431 -0.186085715  0.1373525871 -0.8972707083
#> 49  049 0 -0.005151154 -0.061259941 -0.4168309169 -0.6341929535
#> 50  050 0  0.044662276 -0.141995928  0.1746949264 -0.7780826287
#> 51  051 0  0.368633842  0.214292819 -0.3792465089  0.2307034392
#> 52  052 0 -0.065155002  0.052807499 -0.3226736975 -0.3338884311
#> 53  053 0  0.027727683 -0.497771192 -0.4737297668 -1.8129307833
#> 54  054 0  0.358639153  0.086492003  0.4467105406 -0.0353730715
#> 55  055 0 -0.086446419  0.134216501  0.2392972924 -0.0630504994
#> 56  056 0  0.118604252 -0.048118966 -0.0625494073 -0.5249860487
#> 57  057 0 -0.097953508  0.196505567  0.2350106597  0.1018166258
#> 58  058 0 -0.459528458 -0.094191936  0.6710294837 -0.7525993365
#> 59  059 0  0.032500960 -0.181907384 -0.1864799636 -0.9264180137
#> 60  060 0 -0.022020307  0.285935650  0.2372397971  0.3676360199
#> 61  061 0 -0.553006559  0.187514278  0.1051095588 -0.0747997463
#> 62  062 0  0.242744734 -0.331049965  0.1154461614 -1.2358763663
#> 63  063 0  0.227257321 -0.184558022 -0.0580767821 -0.8610946898
#> 64  064 0  0.276972834 -0.044974267 -0.0082829179 -0.4625634554
#> 65  065 0  0.183635956  0.764069204 -0.2978990907  1.6723712568
#> 66  066 0  0.007937975 -0.155904656  0.2427162732 -0.8201433132
#> 67  067 0 -0.003120742  0.190150825  0.5075906781  0.1410972221
#> 68  068 0  0.008425458  0.114123171  0.2101374234 -0.0914423573
#> 69  069 0  0.130524197  0.058791544  0.2751304188 -0.1975120819
#> 70  070 0  0.108920371  0.009790520  0.0589974559 -0.3587328755
#> 71  071 0  0.116595099  0.350444260 -0.5371820537  0.5067038656
#> 72  072 0 -0.009834597 -0.292177406  0.6754322344 -1.1512422449
#> 73  073 0 -0.466219253  0.317898044  0.1383998322  0.3084671398
#> 74  074 0 -0.294923901 -0.279374564 -0.4271348580 -1.3149774208
#> 75  075 0 -0.117584653  0.001375246  0.0476818429 -0.4519393705
#> 76  076 0 -0.082397433 -0.021558031  0.4130691610 -0.4664625417
#> 77  077 0  0.127000139 -0.363665093 -0.0122123385 -1.3725441483
#> 78  078 0 -0.028824301  0.138093368 -0.2741197156 -0.0867350875
#> 79  079 0  0.341805422  0.393241347 -0.4134110221  0.7040456139
#> 80  080 0 -0.048531288 -0.339327864 -0.5612954226 -1.4156655658
#> 81  081 0 -0.122362631 -0.244633833  0.2852734655 -1.0961671746
#> 82  082 0 -0.375012375  0.036574330  0.0831815628 -0.4316713687
#> 83  083 0 -0.091607933  0.346401639  0.4694589397  0.5336697882
#> 84  084 0 -0.031114486  0.201947905  0.1805931865  0.1315122092
#> 85  085 0 -0.180829675  0.142930475  0.0578636150 -0.0866016889
#> 86  086 0  0.299990157 -0.200861957  0.3518477691 -0.8416762510
#> 87  087 0 -0.429694178 -0.191846431  0.3443378758 -1.0411114847
#> 88  088 0  0.062958134 -0.113510566 -0.0810661362 -0.7210960085
#> 89  089 0 -0.216050597 -0.045264334  0.5619995306 -0.5565495920
#> 90  090 0 -0.100219759  0.119104016  0.1383201900 -0.1184102571
#> 91  091 0 -0.236236239  0.426482854  0.4582729712  0.7053578784
#> 92  092 0 -0.338054306  0.480632462 -0.2730944966  0.7471790966
#> 93  093 0  0.093802755 -0.165166307 -0.0123191126 -0.8447266529
#> 94  094 0  0.094125956 -0.239624772 -0.0532320430 -1.0505566123
#> 95  095 0 -0.085568348 -0.147391822 -0.1927448845 -0.8696062667
#> 96  096 0  0.133926063 -0.208876209 -0.1341237213 -0.9632142913
#> 97  097 0 -0.098959729 -0.434035827  0.1724294192 -1.6137232529
#> 98  098 0 -0.381516086 -0.031364456  0.1415925211 -0.6118957739
#> 99  099 0  0.297404971  0.161172204 -0.0798663229  0.0951661428
#> 100 100 0  0.165985439  0.455093541 -0.3472757288  0.8245941635
#> 101 101 0  0.153811051  0.081451359  0.3903862115 -0.1173474609
#> 102 102 0  0.032399393 -0.274517639 -0.0496012270 -1.1636319675
#> 103 103 0  0.441922147  0.199012452  0.0212166380  0.2521110501
#> 104 104 0 -0.059186114  0.029179205 -0.2100662148 -0.3847380502
#> 105 105 0 -0.000752335  0.066155810 -0.5980166862 -0.3058050621
#> 106 106 0  0.391568341  0.036738549 -0.6089262923 -0.2666759963
#> 107 107 0 -0.004698007 -0.491981124 -0.0820075343 -1.7676229567
#> 108 108 0 -0.078227253  0.278794803  0.2237284936  0.3297330689
#> 109 109 0  0.021955199  0.180457222 -0.1535014383  0.0557795826
#> 110 110 0 -0.075847432 -0.245054844 -0.2673127379 -1.1388488289
#> 111 111 0  0.132959802 -0.275501855 -0.0021096735 -1.1307590866
#> 112 112 0 -0.355902978 -0.098194908  0.3605569334 -0.7630941631
#> 113 113 0 -0.040524195 -0.125063932 -0.0833520763 -0.7842802843
#> 114 114 0  0.115664483  0.502406236  0.5984878005  1.0328765869
#> 115 115 0  0.112411765  0.393409588  0.1574875097  0.6919724100
#> 116 116 0 -0.261669676 -0.049354865 -0.0963390012 -0.6480197328
#> 117 117 0  0.079758549 -0.092786617 -0.2126575557 -0.6730715416
#> 118 118 0  0.006095567 -0.191755241  0.2882993200 -0.9132703196
#> 119 119 0 -0.060353574 -0.030109061 -0.3851271966 -0.5634462844
#> 120 120 0  0.067608338 -0.018708543  0.1684106423 -0.4375626123
#> 121 121 0  0.130815236 -0.023493714 -0.1260370038 -0.4609188238
#> 122 122 0 -0.723818710  0.101251022 -0.3277241953 -0.4045006530
#> 123 123 0 -0.093263617  0.373340887  0.7247348522  0.6319360424
#> 124 124 0 -0.145233342 -0.009190903  0.2910000683 -0.4644760790
#> 125 125 0  0.253986326 -0.206314189  0.3220386786 -0.8735275685
#> 126 126 0  0.265092786 -0.633849781  0.0207449504 -2.0592591533
#> 127 127 0 -0.229418859 -0.151747884  0.1912851243 -0.8866428701
#> 128 128 0 -0.372761600  0.393488388  0.1942059736  0.5475425445
#> 129 129 0 -0.066726698  0.060888427  0.4813136252 -0.2313347283
#> 130 130 0  0.105675793  0.075560711 -0.1747641342 -0.2050617758
#> 131 131 0  0.109174622  0.035661388  0.3606115917 -0.2581020510
#> 132 132 0 -0.026860071 -0.135790272 -0.3105496867 -0.8321004746
#> 133 133 0 -0.111954089  0.335555934 -0.0291046900  0.4477425047
#> 134 134 0  0.138947662 -0.137415484 -0.3124936574 -0.7860032327
#> 135 135 0  0.185823104  0.026513571 -0.0869147534 -0.3046197943
#> 136 136 0 -0.279494675  0.032167802 -0.1870540510 -0.4416785562
#> 137 137 0  0.447729611 -0.282772627 -0.0706074233 -1.0611314915
#> 138 138 0  0.255074984 -0.126083036  0.5703085680 -0.6306954763
#> 139 139 0  0.032347415 -0.046562431 -0.0508695235 -0.5459630935
#> 140 140 0  0.315110098 -0.008654037 -0.1079477920 -0.3625234980
#> 141 141 0 -0.345840932  0.158836750  0.1064966663 -0.0890387771
#> 142 142 0  0.217497072  0.281559986 -0.3904069212  0.3656749357
#> 143 143 0  0.153725783 -0.191776158 -0.4209958549 -0.9397642600
#> 144 144 0  0.375743135 -0.122346132 -0.3827652126 -0.6798493841
#> 145 145 0 -0.111881752  0.337047994  0.2071048169  0.4756451472
#> 146 146 0  0.367070359 -0.044216242 -0.1126046116 -0.4434878485
#> 147 147 0  0.372934207  0.369280516 -0.2834308429  0.6617408830
#> 148 148 0  0.288323225  0.002376867  0.4593090430 -0.2835738280
#> 149 149 0 -0.279419828  0.137671110 -0.0552366572 -0.1424145161
#> 150 150 0  0.066597733  0.437206542  0.1276720213  0.7936554717
#> 151 151 0  0.102453132 -0.342283788 -0.1765035006 -1.3386805045
#> 152 152 0 -0.141680014 -0.360665042 -0.1249601051 -1.4579439159
#> 153 153 0  0.410163135  0.084982554 -0.3270322809 -0.1017907940
#> 154 154 0  0.071320090  0.143253513  0.2043815005  0.0061578869
#> 155 155 0 -0.248113893 -0.207653525 -0.0790858107 -1.0711607434
#> 156 156 0 -0.290243038 -0.215208867 -0.4538232589 -1.1423351165
#> 157 157 0 -0.011329243 -0.120477107  0.1968472980 -0.7346462660
#> 158 158 0  0.162995731 -0.057866270  0.3914423730 -0.4920163942
#> 159 159 0  0.002473891  0.440064078 -0.2207530269  0.7466325438
#> 160 160 0  0.245292857 -0.192961152 -0.1651026218 -0.8891548719
#> 161 161 0  0.330957846 -0.050605335 -0.0828253969 -0.4688406144
#> 162 162 0 -0.038113158  0.227620287  0.5107137561  0.2322640805
#> 163 163 0  0.015734918 -0.021188779  0.1226752617 -0.4647610030
#> 164 164 0  0.015938799 -0.203820385  0.3014688888 -0.9416310297
#> 165 165 0  0.012781890  0.307273034 -0.5233537928  0.3593526461
#> 166 166 0  0.462654454 -0.273292016  0.1329125367 -1.0103354213
#> 167 167 0 -0.096338404  0.009512520  0.0355542945 -0.4246129054
#> 168 168 0  0.114900327 -0.294989451 -0.3040704255 -1.2195689106
#> 169 169 0  0.169035824  0.109595367 -0.3497134074 -0.1111012371
#> 170 170 0 -0.201468030  0.001862509 -0.0072131778 -0.4818069745
#> 171 171 0  0.214842154  0.456594524 -0.6620311532  0.8118378200
#> 172 172 0  0.013379877 -0.144919552 -0.3349682413 -0.8470032012
#> 173 173 0 -0.231153645  0.276328404 -0.0671480656  0.2469357479
#> 174 174 0 -0.109110670 -0.061228456 -0.0204662516 -0.6258962793
#> 175 175 0  0.193615488 -0.072971092  0.6955963408 -0.4928984230
#> 176 176 0 -0.365157566  0.321175108  0.3823199536  0.3728646333
#> 177 177 0  0.064895201 -0.100035743  0.3138265898 -0.6441338129
#> 178 178 0  0.181681968  0.124998103  0.0662663170 -0.0235115228
#> 179 179 0 -0.395839861 -0.218330014 -0.0534331412 -1.1426772311
#> 180 180 0  0.029144214  0.331650177 -0.0590482178  0.4772781318
#> 181 181 0  0.152845547 -0.186748836 -0.2934561363 -0.9135377115
#> 182 182 0 -0.033643092  0.136902398  0.2772669132 -0.0357940522
#> 183 183 0 -0.506551393  0.014966346  0.2783562142 -0.5107583814
#> 184 184 0 -0.303502258 -0.013163173  0.2407676899 -0.5287040149
#> 185 185 0 -0.399653793  0.160190214  0.3564237053 -0.0766036631
#> 186 186 0 -0.050735961 -0.044529667 -0.3049914167 -0.5915020591
#> 187 187 0 -0.146276552 -0.107165639 -0.4525747842 -0.8053664747
#> 188 188 0 -0.057549316 -0.252072594 -0.0609952331 -1.1314534963
#> 189 189 0  0.107567812  0.156230215 -0.5533147609 -0.0240505528
#> 190 190 0  0.223228648 -0.436154194 -0.1826391961 -1.5567819016
#> 191 191 0  0.060336687  0.351821399  0.4385748318  0.5917012861
#> 192 192 0  0.141223285 -0.538117558 -0.3572040031 -1.8758170829
#> 193 193 0 -0.149332370 -0.008032229 -0.3065544546 -0.5228903510
#> 194 194 0 -0.002203537 -0.347223982 -0.2200600086 -1.3884653353
#> 195 195 0  0.357998299  0.011015072 -0.0131740340 -0.2865379147
#> 196 196 0 -0.083983383  0.207893101 -0.1647862014  0.0966062837
#> 197 197 0  0.211791398  0.194314401 -0.3501519413  0.1315363507
#> 198 198 0 -0.294154615  0.259858188 -0.0336740829  0.1864120584
#> 199 199 0  0.103141838 -0.293731578 -0.0869641333 -1.1978461875
#> 200 200 0 -0.051607748  0.357024759 -0.1572021125  0.5114531518
#> 201 201 0  0.032653066  0.052417398  0.3409841054 -0.2380675382
#> 202 202 0  0.135866126  0.016955231 -0.4907867537 -0.3865563056
#> 203 203 0 -0.269410633 -0.182342846  0.2761353127 -0.9732280333
#> 204 204 0 -0.138157358  0.258733108 -0.6090378532  0.1729993825
#> 205 205 0  0.112935735 -0.130065664 -0.0330418884 -0.7458365908
#> 206 206 0 -0.029412603  0.029872393  0.1806780715 -0.3343243541
#> 207 207 0 -0.172869958 -0.011580911 -0.3855514774 -0.5476769358
#> 208 208 0  0.512622443  0.342074042  0.4111065977  0.7008046738
#> 209 209 0  0.301203215  0.004821286  0.2997285699 -0.2891144237
#> 210 210 0 -0.076730378  0.080338594  0.2248204349 -0.2075625090
#> 211 211 0  0.009504739  0.214080862 -0.2223109161  0.1361569284
#> 212 212 0  0.062029015 -0.287953760  0.1893129397 -1.1668777331
#> 213 213 0 -0.090425916 -0.206647987  0.2146629259 -0.9905769397
#> 214 214 0 -0.082492138  0.382132066 -0.2012253232  0.5656140373
#> 215 215 0 -0.075498074  0.015979127  0.0219445414 -0.4020880380
#> 216 216 0  0.387334750 -0.053914844  0.3254671081 -0.4193699555
#> 217 217 0  0.413933890  0.025184955  0.3742900152 -0.1919307567
#> 218 218 0  0.023790270 -0.015852093 -0.1322271441 -0.4735573262
#> 219 219 0 -0.112524194 -0.279366450 -0.0101086688 -1.2171005657
#> 220 220 0  0.020910196  0.334482586  0.8716013591  0.5763519095
#> 221 221 0  0.230172207 -0.032733146  0.2356347773 -0.4190825375
#> 222 222 0  0.136617590 -0.085002215 -0.4379656267 -0.6573250709
#> 223 223 0  0.070855134  0.054705958 -0.1563185462 -0.2703686258
#> 224 224 0 -0.069969334 -0.014386136 -0.3271836913 -0.5179263114
#> 225 225 0 -0.036752925  0.018594041  0.3627841237 -0.3487588475
#> 226 226 0  0.211252792  0.381899850  0.4725678841  0.7227961786
#> 227 227 0 -0.127160972 -0.128762418 -0.2410497515 -0.8367083123
#> 228 228 0 -0.392884246  0.103089838  0.2078674684 -0.2442806585
#> 229 229 0 -0.081402813 -0.306767594 -0.1089504819 -1.2918228150
#> 230 230 0  0.277771832 -0.053511315 -0.0166895482 -0.4863049201
#> 231 231 0 -0.374399283 -0.282633346 -0.3199219121 -1.3372910333
#> 232 232 0  0.106217071  0.387428664 -0.3451582095  0.6231449143
#> 233 233 0 -0.245373186 -0.374144077 -0.0846230365 -1.5221104505
#> 234 234 0  0.165989479 -0.376606042 -0.8149473989 -1.4767023021
#> 235 235 0  0.209269217  0.075328222 -0.5049877260 -0.2073406416
#> 236 236 0 -0.052133207 -0.308726744 -0.0592832187 -1.2831708790
#> 237 237 0  0.362115236 -0.335840408  0.6112019540 -1.1623319683
#> 238 238 0  0.040914890  0.106119356 -0.3158938749 -0.1562840510
#> 239 239 0  0.270418585  0.067469841 -0.0948576954 -0.1685538239
#> 240 240 0  0.099066558  0.233332926 -0.0591866942  0.2321808314
#> 241 241 0  0.111549791 -0.030870228  0.4821063192 -0.4254318524
#> 242 242 0 -0.146012154  0.063798032  0.5103857448 -0.2447579253
#> 243 243 0  0.353302911 -0.032432557  0.1119575015 -0.3930995498
#> 244 244 0  0.240089512  0.221989795  0.1727825782  0.2679668637
#> 245 245 0 -0.086952095  0.184740156 -0.2952243963  0.0197856255
#> 246 246 0  0.029310427 -0.176435021 -0.2833082238 -0.9223334253
#> 247 247 0 -0.109026177 -0.078486275 -0.6613344532 -0.7373153740
#> 248 248 0 -0.082423914  0.016320574 -0.0568804244 -0.4112348008
#> 249 249 0  0.223531773  0.231368047  0.8339021099  0.3550371580
#> 250 250 0  0.083520098  0.141232276  0.6041035809  0.0447474703
#> 251 251 1  0.178175206 -0.024952797 -0.0868382894 -0.4464366143
#> 252 252 1  0.218018053  0.644160543 -0.2607470863  1.3616522660
#> 253 253 1  0.120283266  0.261245047 -0.3897194101  0.2809614685
#> 254 254 1  0.302076366  0.409541800 -0.4867462065  0.7286754782
#> 255 255 1 -0.160689062  0.284369059  0.4049287468  0.3379122468
#> 256 256 1 -0.080648841  0.317139482  0.1811821392  0.4286223244
#> 257 257 1  0.307525900  0.108902476  0.7578231860  0.0411314248
#> 258 258 1  0.238929406 -0.089721720  0.8548663960 -0.5083687418
#> 259 259 1  0.246447245  0.106485845 -0.0798458191 -0.0686258748
#> 260 260 1  0.039488992  0.402576569  0.1265545182  0.6913983566
#> 261 261 1  0.209851542  0.379755666 -0.4835164733  0.6200748235
#> 262 262 1  0.019149206  0.102132761  0.3930523350 -0.1022017219
#> 263 263 1  0.211732116  0.470550946 -0.3950581122  0.8756532369
#> 264 264 1  0.250230159  0.076907414 -0.0094460020 -0.1405294857
#> 265 265 1  0.112216845 -0.194990663  0.2687349807 -0.8915652612
#> 266 266 1  0.247688222 -0.183850443 -0.2537124350 -0.8726721812
#> 267 267 1  0.279892827  0.392769989 -0.1457702990  0.7108459605
#> 268 268 1  0.430354581  0.587383313 -0.0507939066  1.2938845455
#> 269 269 1  0.257626345  0.370259484  0.3601690971  0.6940848764
#> 270 270 1  0.191583592  0.377682310 -0.1137977288  0.6461793927
#> 271 271 1 -0.143636811  0.566453119  0.1031673671  1.0771892722
#> 272 272 1 -0.272604917  0.217156660  0.0710365070  0.0878366049
#> 273 273 1  0.339320103  0.076933294  0.2770235904 -0.0843101995
#> 274 274 1  0.730490986  0.180062322 -0.0428542775  0.2825202084
#> 275 275 1  0.220837405  0.192737771  0.0962914387  0.1750813449
#> 276 276 1  0.097019126  0.312025802 -0.4675771558  0.4036191430
#> 277 277 1  0.047586763  0.301607799  0.4728799974  0.4551738775
#> 278 278 1  0.416446438  0.291292314  0.0463813790  0.4969612805
#> 279 279 1  0.405952087  0.386166505 -0.1699301673  0.7290553532
#> 280 280 1  0.097885936 -0.210921437  0.1829806216 -0.9477770832
#> 281 281 1  0.471505985  0.478641142  0.0893589150  1.0258933637
#> 282 282 1  0.324544581  0.318338301  0.3248301822  0.5702611956
#> 283 283 1  0.324397128  0.028712730 -0.3219242986 -0.2800041207
#> 284 284 1  0.149864512 -0.083115413  0.0637319465 -0.5975329782
#> 285 285 1 -0.460982350  0.185133620 -0.2682318222 -0.0907892753
#> 286 286 1  0.147133668 -0.012917796  0.2151091231 -0.3928397852
#> 287 287 1  0.222779462  0.468538732  0.5035897108  0.9642630884
#> 288 288 1  0.232161200  0.484873824  0.4968957370  1.0107281601
#> 289 289 1  0.384399388 -0.102258301  0.4734457159 -0.5363567220
#> 290 290 1  0.477531410  0.597914728 -0.4078229316  1.3008231933
#> 291 291 1  0.109624321  0.388840660  0.4260312055  0.7058368682
#> 292 292 1 -0.061523327  0.228901719 -0.1160961663  0.1653255897
#> 293 293 1  0.519894886  0.183457609  0.9535617873  0.3278810844
#> 294 294 1  0.454786799  0.232716848  0.0900193224  0.3543346306
#> 295 295 1  0.217054874  0.336417129 -0.2601610881  0.5273591545
#> 296 296 1  0.695217488  0.751752653 -0.0411406077  1.8213243278
#> 297 297 1  0.482821718  0.094907557 -0.4716787406 -0.0672720149
#> 298 298 1  0.298959733  0.106766015  0.6046287092  0.0172625120
#> 299 299 1  0.131986978  0.527898684  0.1543042556  1.0621339762
#> 300 300 1 -0.078080156  0.120344750  0.1386469376 -0.1082450944
#> 301 301 1  0.420840081  0.376728514  0.4638223055  0.7719824465
#> 302 302 1  0.336321521  0.491199774 -0.3643686131  0.9728084673
#> 303 303 1  0.298539642  0.493645724 -0.3201957047  0.9723428477
#> 304 304 1  0.301674958  0.478255681  0.5713785550  1.0215625706
#> 305 305 1 -0.004789008  0.226749067 -0.2327387953  0.1650679396
#> 306 306 1  0.098694058 -0.282616785  0.1297626950 -1.1472118064
#> 307 307 1  0.489425440  0.348157628 -0.0832847502  0.6603091388
#> 308 308 1 -0.070317720  0.074476603  0.1053648627 -0.2335437921
#> 309 309 1  0.722723040 -0.256070271  0.3306578586 -0.8641855073
#> 310 310 1 -0.036191054  0.101287095 -0.1915035176 -0.1804042441
#> 311 311 1  0.360101791  0.692069083 -0.3289454994  1.5280578279
#> 312 312 1 -0.004171675  0.123224706  0.5620413605 -0.0351149959
#> 313 313 1  0.271356567  0.069260197 -0.2064458567 -0.1746754647
#> 314 314 1  0.365641525  0.287564875  0.0278164579  0.4694512467
#> 315 315 1  0.266714968  0.604923751  0.1481172502  1.3114608908
#> 316 316 1 -0.320775902  0.383041656  0.2031054404  0.5360228776
#> 317 317 1  0.306136589  0.390668022  0.0151567207  0.7294132264
#> 318 318 1  0.276720828  0.774948166  0.4256305629  1.8033316655
#> 319 319 1  0.256355173 -0.214342418 -0.1726039412 -0.9444777021
#> 320 320 1  0.234726425  0.238005469  0.8882674945  0.3819352677
#> 321 321 1 -0.187472948  0.249866154  0.1910393379  0.2146273249
#> 322 322 1  0.100653909  0.182415413 -0.2782183872  0.0725644185
#> 323 323 1 -0.018253426  0.273155468 -0.0473965913  0.3054268667
#> 324 324 1  0.143637351  0.236291259  0.5110477760  0.3113710439
#> 325 325 1  0.354342059  0.415289580 -0.2393573856  0.7851991520
#> 326 326 1  0.099198213 -0.056478052  0.2982154313 -0.5171688529
#> 327 327 1  0.198837652  0.093565267 -0.2249904594 -0.1328480632
#> 328 328 1  0.408291178  0.438581583 -0.1208545941  0.8767802532
#> 329 329 1  0.449813533  0.425807686  0.2431532437  0.8915892636
#> 330 330 1 -0.420418890  0.395771431  0.5054726472  0.5705691796
#> 331 331 1  0.084795790 -0.148210022 -0.4001815968 -0.8406655976
#> 332 332 1 -0.079588367 -0.195002092 -0.0124610936 -0.9786199006
#> 333 333 1  0.488034277  0.409811817  0.2036344873  0.8559352411
#> 334 334 1  0.108162373  0.356463664  0.3189636859  0.6068359673
#> 335 335 1  0.244145543  0.677535541 -0.2308794282  1.4631093406
#> 336 336 1  0.278847446  0.204351595 -0.3120274095  0.1830901664
#> 337 337 1  0.472982648  0.508499555  0.3003180771  1.1285568573
#> 338 338 1 -0.019281396  0.159770396  0.3376753784  0.0366709692
#> 339 339 1  0.108382716  0.552865929  0.0086908784  1.1078894601
#> 340 340 1  0.423583882  0.340528265  0.0337041823  0.6313054804
#> 341 341 1  0.012613505 -0.422765261 -0.2653942668 -1.5932444474
#> 342 342 1  0.451746099  0.531613500  0.1856122330  1.1731324048
#> 343 343 1  0.169090480  0.639142580  0.4308243623  1.4028808524
#> 344 344 1  0.023129177  0.487263210  0.1491552044  0.9181977244
#> 345 345 1  0.025691644  0.413576657 -0.0289310293  0.7013019028
#> 346 346 1  0.380135916  0.892761786  0.6130835634  2.1731717309
#> 347 347 1 -0.152292131  0.072231250  0.2383800682 -0.2512710956
#> 348 348 1  0.206850624  0.302705890 -0.2071519476  0.4382229205
#> 349 349 1 -0.141771321  0.289831018  0.9672760250  0.4152481650
#> 350 350 1  0.230132143 -0.239452271  0.2209621876 -0.9808333720
#> 351 351 1  0.387451247  0.757188609 -0.1755128344  1.7283931034
#> 352 352 1 -0.067926416  0.591328134  0.4471357012  1.2024670152
#> 353 353 1  0.009872636  0.796002330  0.4044809119  1.7766664512
#> 354 354 1  0.415390506  1.042312012  0.0767551387  2.5351452783
#> 355 355 1  0.323682555  0.347647465  0.1187459790  0.6286357693
#> 356 356 1  0.423832515  0.208666379 -0.0669487140  0.2638472572
#> 357 357 1  0.254325900  0.098028568 -0.0495688243 -0.0860827382
#> 358 358 1  0.387725542  0.030210824  0.3820193699 -0.1855430492
#> 359 359 1  0.164523162 -0.042313002  0.2992791680 -0.4586968049
#> 360 360 1  0.238846620  0.356717937  0.0311417631  0.6184386480
#> 361 361 1 -0.213423432  0.472609728  0.4706630986  0.8385983396
#> 362 362 1  0.343486733  0.698013251  0.5112244534  1.6238719869
#> 363 363 1  0.089036980  0.503871109  0.0545475280  0.9738141692
#> 364 364 1 -0.380043239  0.478814295  0.3642396473  0.7937282135
#> 365 365 1  0.272788147  0.112347381  0.1347534537 -0.0230296782
#> 366 366 1  0.131427425  0.215637194 -0.1201988491  0.1879590631
#> 367 367 1 -0.263971836  0.387201732 -0.2272421417  0.5212385333
#> 368 368 1  0.307700342 -0.093441811  0.1121586297 -0.5723724721
#> 369 369 1  0.206131332  0.474791129 -0.1321437952  0.9119641754
#> 370 370 1  0.295312378  0.089784414 -0.0234964338 -0.0932630900
#> 371 371 1  0.042603669  0.342694383  0.2311664651  0.5406125086
#> 372 372 1  0.014094029  0.140711664 -0.1138146162 -0.0503390583
#> 373 373 1 -0.263137432 -0.243161490 -0.1204677217 -1.1761654142
#> 374 374 1  0.399992985  0.381519646  0.0554686552  0.7373849243
#> 375 375 1  0.350571705  0.185003725  0.0305894137  0.1871581833
#> 376 376 1  0.605058553  0.166017666  0.0285400737  0.2133077040
#> 377 377 1 -0.244787236  0.077224411  0.2062867172 -0.2692588128
#> 378 378 1  0.164562337  0.350167580 -0.3883036470  0.5356444913
#> 379 379 1  0.424163640  0.588282648 -0.0299767025  1.2965297187
#> 380 380 1  0.029799807  0.280996999 -0.2671095409  0.3192002877
#> 381 381 1 -0.290697531  0.461738702  0.0873997002  0.7468312113
#> 382 382 1  0.360474845 -0.093768707  0.2070093609 -0.5475501629
#> 383 383 1  0.219007280  0.217038700  0.4492651661  0.2760028020
#> 384 384 1  0.131450758  0.175683118  0.2467875665  0.1167149834
#> 385 385 1  0.097810415  0.166493455 -0.0006756695  0.0565504510
#> 386 386 1  0.234648986  0.333316661 -0.4007832511  0.5101451548
#> 387 387 1 -0.039356310  0.279502245 -0.0023382214  0.3207226146
#> 388 388 1  0.500207672  0.416220404  0.4217624465  0.8990382452
#> 389 389 1  0.788938973  0.280432992  0.1498404438  0.5918653488
#> 390 390 1  0.562801355  0.220804906  0.0358144224  0.3496074632
#> 391 391 1  0.323099509 -0.474470990  0.2006806690 -1.5914104660
#> 392 392 1  0.454385475  0.027241645  0.0156572008 -0.2101788101
#> 393 393 1  0.022793064  0.299952381  0.5215069010  0.4480133704
#> 394 394 1  0.089532961  0.438720859 -0.0375469407  0.7880996125
#> 395 395 1  0.062522066  0.584974716 -0.4262179243  1.1370012437
#> 396 396 1 -0.059565784  0.388217897  0.0619769240  0.6156787659
#> 397 397 1  0.002229328  0.330266742  0.4547076378  0.5171438914
#> 398 398 1  0.063725529  0.477963007  0.3716579342  0.9278583818
#> 399 399 1  0.213832661  0.136268591  0.4830880929  0.0589275907
#> 400 400 1 -0.254104260  0.600340801 -0.2716689662  1.0974299288
#> 401 401 1  0.423976570  0.358983474 -0.0085245095  0.6771821129
#> 402 402 1  0.156108280  0.613935141  0.6951687931  1.3572690862
#> 403 403 1  0.120624999  0.033950684  0.1683104725 -0.2786430916
#> 404 404 1  0.037750822  0.412336442  0.0116748635  0.7057255698
#> 405 405 1  0.526984055  0.023391373  0.2253698985 -0.1772530826
#> 406 406 1  0.486246059  0.186557178 -0.2145830539  0.2081116467
#> 407 407 1 -0.052129022  0.583654495 -0.1113406477  1.1301422087
#> 408 408 1  0.154460855  0.537644884 -0.1673243298  1.0629635958
#> 409 409 1  0.324152474  0.454432108  0.0234746429  0.9085770511
#> 410 410 1  0.411316733  0.508447982  0.2149303463  1.1009450691
#> 411 411 1  0.225837074  0.587369133 -0.0908761888  1.2272670743
#> 412 412 1 -0.042756141  0.431228543 -0.2658683261  0.7043036698
#> 413 413 1  0.175054440  0.539215542  0.4797240438  1.1388130457
#> 414 414 1  0.076668212  0.562653013  0.3052918947  1.1546486090
#> 415 415 1  0.424758668 -0.122900601  0.0883844136 -0.6188196370
#> 416 416 1  0.411354524  0.634291443 -0.1324727834  1.4069646852
#> 417 417 1  0.463935633  0.544592885 -0.0457434972  1.1886898702
#> 418 418 1  0.174756919  0.114581173  0.1612199571 -0.0442792242
#> 419 419 1  0.303292493  0.222071454  0.0228532414  0.2723835430
#> 420 420 1  0.106725963 -0.071207553  0.5395324197 -0.5304354808
#> 421 421 1  0.366848482  0.297424613  0.9461391608  0.5892135253
#> 422 422 1  0.186934464  0.333170433  0.2145559123  0.5572553568
#> 423 423 1  0.691171382  0.555611614 -0.0932905023  1.2832357540
#> 424 424 1  0.163584026 -0.022499035  0.5646936502 -0.3784995444
#> 425 425 1  0.196284342  0.051634152  0.3639278032 -0.1878423536
#> 426 426 1  0.331668581  0.116989738  0.0364326703 -0.0023661179
#> 427 427 1  0.161736772  0.326915298 -0.1472836933  0.4960834535
#> 428 428 1  0.292541294  0.954588200  0.2738308252  2.2797175433
#> 429 429 1 -0.219902402  0.541180161 -0.3328394069  0.9413753454
#> 430 430 1  0.561582334  0.473963288 -0.2328368677  1.0082435737
#> 431 431 1 -0.048306930  0.517619958  0.3518606119  0.9990847710
#> 432 432 1  0.552578422  0.047541360 -0.2339522172 -0.1503267575
#> 433 433 1  0.340760583  0.436717986 -0.1084637819  0.8523313986
#> 434 434 1  0.080299544  0.154758082  0.1733375147  0.0369508495
#> 435 435 1  0.453056317  0.382527015 -0.0668701190  0.7439944059
#> 436 436 1  0.276617173  0.289832687 -0.0659281262  0.4389169874
#> 437 437 1  0.052343719  0.641770598 -0.0945176310  1.3212923762
#> 438 438 1  0.490829547 -0.073040018  0.2333850941 -0.4488509809
#> 439 439 1  0.341076275  0.616401951  0.1550396459  1.3660051537
#> 440 440 1  0.161083338 -0.157661750 -0.0457069092 -0.8071847778
#> 441 441 1  0.120687837  0.811095252 -0.0608624774  1.8044958639
#> 442 442 1  0.556566359  0.263162689  0.0146480670  0.4603647071
#> 443 443 1  0.471261811  0.201561280  0.5611098190  0.3224724260
#> 444 444 1  0.344146729 -0.082343379  0.5274929138 -0.4892363065
#> 445 445 1  0.232887464  0.336723779  0.1472839749  0.5741479320
#> 446 446 1 -0.326102120  0.703955901  0.1084848725  1.3945996544
#> 447 447 1  0.550176973  0.392921830  0.2041149542  0.8292088836
#> 448 448 1  0.395578173  0.182249737 -0.0446920630  0.1858586998
#> 449 449 1  0.393409817  0.453663485  0.1581148265  0.9412573287
#> 450 450 1  0.137325915  0.137789144  0.2591642398  0.0170586994
#> 451 451 1 -0.018684758  0.279135381  0.1930823554  0.3457694947
#> 452 452 1  0.065286137  0.348231610 -0.4044713473  0.4984108752
#> 453 453 1  0.630842078  0.183897638 -0.0625825150  0.2604548789
#> 454 454 1 -0.270142155  0.145994032  0.5051963927 -0.0604654731
#> 455 455 1 -0.126256887  0.176565523  0.1080909651  0.0263123133
#> 456 456 1  0.544092885  0.163499552 -0.1572108244  0.1690971356
#> 457 457 1  0.178830813  0.271538858  0.2241323220  0.3887077297
#> 458 458 1  0.134221980  0.135870164 -0.1472243485 -0.0301012897
#> 459 459 1  0.023456919 -0.044027373  0.4135601638 -0.4949437368
#> 460 460 1  0.101529161  0.043613192  0.1928564631 -0.2558172148
#> 461 461 1  0.108936461  0.264076603  0.3727302843  0.3621075442
#> 462 462 1  0.337081756  0.027149939 -0.0928226725 -0.2572417379
#> 463 463 1  0.167257847  0.394153309 -0.2495147422  0.6696861076
#> 464 464 1  0.362663121  0.094628099 -0.4291944638 -0.1004824221
#> 465 465 1  0.300646711  0.455798580 -0.1033187512  0.8922981410
#> 466 466 1 -0.267716003  0.261961352 -0.2702456548  0.1763229004
#> 467 467 1  0.199223615  0.453027960 -0.2038826728  0.8436292769
#> 468 468 1  0.218184556  0.102698974  0.4810955303 -0.0309244628
#> 469 469 1 -0.202844119 -0.164073442  0.0287986773 -0.9283194944
#> 470 470 1  0.425321425  0.432074546 -0.4634413980  0.8297803067
#> 471 471 1 -0.168498679  0.597595677 -0.5178060179  1.0913266323
#> 472 472 1  0.601291009  0.131714529 -0.1973120818  0.0963945816
#> 473 473 1  0.112446390  0.647555320  0.3950149976  1.4047479787
#> 474 474 1  0.107125008  0.432028971  0.0492498676  0.7841009869
#> 475 475 1 -0.190250815  0.669278266  0.2110933019  1.3525080974
#> 476 476 1  0.028377595  0.272155724 -0.2499281322  0.2965373021
#> 477 477 1  0.268067808  0.281510686 -0.4305742309  0.3769506416
#> 478 478 1  0.131169657  0.212659821  0.4473257888  0.2370816596
#> 479 479 1  0.434081946  0.332236209  0.3705175455  0.6460309335
#> 480 480 1 -0.176954807 -0.097506827 -0.1982094518 -0.7629002683
#> 481 481 1 -0.232948396 -0.056075492  0.4812947455 -0.5991613352
#> 482 482 1  0.262907945 -0.099785253  0.2847873814 -0.5858400910
#> 483 483 1 -0.200595535 -0.033395141  0.2773401333 -0.5483815778
#> 484 484 1  0.463836179  0.498084231 -0.0545169888  1.0617246017
#> 485 485 1  0.077376278  0.290040375  0.2120507018  0.4066108373
#> 486 486 1  0.389921221  0.056364778  0.5051158390 -0.1015661855
#> 487 487 1  0.589952619  0.302386112 -0.0535980760  0.5699908882
#> 488 488 1  0.008014426  0.609065767  0.3708864668  1.2660656661
#> 489 489 1  0.177152122  0.062421651  0.2134261458 -0.1796432476
#> 490 490 1  0.139458570  0.079076014  0.2851816241 -0.1387902351
#> 491 491 1  0.139216518  0.427545508 -0.1122076660  0.7654689652
#> 492 492 1  0.072866421 -0.028346827 -0.1805448634 -0.4972911754
#> 493 493 1  0.092385063  0.505021511  0.3642597970  1.0092098528
#> 494 494 1  0.338266376 -0.246603644  0.2017351792 -0.9690919369
#> 495 495 1  0.209351781  0.319314549 -0.3750222364  0.4670607465
#> 496 496 1  0.118749834  1.140651738  0.2825105214  2.7317305495
#> 497 497 1  0.090723700  0.120404958  0.1641119336 -0.0538979236
#> 498 498 1  0.091211650  0.198749816  0.0940990422  0.1515192127
#> 499 499 1 -0.079448914  0.392459439  0.0487288937  0.6197578943
#> 500 500 1  0.096625724  0.067969929 -0.1057706049 -0.2214394220