Declare potential outcomes

Declare potential outcomes

declare_potential_outcomes(
  ...,
  handler = potential_outcomes_handler,
  label = NULL
)

potential_outcomes_internal.formula(
  formula,
  conditions = c(0, 1),
  assignment_variables = "Z",
  data,
  level = NULL,
  label = outcome_variable
)

potential_outcomes_internal.NULL(
  formula = stop("Not provided"),
  ...,
  data,
  level = NULL
)

Arguments

...
arguments to be captured, and later passed to the handler

handler
a tidy-in, tidy-out function

label
a string describing the step

formula
a formula to calculate potential outcomes as functions of assignment variables.

conditions
see expand_conditions. Provide values (e.g. conditions = 1:4) for a single assignment variable. If multiple assignment variables, provide named list (e.g. conditions = list(Z1 = 0:1, Z2 = 0:1)). Defaults to 0:1 if no conditions provided.

assignment_variables
The name of the assignment variable. Generally not required as names are taken from conditions.

data
a data.frame

level
a character specifying a level of hierarchy for fabricate to calculate at

Value

a function that returns a data.frame

Details

A declare_potential_outcomes function is used to create outcomes that each unit would express in each possible treatment condition.

Examples

# Potential outcomes can be declared in two ways: 
# by using a formula or as separate variables.

# Using a formula
declare_population(N = 100, U = rnorm(N)) +
  declare_potential_outcomes(Y ~ 0.5*Z + U)

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, U = rnorm(N)) -----------------
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: U 
#>    min median  mean  max   sd N_missing N_unique
#>  -2.75  -0.09 -0.09 2.42 1.09         0      100
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y ~ 0.5 * Z + U) -------
#> 
#> Formula: Y ~ 0.5 * Z + U 
#> 
#> Added variable: Y_Z_0 
#>    min median  mean  max   sd N_missing N_unique
#>  -2.75  -0.09 -0.09 2.42 1.09         0      100
#> 
#> Added variable: Y_Z_1 
#>    min median mean  max   sd N_missing N_unique
#>  -2.25   0.41 0.41 2.92 1.09         0      100
#> 

  
# As separate variables
declare_population(N = 100, U = rnorm(N)) +
  declare_potential_outcomes(Y_Z_0 = U,
                             Y_Z_1 = U + 0.5)

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, U = rnorm(N)) -----------------
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: U 
#>    min median mean  max   sd N_missing N_unique
#>  -2.36   0.05 0.01 1.76 0.98         0      100
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y_Z_0 = U, Y_Z_1 = U + 0.5) 
#> 
#> Added variable: Y_Z_0 
#>    min median mean  max   sd N_missing N_unique
#>  -2.36   0.05 0.01 1.76 0.98         0      100
#> 
#> Added variable: Y_Z_1 
#>    min median mean  max   sd N_missing N_unique
#>  -1.86   0.55 0.51 2.26 0.98         0      100
#> 

# (notice the naming structure: outcome_assignment_condition: Y_Z_1)  

  
# You can change the name of the outcome
declare_population(N = 100, U = rnorm(N)) +
  declare_potential_outcomes(Y2 ~ 0.5*Z + U)

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, U = rnorm(N)) -----------------
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: U 
#>   min median mean  max   sd N_missing N_unique
#>  -3.6   0.14 0.03 1.99 0.97         0      100
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y2 ~ 0.5 * Z + U) ------
#> 
#> Formula: Y2 ~ 0.5 * Z + U 
#> 
#> Added variable: Y2_Z_0 
#>   min median mean  max   sd N_missing N_unique
#>  -3.6   0.14 0.03 1.99 0.97         0      100
#> 
#> Added variable: Y2_Z_1 
#>   min median mean  max   sd N_missing N_unique
#>  -3.1   0.64 0.53 2.49 0.97         0      100
#> 

  
# You can change the name of the assignment_variable
declare_population(N = 100, U = rnorm(N)) +
  declare_potential_outcomes(Y ~ 0.5*D + U, assignment_variable = "D")

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, U = rnorm(N)) -----------------
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: U 
#>    min median  mean  max   sd N_missing N_unique
#>  -2.07   0.06 -0.04 2.48 0.91         0      100
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y ~ 0.5 * D + U, assignment_variable = "D") 
#> 
#> Formula: Y ~ 0.5 * D + U 
#> 
#> Added variable: Y_D_0 
#>    min median  mean  max   sd N_missing N_unique
#>  -2.07   0.06 -0.04 2.48 0.91         0      100
#> 
#> Added variable: Y_D_1 
#>    min median mean  max   sd N_missing N_unique
#>  -1.57   0.56 0.46 2.98 0.91         0      100
#> 

  

# `conditions` defines the "range" of the potential outcomes function
declare_population(N = 100, age = sample(18:65, N, replace = TRUE)) +
  declare_potential_outcomes(formula = Y ~ .05 + .25 * Z + .01 * age * Z,
                             conditions = 1:4)

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, age = sample(18:65, N, replace = TRUE)) 
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: age 
#>  min median  mean max    sd N_missing N_unique
#>   18   43.5 42.22  65 13.18         0       46
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(formula = Y ~ 0.05 + 0.25 * Z + 0.01 *     age * Z, conditions = 1:4) 
#> 
#> Formula: Y ~ 0.05 + 0.25 * Z + 0.01 * age * Z 
#> 
#> Added variable: Y_Z_1 
#>   min median mean  max   sd N_missing N_unique
#>  0.48   0.74 0.72 0.95 0.13         0       46
#> 
#> Added variable: Y_Z_2 
#>   min median mean  max   sd N_missing N_unique
#>  0.91   1.42 1.39 1.85 0.26         0       46
#> 
#> Added variable: Y_Z_3 
#>   min median mean  max  sd N_missing N_unique
#>  1.34    2.1 2.07 2.75 0.4         0       46
#> 
#> Added variable: Y_Z_4 
#>   min median mean  max   sd N_missing N_unique
#>  1.77   2.79 2.74 3.65 0.53         0       46
#> 

# Multiple assignment variables can be specified in `conditions`. For example,
# in a 2x2 factorial potential outcome:

declare_population(N = 100, age = sample(18:65, N, replace = TRUE)) +
  declare_potential_outcomes(formula = Y ~ .05 + .25 * Z1 + .01 * age * Z2,
                             conditions = list(Z1 = 0:1, Z2 = 0:1))

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, age = sample(18:65, N, replace = TRUE)) 
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: age 
#>  min median mean max    sd N_missing N_unique
#>   18   40.5 39.2  65 13.29         0       42
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(formula = Y ~ 0.05 + 0.25 * Z1 + 0.01 *     age * Z2, conditions = list(Z1 = 0:1, Z2 = 0:1)) 
#> 
#> Formula: Y ~ 0.05 + 0.25 * Z1 + 0.01 * age * Z2 
#> 
#> Added variable: Y_Z1_0_Z2_0 
#>  0.05
#>   100
#>  1.00
#> 
#> Added variable: Y_Z1_1_Z2_0 
#>   0.3
#>   100
#>  1.00
#> 
#> Added variable: Y_Z1_0_Z2_1 
#>   min median mean max   sd N_missing N_unique
#>  0.23   0.46 0.44 0.7 0.13         0       42
#> 
#> Added variable: Y_Z1_1_Z2_1 
#>   min median mean  max   sd N_missing N_unique
#>  0.48    0.7 0.69 0.95 0.13         0       42
#>