Get estimates, estimands, assignment vectors, or samples from a design given data

Get estimates, estimands, assignment vectors, or samples from a design given data

get_estimates(design, data = NULL, start = 1, end = length(design))

draw_assignment(design, data = NULL, start = 1, end = length(design))

draw_sample(design, data = NULL, start = 1, end = length(design))

Arguments

design
A design object, typically created using the + operator

data
A data.frame object with sufficient information to get the data, estimates, estimands, an assignment vector, or a sample.

start
(Defaults to 1) a scalar indicating which step in the design to begin with. By default all data steps are drawn, from step 1 to the last step of the design.

end
(Defaults to length(design)) a scalar indicating which step in the design to finish with.

Examples

design <- declare_population(N = 100, u = rnorm(N)) +
  declare_potential_outcomes(Y ~ Z + u) +
  declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(n = 75) +
  declare_assignment(m = 50) +
  reveal_outcomes(Y, Z) +
  declare_estimator(Y ~ Z, estimand = "ATE")

dat <- draw_data(design)

draw_data(design, data = dat, start = 2)

#>     ID           u       Y_Z_0       Y_Z_1 S_inclusion_prob Z Z_cond_prob
#> 1  001 -0.51979611 -0.51979611  0.48020389                1 1   0.6666667
#> 2  002 -0.13264612 -0.13264612  0.86735388                1 1   0.6666667
#> 3  003 -0.52255268 -0.52255268  0.47744732                1 0   0.3333333
#> 4  004  2.63278919  2.63278919  3.63278919                1 1   0.6666667
#> 5  005 -0.09344709 -0.09344709  0.90655291                1 1   0.6666667
#> 6  006  0.34989607  0.34989607  1.34989607                1 0   0.3333333
#> 7  007  0.74039870  0.74039870  1.74039870                1 1   0.6666667
#> 8  008 -0.37191521 -0.37191521  0.62808479                1 0   0.3333333
#> 9  010  1.13625019  1.13625019  2.13625019                1 1   0.6666667
#> 10 012 -0.86090056 -0.86090056  0.13909944                1 1   0.6666667
#> 11 013  1.17239914  1.17239914  2.17239914                1 1   0.6666667
#> 12 014 -0.73922234 -0.73922234  0.26077766                1 0   0.3333333
#> 13 016  0.07877580  0.07877580  1.07877580                1 0   0.3333333
#> 14 017 -0.56752746 -0.56752746  0.43247254                1 1   0.6666667
#> 15 018  0.38294164  0.38294164  1.38294164                1 1   0.6666667
#> 16 019 -1.49740740 -1.49740740 -0.49740740                1 1   0.6666667
#> 17 020  0.49630726  0.49630726  1.49630726                1 0   0.3333333
#> 18 021 -1.01739566 -1.01739566 -0.01739566                1 1   0.6666667
#> 19 023  0.95483502  0.95483502  1.95483502                1 1   0.6666667
#> 20 024 -1.03694606 -1.03694606 -0.03694606                1 1   0.6666667
#> 21 025  0.04277836  0.04277836  1.04277836                1 0   0.3333333
#> 22 027 -2.31588271 -2.31588271 -1.31588271                1 1   0.6666667
#> 23 028 -0.53409030 -0.53409030  0.46590970                1 0   0.3333333
#> 24 029 -0.30850671 -0.30850671  0.69149329                1 0   0.3333333
#> 25 031 -0.45516167 -0.45516167  0.54483833                1 0   0.3333333
#> 26 032 -0.43145450 -0.43145450  0.56854550                1 0   0.3333333
#> 27 033  0.05970660  0.05970660  1.05970660                1 0   0.3333333
#> 28 034  0.08242633  0.08242633  1.08242633                1 1   0.6666667
#> 29 036  0.42072046  0.42072046  1.42072046                1 1   0.6666667
#> 30 037  0.57957101  0.57957101  1.57957101                1 0   0.3333333
#> 31 038 -0.30059058 -0.30059058  0.69940942                1 0   0.3333333
#> 32 039 -1.08443287 -1.08443287 -0.08443287                1 1   0.6666667
#> 33 041  0.58635030  0.58635030  1.58635030                1 0   0.3333333
#> 34 042  2.07239306  2.07239306  3.07239306                1 1   0.6666667
#> 35 043  0.31096860  0.31096860  1.31096860                1 1   0.6666667
#> 36 044  1.14985228  1.14985228  2.14985228                1 1   0.6666667
#> 37 045 -0.28359526 -0.28359526  0.71640474                1 0   0.3333333
#> 38 046  0.66621454  0.66621454  1.66621454                1 1   0.6666667
#> 39 047  0.60092926  0.60092926  1.60092926                1 0   0.3333333
#> 40 048 -0.31310830 -0.31310830  0.68689170                1 1   0.6666667
#> 41 049 -1.90353003 -1.90353003 -0.90353003                1 0   0.3333333
#> 42 051  0.32661517  0.32661517  1.32661517                1 0   0.3333333
#> 43 052 -1.34042234 -1.34042234 -0.34042234                1 1   0.6666667
#> 44 053 -0.82449350 -0.82449350  0.17550650                1 0   0.3333333
#> 45 054 -0.57687152 -0.57687152  0.42312848                1 1   0.6666667
#> 46 055 -0.14616876 -0.14616876  0.85383124                1 0   0.3333333
#> 47 056 -0.57479533 -0.57479533  0.42520467                1 1   0.6666667
#> 48 058  0.32945415  0.32945415  1.32945415                1 0   0.3333333
#> 49 059 -1.22809661 -1.22809661 -0.22809661                1 1   0.6666667
#> 50 060  0.88945072  0.88945072  1.88945072                1 1   0.6666667
#> 51 061  1.00387199  1.00387199  2.00387199                1 1   0.6666667
#> 52 062  2.12083694  2.12083694  3.12083694                1 1   0.6666667
#> 53 063  0.32126235  0.32126235  1.32126235                1 1   0.6666667
#> 54 064  2.56209810  2.56209810  3.56209810                1 1   0.6666667
#> 55 065  1.02869982  1.02869982  2.02869982                1 1   0.6666667
#> 56 066 -0.46991179 -0.46991179  0.53008821                1 1   0.6666667
#> 57 067  0.61592753  0.61592753  1.61592753                1 1   0.6666667
#> 58 068 -0.72461316 -0.72461316  0.27538684                1 1   0.6666667
#> 59 070  0.04572569  0.04572569  1.04572569                1 1   0.6666667
#> 60 071 -0.83314033 -0.83314033  0.16685967                1 0   0.3333333
#> 61 074  2.35496457  2.35496457  3.35496457                1 1   0.6666667
#> 62 076  0.34855173  0.34855173  1.34855173                1 0   0.3333333
#> 63 077  0.94349207  0.94349207  1.94349207                1 1   0.6666667
#> 64 079 -0.29468103 -0.29468103  0.70531897                1 1   0.6666667
#> 65 080  0.53233468  0.53233468  1.53233468                1 1   0.6666667
#> 66 082 -0.12516370 -0.12516370  0.87483630                1 1   0.6666667
#> 67 085 -1.32168985 -1.32168985 -0.32168985                1 1   0.6666667
#> 68 086  0.35721991  0.35721991  1.35721991                1 1   0.6666667
#> 69 087  1.19207207  1.19207207  2.19207207                1 1   0.6666667
#> 70 088  0.54761879  0.54761879  1.54761879                1 1   0.6666667
#> 71 090 -0.04909723 -0.04909723  0.95090277                1 1   0.6666667
#> 72 092 -0.90829756 -0.90829756  0.09170244                1 1   0.6666667
#> 73 095  1.41737087  1.41737087  2.41737087                1 0   0.3333333
#> 74 096 -1.01582558 -1.01582558 -0.01582558                1 1   0.6666667
#> 75 100 -0.32880811 -0.32880811  0.67119189                1 1   0.6666667
#>              Y
#> 1   0.48020389
#> 2   0.86735388
#> 3  -0.52255268
#> 4   3.63278919
#> 5   0.90655291
#> 6   0.34989607
#> 7   1.74039870
#> 8  -0.37191521
#> 9   2.13625019
#> 10  0.13909944
#> 11  2.17239914
#> 12 -0.73922234
#> 13  0.07877580
#> 14  0.43247254
#> 15  1.38294164
#> 16 -0.49740740
#> 17  0.49630726
#> 18 -0.01739566
#> 19  1.95483502
#> 20 -0.03694606
#> 21  0.04277836
#> 22 -1.31588271
#> 23 -0.53409030
#> 24 -0.30850671
#> 25 -0.45516167
#> 26 -0.43145450
#> 27  0.05970660
#> 28  1.08242633
#> 29  1.42072046
#> 30  0.57957101
#> 31 -0.30059058
#> 32 -0.08443287
#> 33  0.58635030
#> 34  3.07239306
#> 35  1.31096860
#> 36  2.14985228
#> 37 -0.28359526
#> 38  1.66621454
#> 39  0.60092926
#> 40  0.68689170
#> 41 -1.90353003
#> 42  0.32661517
#> 43 -0.34042234
#> 44 -0.82449350
#> 45  0.42312848
#> 46 -0.14616876
#> 47  0.42520467
#> 48  0.32945415
#> 49 -0.22809661
#> 50  1.88945072
#> 51  2.00387199
#> 52  3.12083694
#> 53  1.32126235
#> 54  3.56209810
#> 55  2.02869982
#> 56  0.53008821
#> 57  1.61592753
#> 58  0.27538684
#> 59  1.04572569
#> 60 -0.83314033
#> 61  3.35496457
#> 62  0.34855173
#> 63  1.94349207
#> 64  0.70531897
#> 65  1.53233468
#> 66  0.87483630
#> 67 -0.32168985
#> 68  1.35721991
#> 69  2.19207207
#> 70  1.54761879
#> 71  0.95090277
#> 72  0.09170244
#> 73  1.41737087
#> 74 -0.01582558
#> 75  0.67119189

get_estimates(design, data = dat)

#>   estimator_label term estimate std.error statistic      p.value  conf.low
#> 1       estimator    Z 1.158243 0.2280352  5.079228 4.441659e-06 0.7015092
#>   conf.high       df outcome estimand_label
#> 1  1.614976 56.42459       Y            ATE

draw_assignment(design, data = dat)

#>     ID           u       Y_Z_0       Y_Z_1 S_inclusion_prob Z Z_cond_prob
#> 1  001 -0.51979611 -0.51979611  0.48020389             0.75 0   0.3333333
#> 2  002 -0.13264612 -0.13264612  0.86735388             0.75 0   0.3333333
#> 3  003 -0.52255268 -0.52255268  0.47744732             0.75 1   0.6666667
#> 4  004  2.63278919  2.63278919  3.63278919             0.75 0   0.3333333
#> 5  005 -0.09344709 -0.09344709  0.90655291             0.75 1   0.6666667
#> 6  006  0.34989607  0.34989607  1.34989607             0.75 1   0.6666667
#> 7  007  0.74039870  0.74039870  1.74039870             0.75 1   0.6666667
#> 8  008 -0.37191521 -0.37191521  0.62808479             0.75 1   0.6666667
#> 9  010  1.13625019  1.13625019  2.13625019             0.75 1   0.6666667
#> 10 012 -0.86090056 -0.86090056  0.13909944             0.75 0   0.3333333
#> 11 013  1.17239914  1.17239914  2.17239914             0.75 1   0.6666667
#> 12 014 -0.73922234 -0.73922234  0.26077766             0.75 1   0.6666667
#> 13 016  0.07877580  0.07877580  1.07877580             0.75 1   0.6666667
#> 14 017 -0.56752746 -0.56752746  0.43247254             0.75 1   0.6666667
#> 15 018  0.38294164  0.38294164  1.38294164             0.75 1   0.6666667
#> 16 019 -1.49740740 -1.49740740 -0.49740740             0.75 1   0.6666667
#> 17 020  0.49630726  0.49630726  1.49630726             0.75 1   0.6666667
#> 18 021 -1.01739566 -1.01739566 -0.01739566             0.75 0   0.3333333
#> 19 023  0.95483502  0.95483502  1.95483502             0.75 1   0.6666667
#> 20 024 -1.03694606 -1.03694606 -0.03694606             0.75 0   0.3333333
#> 21 025  0.04277836  0.04277836  1.04277836             0.75 1   0.6666667
#> 22 027 -2.31588271 -2.31588271 -1.31588271             0.75 0   0.3333333
#> 23 028 -0.53409030 -0.53409030  0.46590970             0.75 1   0.6666667
#> 24 029 -0.30850671 -0.30850671  0.69149329             0.75 0   0.3333333
#> 25 031 -0.45516167 -0.45516167  0.54483833             0.75 1   0.6666667
#> 26 032 -0.43145450 -0.43145450  0.56854550             0.75 1   0.6666667
#> 27 033  0.05970660  0.05970660  1.05970660             0.75 1   0.6666667
#> 28 034  0.08242633  0.08242633  1.08242633             0.75 1   0.6666667
#> 29 036  0.42072046  0.42072046  1.42072046             0.75 0   0.3333333
#> 30 037  0.57957101  0.57957101  1.57957101             0.75 1   0.6666667
#> 31 038 -0.30059058 -0.30059058  0.69940942             0.75 1   0.6666667
#> 32 039 -1.08443287 -1.08443287 -0.08443287             0.75 0   0.3333333
#> 33 041  0.58635030  0.58635030  1.58635030             0.75 1   0.6666667
#> 34 042  2.07239306  2.07239306  3.07239306             0.75 0   0.3333333
#> 35 043  0.31096860  0.31096860  1.31096860             0.75 0   0.3333333
#> 36 044  1.14985228  1.14985228  2.14985228             0.75 0   0.3333333
#> 37 045 -0.28359526 -0.28359526  0.71640474             0.75 1   0.6666667
#> 38 046  0.66621454  0.66621454  1.66621454             0.75 1   0.6666667
#> 39 047  0.60092926  0.60092926  1.60092926             0.75 1   0.6666667
#> 40 048 -0.31310830 -0.31310830  0.68689170             0.75 1   0.6666667
#> 41 049 -1.90353003 -1.90353003 -0.90353003             0.75 1   0.6666667
#> 42 051  0.32661517  0.32661517  1.32661517             0.75 1   0.6666667
#> 43 052 -1.34042234 -1.34042234 -0.34042234             0.75 0   0.3333333
#> 44 053 -0.82449350 -0.82449350  0.17550650             0.75 0   0.3333333
#> 45 054 -0.57687152 -0.57687152  0.42312848             0.75 1   0.6666667
#> 46 055 -0.14616876 -0.14616876  0.85383124             0.75 1   0.6666667
#> 47 056 -0.57479533 -0.57479533  0.42520467             0.75 0   0.3333333
#> 48 058  0.32945415  0.32945415  1.32945415             0.75 1   0.6666667
#> 49 059 -1.22809661 -1.22809661 -0.22809661             0.75 1   0.6666667
#> 50 060  0.88945072  0.88945072  1.88945072             0.75 1   0.6666667
#> 51 061  1.00387199  1.00387199  2.00387199             0.75 1   0.6666667
#> 52 062  2.12083694  2.12083694  3.12083694             0.75 1   0.6666667
#> 53 063  0.32126235  0.32126235  1.32126235             0.75 1   0.6666667
#> 54 064  2.56209810  2.56209810  3.56209810             0.75 0   0.3333333
#> 55 065  1.02869982  1.02869982  2.02869982             0.75 0   0.3333333
#> 56 066 -0.46991179 -0.46991179  0.53008821             0.75 0   0.3333333
#> 57 067  0.61592753  0.61592753  1.61592753             0.75 1   0.6666667
#> 58 068 -0.72461316 -0.72461316  0.27538684             0.75 1   0.6666667
#> 59 070  0.04572569  0.04572569  1.04572569             0.75 1   0.6666667
#> 60 071 -0.83314033 -0.83314033  0.16685967             0.75 1   0.6666667
#> 61 074  2.35496457  2.35496457  3.35496457             0.75 1   0.6666667
#> 62 076  0.34855173  0.34855173  1.34855173             0.75 1   0.6666667
#> 63 077  0.94349207  0.94349207  1.94349207             0.75 0   0.3333333
#> 64 079 -0.29468103 -0.29468103  0.70531897             0.75 0   0.3333333
#> 65 080  0.53233468  0.53233468  1.53233468             0.75 1   0.6666667
#> 66 082 -0.12516370 -0.12516370  0.87483630             0.75 1   0.6666667
#> 67 085 -1.32168985 -1.32168985 -0.32168985             0.75 0   0.3333333
#> 68 086  0.35721991  0.35721991  1.35721991             0.75 0   0.3333333
#> 69 087  1.19207207  1.19207207  2.19207207             0.75 1   0.6666667
#> 70 088  0.54761879  0.54761879  1.54761879             0.75 1   0.6666667
#> 71 090 -0.04909723 -0.04909723  0.95090277             0.75 1   0.6666667
#> 72 092 -0.90829756 -0.90829756  0.09170244             0.75 0   0.3333333
#> 73 095  1.41737087  1.41737087  2.41737087             0.75 0   0.3333333
#> 74 096 -1.01582558 -1.01582558 -0.01582558             0.75 1   0.6666667
#> 75 100 -0.32880811 -0.32880811  0.67119189             0.75 1   0.6666667
#>              Y
#> 1  -0.51979611
#> 2   0.86735388
#> 3  -0.52255268
#> 4   3.63278919
#> 5   0.90655291
#> 6   0.34989607
#> 7   1.74039870
#> 8   0.62808479
#> 9   1.13625019
#> 10  0.13909944
#> 11  2.17239914
#> 12  0.26077766
#> 13  1.07877580
#> 14 -0.56752746
#> 15  1.38294164
#> 16 -0.49740740
#> 17  1.49630726
#> 18 -0.01739566
#> 19  1.95483502
#> 20 -0.03694606
#> 21  1.04277836
#> 22 -2.31588271
#> 23 -0.53409030
#> 24 -0.30850671
#> 25  0.54483833
#> 26  0.56854550
#> 27  1.05970660
#> 28  1.08242633
#> 29  1.42072046
#> 30  1.57957101
#> 31 -0.30059058
#> 32 -0.08443287
#> 33  0.58635030
#> 34  3.07239306
#> 35  1.31096860
#> 36  2.14985228
#> 37  0.71640474
#> 38  0.66621454
#> 39  1.60092926
#> 40  0.68689170
#> 41 -0.90353003
#> 42  1.32661517
#> 43 -1.34042234
#> 44  0.17550650
#> 45 -0.57687152
#> 46  0.85383124
#> 47  0.42520467
#> 48  1.32945415
#> 49 -0.22809661
#> 50  1.88945072
#> 51  2.00387199
#> 52  3.12083694
#> 53  0.32126235
#> 54  3.56209810
#> 55  1.02869982
#> 56  0.53008821
#> 57  0.61592753
#> 58  0.27538684
#> 59  1.04572569
#> 60 -0.83314033
#> 61  3.35496457
#> 62  1.34855173
#> 63  1.94349207
#> 64  0.70531897
#> 65  0.53233468
#> 66 -0.12516370
#> 67 -0.32168985
#> 68  1.35721991
#> 69  1.19207207
#> 70  0.54761879
#> 71 -0.04909723
#> 72 -0.90829756
#> 73  1.41737087
#> 74 -0.01582558
#> 75 -0.32880811

draw_sample(design, data = dat)

#>     ID           u       Y_Z_0       Y_Z_1 S_inclusion_prob Z Z_cond_prob
#> 1  001 -0.51979611 -0.51979611  0.48020389                1 0   0.3333333
#> 2  002 -0.13264612 -0.13264612  0.86735388                1 1   0.6666667
#> 3  003 -0.52255268 -0.52255268  0.47744732                1 0   0.3333333
#> 4  004  2.63278919  2.63278919  3.63278919                1 1   0.6666667
#> 5  005 -0.09344709 -0.09344709  0.90655291                1 1   0.6666667
#> 6  006  0.34989607  0.34989607  1.34989607                1 0   0.3333333
#> 7  007  0.74039870  0.74039870  1.74039870                1 1   0.6666667
#> 8  008 -0.37191521 -0.37191521  0.62808479                1 1   0.6666667
#> 9  010  1.13625019  1.13625019  2.13625019                1 0   0.3333333
#> 10 012 -0.86090056 -0.86090056  0.13909944                1 1   0.6666667
#> 11 013  1.17239914  1.17239914  2.17239914                1 1   0.6666667
#> 12 014 -0.73922234 -0.73922234  0.26077766                1 1   0.6666667
#> 13 016  0.07877580  0.07877580  1.07877580                1 1   0.6666667
#> 14 017 -0.56752746 -0.56752746  0.43247254                1 0   0.3333333
#> 15 018  0.38294164  0.38294164  1.38294164                1 1   0.6666667
#> 16 019 -1.49740740 -1.49740740 -0.49740740                1 1   0.6666667
#> 17 020  0.49630726  0.49630726  1.49630726                1 1   0.6666667
#> 18 021 -1.01739566 -1.01739566 -0.01739566                1 1   0.6666667
#> 19 023  0.95483502  0.95483502  1.95483502                1 1   0.6666667
#> 20 024 -1.03694606 -1.03694606 -0.03694606                1 1   0.6666667
#> 21 025  0.04277836  0.04277836  1.04277836                1 1   0.6666667
#> 22 027 -2.31588271 -2.31588271 -1.31588271                1 0   0.3333333
#> 23 028 -0.53409030 -0.53409030  0.46590970                1 0   0.3333333
#> 24 029 -0.30850671 -0.30850671  0.69149329                1 0   0.3333333
#> 25 031 -0.45516167 -0.45516167  0.54483833                1 1   0.6666667
#> 26 032 -0.43145450 -0.43145450  0.56854550                1 1   0.6666667
#> 27 033  0.05970660  0.05970660  1.05970660                1 1   0.6666667
#> 28 034  0.08242633  0.08242633  1.08242633                1 1   0.6666667
#> 29 036  0.42072046  0.42072046  1.42072046                1 1   0.6666667
#> 30 037  0.57957101  0.57957101  1.57957101                1 1   0.6666667
#> 31 038 -0.30059058 -0.30059058  0.69940942                1 0   0.3333333
#> 32 039 -1.08443287 -1.08443287 -0.08443287                1 1   0.6666667
#> 33 041  0.58635030  0.58635030  1.58635030                1 0   0.3333333
#> 34 042  2.07239306  2.07239306  3.07239306                1 1   0.6666667
#> 35 043  0.31096860  0.31096860  1.31096860                1 1   0.6666667
#> 36 044  1.14985228  1.14985228  2.14985228                1 1   0.6666667
#> 37 045 -0.28359526 -0.28359526  0.71640474                1 1   0.6666667
#> 38 046  0.66621454  0.66621454  1.66621454                1 0   0.3333333
#> 39 047  0.60092926  0.60092926  1.60092926                1 1   0.6666667
#> 40 048 -0.31310830 -0.31310830  0.68689170                1 1   0.6666667
#> 41 049 -1.90353003 -1.90353003 -0.90353003                1 1   0.6666667
#> 42 051  0.32661517  0.32661517  1.32661517                1 1   0.6666667
#> 43 052 -1.34042234 -1.34042234 -0.34042234                1 0   0.3333333
#> 44 053 -0.82449350 -0.82449350  0.17550650                1 1   0.6666667
#> 45 054 -0.57687152 -0.57687152  0.42312848                1 0   0.3333333
#> 46 055 -0.14616876 -0.14616876  0.85383124                1 1   0.6666667
#> 47 056 -0.57479533 -0.57479533  0.42520467                1 1   0.6666667
#> 48 058  0.32945415  0.32945415  1.32945415                1 1   0.6666667
#> 49 059 -1.22809661 -1.22809661 -0.22809661                1 1   0.6666667
#> 50 060  0.88945072  0.88945072  1.88945072                1 1   0.6666667
#> 51 061  1.00387199  1.00387199  2.00387199                1 1   0.6666667
#> 52 062  2.12083694  2.12083694  3.12083694                1 1   0.6666667
#> 53 063  0.32126235  0.32126235  1.32126235                1 0   0.3333333
#> 54 064  2.56209810  2.56209810  3.56209810                1 1   0.6666667
#> 55 065  1.02869982  1.02869982  2.02869982                1 0   0.3333333
#> 56 066 -0.46991179 -0.46991179  0.53008821                1 1   0.6666667
#> 57 067  0.61592753  0.61592753  1.61592753                1 0   0.3333333
#> 58 068 -0.72461316 -0.72461316  0.27538684                1 1   0.6666667
#> 59 070  0.04572569  0.04572569  1.04572569                1 1   0.6666667
#> 60 071 -0.83314033 -0.83314033  0.16685967                1 0   0.3333333
#> 61 074  2.35496457  2.35496457  3.35496457                1 1   0.6666667
#> 62 076  0.34855173  0.34855173  1.34855173                1 1   0.6666667
#> 63 077  0.94349207  0.94349207  1.94349207                1 1   0.6666667
#> 64 079 -0.29468103 -0.29468103  0.70531897                1 1   0.6666667
#> 65 080  0.53233468  0.53233468  1.53233468                1 0   0.3333333
#> 66 082 -0.12516370 -0.12516370  0.87483630                1 0   0.3333333
#> 67 085 -1.32168985 -1.32168985 -0.32168985                1 1   0.6666667
#> 68 086  0.35721991  0.35721991  1.35721991                1 1   0.6666667
#> 69 087  1.19207207  1.19207207  2.19207207                1 0   0.3333333
#> 70 088  0.54761879  0.54761879  1.54761879                1 0   0.3333333
#> 71 090 -0.04909723 -0.04909723  0.95090277                1 0   0.3333333
#> 72 092 -0.90829756 -0.90829756  0.09170244                1 0   0.3333333
#> 73 095  1.41737087  1.41737087  2.41737087                1 0   0.3333333
#> 74 096 -1.01582558 -1.01582558 -0.01582558                1 1   0.6666667
#> 75 100 -0.32880811 -0.32880811  0.67119189                1 0   0.3333333
#>              Y
#> 1  -0.51979611
#> 2   0.86735388
#> 3  -0.52255268
#> 4   3.63278919
#> 5   0.90655291
#> 6   0.34989607
#> 7   1.74039870
#> 8   0.62808479
#> 9   1.13625019
#> 10  0.13909944
#> 11  2.17239914
#> 12  0.26077766
#> 13  1.07877580
#> 14 -0.56752746
#> 15  1.38294164
#> 16 -0.49740740
#> 17  1.49630726
#> 18 -0.01739566
#> 19  1.95483502
#> 20 -0.03694606
#> 21  1.04277836
#> 22 -2.31588271
#> 23 -0.53409030
#> 24 -0.30850671
#> 25  0.54483833
#> 26  0.56854550
#> 27  1.05970660
#> 28  1.08242633
#> 29  1.42072046
#> 30  1.57957101
#> 31 -0.30059058
#> 32 -0.08443287
#> 33  0.58635030
#> 34  3.07239306
#> 35  1.31096860
#> 36  2.14985228
#> 37  0.71640474
#> 38  0.66621454
#> 39  1.60092926
#> 40  0.68689170
#> 41 -0.90353003
#> 42  1.32661517
#> 43 -1.34042234
#> 44  0.17550650
#> 45 -0.57687152
#> 46  0.85383124
#> 47  0.42520467
#> 48  1.32945415
#> 49 -0.22809661
#> 50  1.88945072
#> 51  2.00387199
#> 52  3.12083694
#> 53  0.32126235
#> 54  3.56209810
#> 55  1.02869982
#> 56  0.53008821
#> 57  0.61592753
#> 58  0.27538684
#> 59  1.04572569
#> 60 -0.83314033
#> 61  3.35496457
#> 62  1.34855173
#> 63  1.94349207
#> 64  0.70531897
#> 65  0.53233468
#> 66 -0.12516370
#> 67 -0.32168985
#> 68  1.35721991
#> 69  1.19207207
#> 70  0.54761879
#> 71 -0.04909723
#> 72 -0.90829756
#> 73  1.41737087
#> 74 -0.01582558
#> 75 -0.32880811