Get estimates, inquiries, assignment vectors, or samples from a design given data

Get estimates, inquiries, assignment vectors, or samples from a design given data

get_estimates(design, data = NULL, start = 1, end = length(design))

draw_assignment(design, data = NULL, start = 1, end = length(design))

draw_sample(design, data = NULL, start = 1, end = length(design))

Arguments

design
A design object, typically created using the + operator

data
A data.frame object with sufficient information to get the data, estimates, inquiries, an assignment vector, or a sample.

start
(Defaults to 1) a scalar indicating which step in the design to begin with. By default all data steps are drawn, from step 1 to the last step of the design.

end
(Defaults to length(design)) a scalar indicating which step in the design to finish with.

Examples

design <- 
  declare_model(
    N = 100, 
    U = rnorm(N),
    potential_outcomes(Y ~ Z + U)
  ) +
  declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(S = complete_rs(N, n = 75)) +
  declare_assignment(Z = complete_ra(N, m = 50)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z)) +
  declare_estimator(Y ~ Z, inquiry = "ATE")

dat <- draw_data(design)

draw_data(design, data = dat, start = 2)

#>     ID             U         Y_Z_0       Y_Z_1 S Z             Y
#> 1  001  1.141986e-01  1.141986e-01  1.11419864 1 1  1.114199e+00
#> 2  002 -6.229922e-01 -6.229922e-01  0.37700775 1 1  3.770078e-01
#> 3  004 -1.787814e+00 -1.787814e+00 -0.78781386 1 1 -7.878139e-01
#> 4  005 -6.022824e-01 -6.022824e-01  0.39771755 1 1  3.977176e-01
#> 5  006 -9.427656e-01 -9.427656e-01  0.05723435 1 0 -9.427656e-01
#> 6  007 -5.190466e-01 -5.190466e-01  0.48095339 1 1  4.809534e-01
#> 7  009 -1.090758e+00 -1.090758e+00 -0.09075827 1 1 -9.075827e-02
#> 8  010 -1.427400e+00 -1.427400e+00 -0.42740029 1 1 -4.274003e-01
#> 9  011  7.209845e-01  7.209845e-01  1.72098454 1 1  1.720985e+00
#> 10 012  3.549768e-02  3.549768e-02  1.03549768 1 1  1.035498e+00
#> 11 013 -1.409914e+00 -1.409914e+00 -0.40991436 1 1 -4.099144e-01
#> 12 014 -3.113561e-01 -3.113561e-01  0.68864385 1 1  6.886439e-01
#> 13 015  6.927429e-01  6.927429e-01  1.69274289 1 1  1.692743e+00
#> 14 016 -7.559372e-01 -7.559372e-01  0.24406284 1 1  2.440628e-01
#> 15 019 -1.440200e+00 -1.440200e+00 -0.44019996 1 1 -4.402000e-01
#> 16 020 -3.203963e-01 -3.203963e-01  0.67960365 1 1  6.796037e-01
#> 17 021  6.326422e-01  6.326422e-01  1.63264220 1 0  6.326422e-01
#> 18 023 -8.171946e-01 -8.171946e-01  0.18280538 1 1  1.828054e-01
#> 19 024  1.179221e-01  1.179221e-01  1.11792208 1 1  1.117922e+00
#> 20 025 -1.083568e+00 -1.083568e+00 -0.08356795 1 1 -8.356795e-02
#> 21 026 -6.852012e-01 -6.852012e-01  0.31479878 1 1  3.147988e-01
#> 22 027  1.221910e-01  1.221910e-01  1.12219099 1 0  1.221910e-01
#> 23 028 -1.223529e+00 -1.223529e+00 -0.22352919 1 0 -1.223529e+00
#> 24 029  7.976117e-01  7.976117e-01  1.79761169 1 0  7.976117e-01
#> 25 030 -2.597459e-01 -2.597459e-01  0.74025408 1 0 -2.597459e-01
#> 26 031 -1.804010e+00 -1.804010e+00 -0.80401040 1 1 -8.040104e-01
#> 27 032  6.610451e-02  6.610451e-02  1.06610451 1 1  1.066105e+00
#> 28 033  1.833007e-01  1.833007e-01  1.18330070 1 0  1.833007e-01
#> 29 034 -1.204074e+00 -1.204074e+00 -0.20407389 1 0 -1.204074e+00
#> 30 035  1.253185e+00  1.253185e+00  2.25318540 1 1  2.253185e+00
#> 31 037  9.988050e-02  9.988050e-02  1.09988050 1 1  1.099880e+00
#> 32 042 -2.080653e-01 -2.080653e-01  0.79193470 1 0 -2.080653e-01
#> 33 043 -4.019219e-01 -4.019219e-01  0.59807811 1 0 -4.019219e-01
#> 34 045  5.497174e-01  5.497174e-01  1.54971742 1 1  1.549717e+00
#> 35 047 -3.203339e-01 -3.203339e-01  0.67966610 1 1  6.796661e-01
#> 36 048  1.460633e+00  1.460633e+00  2.46063321 1 1  2.460633e+00
#> 37 049 -2.191702e+00 -2.191702e+00 -1.19170196 1 0 -2.191702e+00
#> 38 051 -1.214878e+00 -1.214878e+00 -0.21487832 1 1 -2.148783e-01
#> 39 053  6.390103e-01  6.390103e-01  1.63901030 1 1  1.639010e+00
#> 40 055 -1.106432e+00 -1.106432e+00 -0.10643238 1 1 -1.064324e-01
#> 41 056  8.746173e-02  8.746173e-02  1.08746173 1 0  8.746173e-02
#> 42 057  1.813206e-01  1.813206e-01  1.18132059 1 1  1.181321e+00
#> 43 058  3.181688e-01  3.181688e-01  1.31816881 1 1  1.318169e+00
#> 44 060  2.874167e-02  2.874167e-02  1.02874167 1 0  2.874167e-02
#> 45 061 -9.196080e-01 -9.196080e-01  0.08039198 1 1  8.039198e-02
#> 46 062  1.198677e-01  1.198677e-01  1.11986774 1 1  1.119868e+00
#> 47 064  1.108648e+00  1.108648e+00  2.10864771 1 1  2.108648e+00
#> 48 065 -5.972366e-02 -5.972366e-02  0.94027634 1 0 -5.972366e-02
#> 49 066 -1.073280e+00 -1.073280e+00 -0.07328024 1 0 -1.073280e+00
#> 50 067 -1.692392e+00 -1.692392e+00 -0.69239225 1 1 -6.923922e-01
#> 51 068  1.024947e+00  1.024947e+00  2.02494725 1 1  2.024947e+00
#> 52 069  2.966568e-01  2.966568e-01  1.29665676 1 1  1.296657e+00
#> 53 070  2.317201e-01  2.317201e-01  1.23172010 1 0  2.317201e-01
#> 54 072  2.797606e-06  2.797606e-06  1.00000280 1 0  2.797606e-06
#> 55 073  1.632610e+00  1.632610e+00  2.63260962 1 0  1.632610e+00
#> 56 074  1.501509e+00  1.501509e+00  2.50150912 1 1  2.501509e+00
#> 57 076  5.106322e-01  5.106322e-01  1.51063216 1 0  5.106322e-01
#> 58 080 -8.705967e-01 -8.705967e-01  0.12940331 1 0 -8.705967e-01
#> 59 081  6.416779e-01  6.416779e-01  1.64167789 1 1  1.641678e+00
#> 60 082 -6.129715e-01 -6.129715e-01  0.38702850 1 0 -6.129715e-01
#> 61 084 -4.325426e-01 -4.325426e-01  0.56745745 1 1  5.674574e-01
#> 62 085 -5.280970e-01 -5.280970e-01  0.47190301 1 1  4.719030e-01
#> 63 087  8.031862e-01  8.031862e-01  1.80318616 1 1  1.803186e+00
#> 64 088  1.281532e+00  1.281532e+00  2.28153243 1 1  2.281532e+00
#> 65 089 -4.205917e-01 -4.205917e-01  0.57940829 1 1  5.794083e-01
#> 66 090 -1.123699e+00 -1.123699e+00 -0.12369863 1 1 -1.236986e-01
#> 67 091 -1.186798e+00 -1.186798e+00 -0.18679819 1 1 -1.867982e-01
#> 68 092  1.235737e+00  1.235737e+00  2.23573685 1 0  1.235737e+00
#> 69 093 -4.121528e-01 -4.121528e-01  0.58784720 1 1  5.878472e-01
#> 70 094  8.327458e-01  8.327458e-01  1.83274583 1 1  1.832746e+00
#> 71 095  1.198868e+00  1.198868e+00  2.19886767 1 1  2.198868e+00
#> 72 096  1.618880e+00  1.618880e+00  2.61888016 1 0  1.618880e+00
#> 73 098  1.254281e-01  1.254281e-01  1.12542811 1 1  1.125428e+00
#> 74 099  6.833356e-01  6.833356e-01  1.68333560 1 0  6.833356e-01
#> 75 100 -6.135761e-01 -6.135761e-01  0.38642386 1 0 -6.135761e-01

get_estimates(design, data = dat)

#>   estimator term  estimate std.error statistic    p.value  conf.low conf.high
#> 1 estimator    Z 0.5641645 0.2187808  2.578674 0.01193294 0.1281349  1.000194
#>   df outcome inquiry
#> 1 73       Y     ATE

draw_assignment(design, data = dat)

#>     ID             U         Y_Z_0       Y_Z_1 S Z           Y
#> 1  001  1.141986e-01  1.141986e-01  1.11419864 1 0  0.11419864
#> 2  002 -6.229922e-01 -6.229922e-01  0.37700775 1 1 -0.62299225
#> 3  004 -1.787814e+00 -1.787814e+00 -0.78781386 1 1 -0.78781386
#> 4  005 -6.022824e-01 -6.022824e-01  0.39771755 1 0  0.39771755
#> 5  006 -9.427656e-01 -9.427656e-01  0.05723435 1 1  0.05723435
#> 6  007 -5.190466e-01 -5.190466e-01  0.48095339 1 0  0.48095339
#> 7  009 -1.090758e+00 -1.090758e+00 -0.09075827 1 0 -0.09075827
#> 8  010 -1.427400e+00 -1.427400e+00 -0.42740029 1 1 -0.42740029
#> 9  011  7.209845e-01  7.209845e-01  1.72098454 1 1  1.72098454
#> 10 012  3.549768e-02  3.549768e-02  1.03549768 1 1  0.03549768
#> 11 013 -1.409914e+00 -1.409914e+00 -0.40991436 1 0 -0.40991436
#> 12 014 -3.113561e-01 -3.113561e-01  0.68864385 1 0  0.68864385
#> 13 015  6.927429e-01  6.927429e-01  1.69274289 1 1  0.69274289
#> 14 016 -7.559372e-01 -7.559372e-01  0.24406284 1 0  0.24406284
#> 15 019 -1.440200e+00 -1.440200e+00 -0.44019996 1 1 -0.44019996
#> 16 020 -3.203963e-01 -3.203963e-01  0.67960365 1 0  0.67960365
#> 17 021  6.326422e-01  6.326422e-01  1.63264220 1 0  1.63264220
#> 18 023 -8.171946e-01 -8.171946e-01  0.18280538 1 1 -0.81719462
#> 19 024  1.179221e-01  1.179221e-01  1.11792208 1 1  1.11792208
#> 20 025 -1.083568e+00 -1.083568e+00 -0.08356795 1 1 -0.08356795
#> 21 026 -6.852012e-01 -6.852012e-01  0.31479878 1 0 -0.68520122
#> 22 027  1.221910e-01  1.221910e-01  1.12219099 1 1  1.12219099
#> 23 028 -1.223529e+00 -1.223529e+00 -0.22352919 1 1 -0.22352919
#> 24 029  7.976117e-01  7.976117e-01  1.79761169 1 1  1.79761169
#> 25 030 -2.597459e-01 -2.597459e-01  0.74025408 1 1  0.74025408
#> 26 031 -1.804010e+00 -1.804010e+00 -0.80401040 1 0 -1.80401040
#> 27 032  6.610451e-02  6.610451e-02  1.06610451 1 1  0.06610451
#> 28 033  1.833007e-01  1.833007e-01  1.18330070 1 1  1.18330070
#> 29 034 -1.204074e+00 -1.204074e+00 -0.20407389 1 0 -0.20407389
#> 30 035  1.253185e+00  1.253185e+00  2.25318540 1 0  2.25318540
#> 31 037  9.988050e-02  9.988050e-02  1.09988050 1 0  1.09988050
#> 32 042 -2.080653e-01 -2.080653e-01  0.79193470 1 1  0.79193470
#> 33 043 -4.019219e-01 -4.019219e-01  0.59807811 1 1 -0.40192189
#> 34 045  5.497174e-01  5.497174e-01  1.54971742 1 1  1.54971742
#> 35 047 -3.203339e-01 -3.203339e-01  0.67966610 1 0 -0.32033390
#> 36 048  1.460633e+00  1.460633e+00  2.46063321 1 1  1.46063321
#> 37 049 -2.191702e+00 -2.191702e+00 -1.19170196 1 0 -1.19170196
#> 38 051 -1.214878e+00 -1.214878e+00 -0.21487832 1 0 -0.21487832
#> 39 053  6.390103e-01  6.390103e-01  1.63901030 1 1  0.63901030
#> 40 055 -1.106432e+00 -1.106432e+00 -0.10643238 1 1 -0.10643238
#> 41 056  8.746173e-02  8.746173e-02  1.08746173 1 1  0.08746173
#> 42 057  1.813206e-01  1.813206e-01  1.18132059 1 0  0.18132059
#> 43 058  3.181688e-01  3.181688e-01  1.31816881 1 1  1.31816881
#> 44 060  2.874167e-02  2.874167e-02  1.02874167 1 0  1.02874167
#> 45 061 -9.196080e-01 -9.196080e-01  0.08039198 1 1  0.08039198
#> 46 062  1.198677e-01  1.198677e-01  1.11986774 1 1  1.11986774
#> 47 064  1.108648e+00  1.108648e+00  2.10864771 1 1  1.10864771
#> 48 065 -5.972366e-02 -5.972366e-02  0.94027634 1 0  0.94027634
#> 49 066 -1.073280e+00 -1.073280e+00 -0.07328024 1 1 -0.07328024
#> 50 067 -1.692392e+00 -1.692392e+00 -0.69239225 1 1 -1.69239225
#> 51 068  1.024947e+00  1.024947e+00  2.02494725 1 1  1.02494725
#> 52 069  2.966568e-01  2.966568e-01  1.29665676 1 1  0.29665676
#> 53 070  2.317201e-01  2.317201e-01  1.23172010 1 0  1.23172010
#> 54 072  2.797606e-06  2.797606e-06  1.00000280 1 1  1.00000280
#> 55 073  1.632610e+00  1.632610e+00  2.63260962 1 1  1.63260962
#> 56 074  1.501509e+00  1.501509e+00  2.50150912 1 1  2.50150912
#> 57 076  5.106322e-01  5.106322e-01  1.51063216 1 1  1.51063216
#> 58 080 -8.705967e-01 -8.705967e-01  0.12940331 1 0  0.12940331
#> 59 081  6.416779e-01  6.416779e-01  1.64167789 1 1  1.64167789
#> 60 082 -6.129715e-01 -6.129715e-01  0.38702850 1 1  0.38702850
#> 61 084 -4.325426e-01 -4.325426e-01  0.56745745 1 0  0.56745745
#> 62 085 -5.280970e-01 -5.280970e-01  0.47190301 1 1  0.47190301
#> 63 087  8.031862e-01  8.031862e-01  1.80318616 1 0  0.80318616
#> 64 088  1.281532e+00  1.281532e+00  2.28153243 1 0  2.28153243
#> 65 089 -4.205917e-01 -4.205917e-01  0.57940829 1 1 -0.42059171
#> 66 090 -1.123699e+00 -1.123699e+00 -0.12369863 1 1 -0.12369863
#> 67 091 -1.186798e+00 -1.186798e+00 -0.18679819 1 1 -0.18679819
#> 68 092  1.235737e+00  1.235737e+00  2.23573685 1 1  2.23573685
#> 69 093 -4.121528e-01 -4.121528e-01  0.58784720 1 1 -0.41215280
#> 70 094  8.327458e-01  8.327458e-01  1.83274583 1 1  0.83274583
#> 71 095  1.198868e+00  1.198868e+00  2.19886767 1 1  1.19886767
#> 72 096  1.618880e+00  1.618880e+00  2.61888016 1 1  2.61888016
#> 73 098  1.254281e-01  1.254281e-01  1.12542811 1 1  1.12542811
#> 74 099  6.833356e-01  6.833356e-01  1.68333560 1 1  0.68333560
#> 75 100 -6.135761e-01 -6.135761e-01  0.38642386 1 1  0.38642386

draw_sample(design, data = dat)

#>     ID             U         Y_Z_0       Y_Z_1 S Z           Y
#> 1  001  1.141986e-01  1.141986e-01  1.11419864 1 0  0.11419864
#> 2  002 -6.229922e-01 -6.229922e-01  0.37700775 1 0 -0.62299225
#> 3  004 -1.787814e+00 -1.787814e+00 -0.78781386 1 1 -0.78781386
#> 4  005 -6.022824e-01 -6.022824e-01  0.39771755 1 1  0.39771755
#> 5  006 -9.427656e-01 -9.427656e-01  0.05723435 1 1  0.05723435
#> 6  007 -5.190466e-01 -5.190466e-01  0.48095339 1 1  0.48095339
#> 7  009 -1.090758e+00 -1.090758e+00 -0.09075827 1 1 -0.09075827
#> 8  010 -1.427400e+00 -1.427400e+00 -0.42740029 1 1 -0.42740029
#> 9  011  7.209845e-01  7.209845e-01  1.72098454 1 1  1.72098454
#> 10 012  3.549768e-02  3.549768e-02  1.03549768 1 0  0.03549768
#> 11 013 -1.409914e+00 -1.409914e+00 -0.40991436 1 1 -0.40991436
#> 12 014 -3.113561e-01 -3.113561e-01  0.68864385 1 1  0.68864385
#> 13 015  6.927429e-01  6.927429e-01  1.69274289 1 0  0.69274289
#> 14 016 -7.559372e-01 -7.559372e-01  0.24406284 1 1  0.24406284
#> 15 019 -1.440200e+00 -1.440200e+00 -0.44019996 1 1 -0.44019996
#> 16 020 -3.203963e-01 -3.203963e-01  0.67960365 1 1  0.67960365
#> 17 021  6.326422e-01  6.326422e-01  1.63264220 1 1  1.63264220
#> 18 023 -8.171946e-01 -8.171946e-01  0.18280538 1 0 -0.81719462
#> 19 024  1.179221e-01  1.179221e-01  1.11792208 1 1  1.11792208
#> 20 025 -1.083568e+00 -1.083568e+00 -0.08356795 1 1 -0.08356795
#> 21 026 -6.852012e-01 -6.852012e-01  0.31479878 1 0 -0.68520122
#> 22 027  1.221910e-01  1.221910e-01  1.12219099 1 1  1.12219099
#> 23 028 -1.223529e+00 -1.223529e+00 -0.22352919 1 1 -0.22352919
#> 24 029  7.976117e-01  7.976117e-01  1.79761169 1 1  1.79761169
#> 25 030 -2.597459e-01 -2.597459e-01  0.74025408 1 1  0.74025408
#> 26 031 -1.804010e+00 -1.804010e+00 -0.80401040 1 0 -1.80401040
#> 27 032  6.610451e-02  6.610451e-02  1.06610451 1 0  0.06610451
#> 28 033  1.833007e-01  1.833007e-01  1.18330070 1 1  1.18330070
#> 29 034 -1.204074e+00 -1.204074e+00 -0.20407389 1 1 -0.20407389
#> 30 035  1.253185e+00  1.253185e+00  2.25318540 1 1  2.25318540
#> 31 037  9.988050e-02  9.988050e-02  1.09988050 1 1  1.09988050
#> 32 042 -2.080653e-01 -2.080653e-01  0.79193470 1 1  0.79193470
#> 33 043 -4.019219e-01 -4.019219e-01  0.59807811 1 0 -0.40192189
#> 34 045  5.497174e-01  5.497174e-01  1.54971742 1 1  1.54971742
#> 35 047 -3.203339e-01 -3.203339e-01  0.67966610 1 0 -0.32033390
#> 36 048  1.460633e+00  1.460633e+00  2.46063321 1 0  1.46063321
#> 37 049 -2.191702e+00 -2.191702e+00 -1.19170196 1 1 -1.19170196
#> 38 051 -1.214878e+00 -1.214878e+00 -0.21487832 1 1 -0.21487832
#> 39 053  6.390103e-01  6.390103e-01  1.63901030 1 0  0.63901030
#> 40 055 -1.106432e+00 -1.106432e+00 -0.10643238 1 1 -0.10643238
#> 41 056  8.746173e-02  8.746173e-02  1.08746173 1 0  0.08746173
#> 42 057  1.813206e-01  1.813206e-01  1.18132059 1 0  0.18132059
#> 43 058  3.181688e-01  3.181688e-01  1.31816881 1 1  1.31816881
#> 44 060  2.874167e-02  2.874167e-02  1.02874167 1 1  1.02874167
#> 45 061 -9.196080e-01 -9.196080e-01  0.08039198 1 1  0.08039198
#> 46 062  1.198677e-01  1.198677e-01  1.11986774 1 1  1.11986774
#> 47 064  1.108648e+00  1.108648e+00  2.10864771 1 0  1.10864771
#> 48 065 -5.972366e-02 -5.972366e-02  0.94027634 1 1  0.94027634
#> 49 066 -1.073280e+00 -1.073280e+00 -0.07328024 1 1 -0.07328024
#> 50 067 -1.692392e+00 -1.692392e+00 -0.69239225 1 0 -1.69239225
#> 51 068  1.024947e+00  1.024947e+00  2.02494725 1 0  1.02494725
#> 52 069  2.966568e-01  2.966568e-01  1.29665676 1 0  0.29665676
#> 53 070  2.317201e-01  2.317201e-01  1.23172010 1 1  1.23172010
#> 54 072  2.797606e-06  2.797606e-06  1.00000280 1 1  1.00000280
#> 55 073  1.632610e+00  1.632610e+00  2.63260962 1 0  1.63260962
#> 56 074  1.501509e+00  1.501509e+00  2.50150912 1 1  2.50150912
#> 57 076  5.106322e-01  5.106322e-01  1.51063216 1 1  1.51063216
#> 58 080 -8.705967e-01 -8.705967e-01  0.12940331 1 1  0.12940331
#> 59 081  6.416779e-01  6.416779e-01  1.64167789 1 1  1.64167789
#> 60 082 -6.129715e-01 -6.129715e-01  0.38702850 1 1  0.38702850
#> 61 084 -4.325426e-01 -4.325426e-01  0.56745745 1 1  0.56745745
#> 62 085 -5.280970e-01 -5.280970e-01  0.47190301 1 1  0.47190301
#> 63 087  8.031862e-01  8.031862e-01  1.80318616 1 0  0.80318616
#> 64 088  1.281532e+00  1.281532e+00  2.28153243 1 1  2.28153243
#> 65 089 -4.205917e-01 -4.205917e-01  0.57940829 1 0 -0.42059171
#> 66 090 -1.123699e+00 -1.123699e+00 -0.12369863 1 1 -0.12369863
#> 67 091 -1.186798e+00 -1.186798e+00 -0.18679819 1 1 -0.18679819
#> 68 092  1.235737e+00  1.235737e+00  2.23573685 1 1  2.23573685
#> 69 093 -4.121528e-01 -4.121528e-01  0.58784720 1 0 -0.41215280
#> 70 094  8.327458e-01  8.327458e-01  1.83274583 1 0  0.83274583
#> 71 095  1.198868e+00  1.198868e+00  2.19886767 1 0  1.19886767
#> 72 096  1.618880e+00  1.618880e+00  2.61888016 1 1  2.61888016
#> 73 098  1.254281e-01  1.254281e-01  1.12542811 1 1  1.12542811
#> 74 099  6.833356e-01  6.833356e-01  1.68333560 1 0  0.68333560
#> 75 100 -6.135761e-01 -6.135761e-01  0.38642386 1 1  0.38642386