Get estimates, estimands, assignment vectors, or samples from a design given data

Get estimates, estimands, assignment vectors, or samples from a design given data

get_estimates(design, data = NULL, start = 1, end = length(design))

draw_assignment(design, data = NULL, start = 1, end = length(design))

draw_sample(design, data = NULL, start = 1, end = length(design))

Arguments

design
A design object, typically created using the + operator

data
A data.frame object with sufficient information to get the data, estimates, estimands, an assignment vector, or a sample.

start
(Defaults to 1) a scalar indicating which step in the design to begin with. By default all data steps are drawn, from step 1 to the last step of the design.

end
(Defaults to length(design)) a scalar indicating which step in the design to finish with.

Examples

design <- declare_population(N = 100, u = rnorm(N)) +
  declare_potential_outcomes(Y ~ Z + u) +
  declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(n = 75) +
  declare_assignment(m = 50) +
  reveal_outcomes(Y, Z) +
  declare_estimator(Y ~ Z, estimand = "ATE")

dat <- draw_data(design)

draw_data(design, data = dat, start = 2)

#>     ID            u        Y_Z_0       Y_Z_1 S_inclusion_prob Z Z_cond_prob
#> 1  002  0.390191956  0.390191956  1.39019196                1 0   0.3333333
#> 2  003  1.871733816  1.871733816  2.87173382                1 1   0.6666667
#> 3  004 -0.488486270 -0.488486270  0.51151373                1 0   0.3333333
#> 4  005 -0.248771151 -0.248771151  0.75122885                1 0   0.3333333
#> 5  008  1.706144150  1.706144150  2.70614415                1 0   0.3333333
#> 6  011 -1.123982219 -1.123982219 -0.12398222                1 1   0.6666667
#> 7  012  1.589642326  1.589642326  2.58964233                1 1   0.6666667
#> 8  013 -0.779077571 -0.779077571  0.22092243                1 1   0.6666667
#> 9  014 -0.976976771 -0.976976771  0.02302323                1 1   0.6666667
#> 10 015  0.525384860  0.525384860  1.52538486                1 1   0.6666667
#> 11 018 -0.195638053 -0.195638053  0.80436195                1 0   0.3333333
#> 12 019  1.189035632  1.189035632  2.18903563                1 1   0.6666667
#> 13 020 -0.155748325 -0.155748325  0.84425167                1 1   0.6666667
#> 14 021 -1.248727456 -1.248727456 -0.24872746                1 1   0.6666667
#> 15 022  0.430798701  0.430798701  1.43079870                1 1   0.6666667
#> 16 024 -0.341035256 -0.341035256  0.65896474                1 0   0.3333333
#> 17 025 -2.444042764 -2.444042764 -1.44404276                1 0   0.3333333
#> 18 026  0.682177629  0.682177629  1.68217763                1 0   0.3333333
#> 19 027  1.649690872  1.649690872  2.64969087                1 1   0.6666667
#> 20 028 -0.036813680 -0.036813680  0.96318632                1 1   0.6666667
#> 21 030  0.061242727  0.061242727  1.06124273                1 0   0.3333333
#> 22 031  1.463657329  1.463657329  2.46365733                1 1   0.6666667
#> 23 033 -2.455706303 -2.455706303 -1.45570630                1 1   0.6666667
#> 24 034  0.437567170  0.437567170  1.43756717                1 1   0.6666667
#> 25 036 -1.277149137 -1.277149137 -0.27714914                1 1   0.6666667
#> 26 037  0.109574799  0.109574799  1.10957480                1 1   0.6666667
#> 27 038  0.048517072  0.048517072  1.04851707                1 0   0.3333333
#> 28 039 -1.066930718 -1.066930718 -0.06693072                1 1   0.6666667
#> 29 040 -0.169838706 -0.169838706  0.83016129                1 0   0.3333333
#> 30 041 -0.782161548 -0.782161548  0.21783845                1 0   0.3333333
#> 31 043  0.100579496  0.100579496  1.10057950                1 1   0.6666667
#> 32 044 -0.595647705 -0.595647705  0.40435230                1 0   0.3333333
#> 33 045  0.337330988  0.337330988  1.33733099                1 1   0.6666667
#> 34 046  0.691659765  0.691659765  1.69165977                1 1   0.6666667
#> 35 047  0.520009139  0.520009139  1.52000914                1 0   0.3333333
#> 36 048  0.262511486  0.262511486  1.26251149                1 0   0.3333333
#> 37 049  1.340805674  1.340805674  2.34080567                1 1   0.6666667
#> 38 050 -1.466248992 -1.466248992 -0.46624899                1 0   0.3333333
#> 39 052  1.508366928  1.508366928  2.50836693                1 1   0.6666667
#> 40 053 -0.056882551 -0.056882551  0.94311745                1 0   0.3333333
#> 41 054 -0.473658670 -0.473658670  0.52634133                1 1   0.6666667
#> 42 055 -0.230609318 -0.230609318  0.76939068                1 1   0.6666667
#> 43 056 -0.975971783 -0.975971783  0.02402822                1 1   0.6666667
#> 44 057 -1.519560103 -1.519560103 -0.51956010                1 1   0.6666667
#> 45 058 -1.132115492 -1.132115492 -0.13211549                1 1   0.6666667
#> 46 059  1.732456622  1.732456622  2.73245662                1 1   0.6666667
#> 47 060  1.456327813  1.456327813  2.45632781                1 0   0.3333333
#> 48 061  0.136631624  0.136631624  1.13663162                1 1   0.6666667
#> 49 062 -2.527272469 -2.527272469 -1.52727247                1 1   0.6666667
#> 50 064 -1.922911742 -1.922911742 -0.92291174                1 1   0.6666667
#> 51 065 -2.166600121 -2.166600121 -1.16660012                1 1   0.6666667
#> 52 067 -1.224366171 -1.224366171 -0.22436617                1 0   0.3333333
#> 53 068 -2.367837184 -2.367837184 -1.36783718                1 1   0.6666667
#> 54 069  0.222758476  0.222758476  1.22275848                1 0   0.3333333
#> 55 071 -0.890538330 -0.890538330  0.10946167                1 1   0.6666667
#> 56 072  1.865303159  1.865303159  2.86530316                1 1   0.6666667
#> 57 073 -0.098196677 -0.098196677  0.90180332                1 1   0.6666667
#> 58 074 -0.397300713 -0.397300713  0.60269929                1 0   0.3333333
#> 59 075 -0.209680583 -0.209680583  0.79031942                1 0   0.3333333
#> 60 076 -0.345801617 -0.345801617  0.65419838                1 0   0.3333333
#> 61 077  0.005740578  0.005740578  1.00574058                1 1   0.6666667
#> 62 079  0.052601925  0.052601925  1.05260193                1 1   0.6666667
#> 63 080  0.089003583  0.089003583  1.08900358                1 1   0.6666667
#> 64 082 -0.986659530 -0.986659530  0.01334047                1 1   0.6666667
#> 65 083  0.454566178  0.454566178  1.45456618                1 1   0.6666667
#> 66 087  1.642000795  1.642000795  2.64200080                1 1   0.6666667
#> 67 088 -1.319461448 -1.319461448 -0.31946145                1 0   0.3333333
#> 68 090 -0.390412961 -0.390412961  0.60958704                1 1   0.6666667
#> 69 091 -1.301609944 -1.301609944 -0.30160994                1 1   0.6666667
#> 70 092 -0.390965912 -0.390965912  0.60903409                1 1   0.6666667
#> 71 095  0.526596714  0.526596714  1.52659671                1 1   0.6666667
#> 72 096  0.767680910  0.767680910  1.76768091                1 1   0.6666667
#> 73 097 -0.134420116 -0.134420116  0.86557988                1 1   0.6666667
#> 74 099  1.878533362  1.878533362  2.87853336                1 1   0.6666667
#> 75 100  0.434136018  0.434136018  1.43413602                1 0   0.3333333
#>              Y
#> 1   0.39019196
#> 2   2.87173382
#> 3  -0.48848627
#> 4  -0.24877115
#> 5   1.70614415
#> 6  -0.12398222
#> 7   2.58964233
#> 8   0.22092243
#> 9   0.02302323
#> 10  1.52538486
#> 11 -0.19563805
#> 12  2.18903563
#> 13  0.84425167
#> 14 -0.24872746
#> 15  1.43079870
#> 16 -0.34103526
#> 17 -2.44404276
#> 18  0.68217763
#> 19  2.64969087
#> 20  0.96318632
#> 21  0.06124273
#> 22  2.46365733
#> 23 -1.45570630
#> 24  1.43756717
#> 25 -0.27714914
#> 26  1.10957480
#> 27  0.04851707
#> 28 -0.06693072
#> 29 -0.16983871
#> 30 -0.78216155
#> 31  1.10057950
#> 32 -0.59564770
#> 33  1.33733099
#> 34  1.69165977
#> 35  0.52000914
#> 36  0.26251149
#> 37  2.34080567
#> 38 -1.46624899
#> 39  2.50836693
#> 40 -0.05688255
#> 41  0.52634133
#> 42  0.76939068
#> 43  0.02402822
#> 44 -0.51956010
#> 45 -0.13211549
#> 46  2.73245662
#> 47  1.45632781
#> 48  1.13663162
#> 49 -1.52727247
#> 50 -0.92291174
#> 51 -1.16660012
#> 52 -1.22436617
#> 53 -1.36783718
#> 54  0.22275848
#> 55  0.10946167
#> 56  2.86530316
#> 57  0.90180332
#> 58 -0.39730071
#> 59 -0.20968058
#> 60 -0.34580162
#> 61  1.00574058
#> 62  1.05260193
#> 63  1.08900358
#> 64  0.01334047
#> 65  1.45456618
#> 66  2.64200080
#> 67 -1.31946145
#> 68  0.60958704
#> 69 -0.30160994
#> 70  0.60903409
#> 71  1.52659671
#> 72  1.76768091
#> 73  0.86557988
#> 74  2.87853336
#> 75  0.43413602

get_estimates(design, data = dat)

#>   estimator_label term estimate std.error statistic    p.value conf.low
#> 1       estimator    Z 0.762782  0.313876  2.430201 0.02043192 0.125271
#>   conf.high       df outcome estimand_label
#> 1  1.400293 34.53212       Y            ATE

draw_assignment(design, data = dat)

#>     ID            u        Y_Z_0       Y_Z_1 S_inclusion_prob Z Z_cond_prob
#> 1  002  0.390191956  0.390191956  1.39019196             0.75 0   0.3333333
#> 2  003  1.871733816  1.871733816  2.87173382             0.75 0   0.3333333
#> 3  004 -0.488486270 -0.488486270  0.51151373             0.75 1   0.6666667
#> 4  005 -0.248771151 -0.248771151  0.75122885             0.75 0   0.3333333
#> 5  008  1.706144150  1.706144150  2.70614415             0.75 1   0.6666667
#> 6  011 -1.123982219 -1.123982219 -0.12398222             0.75 1   0.6666667
#> 7  012  1.589642326  1.589642326  2.58964233             0.75 0   0.3333333
#> 8  013 -0.779077571 -0.779077571  0.22092243             0.75 1   0.6666667
#> 9  014 -0.976976771 -0.976976771  0.02302323             0.75 1   0.6666667
#> 10 015  0.525384860  0.525384860  1.52538486             0.75 1   0.6666667
#> 11 018 -0.195638053 -0.195638053  0.80436195             0.75 1   0.6666667
#> 12 019  1.189035632  1.189035632  2.18903563             0.75 1   0.6666667
#> 13 020 -0.155748325 -0.155748325  0.84425167             0.75 1   0.6666667
#> 14 021 -1.248727456 -1.248727456 -0.24872746             0.75 1   0.6666667
#> 15 022  0.430798701  0.430798701  1.43079870             0.75 0   0.3333333
#> 16 024 -0.341035256 -0.341035256  0.65896474             0.75 0   0.3333333
#> 17 025 -2.444042764 -2.444042764 -1.44404276             0.75 1   0.6666667
#> 18 026  0.682177629  0.682177629  1.68217763             0.75 1   0.6666667
#> 19 027  1.649690872  1.649690872  2.64969087             0.75 0   0.3333333
#> 20 028 -0.036813680 -0.036813680  0.96318632             0.75 0   0.3333333
#> 21 030  0.061242727  0.061242727  1.06124273             0.75 0   0.3333333
#> 22 031  1.463657329  1.463657329  2.46365733             0.75 1   0.6666667
#> 23 033 -2.455706303 -2.455706303 -1.45570630             0.75 1   0.6666667
#> 24 034  0.437567170  0.437567170  1.43756717             0.75 1   0.6666667
#> 25 036 -1.277149137 -1.277149137 -0.27714914             0.75 1   0.6666667
#> 26 037  0.109574799  0.109574799  1.10957480             0.75 1   0.6666667
#> 27 038  0.048517072  0.048517072  1.04851707             0.75 1   0.6666667
#> 28 039 -1.066930718 -1.066930718 -0.06693072             0.75 0   0.3333333
#> 29 040 -0.169838706 -0.169838706  0.83016129             0.75 0   0.3333333
#> 30 041 -0.782161548 -0.782161548  0.21783845             0.75 1   0.6666667
#> 31 043  0.100579496  0.100579496  1.10057950             0.75 1   0.6666667
#> 32 044 -0.595647705 -0.595647705  0.40435230             0.75 1   0.6666667
#> 33 045  0.337330988  0.337330988  1.33733099             0.75 1   0.6666667
#> 34 046  0.691659765  0.691659765  1.69165977             0.75 1   0.6666667
#> 35 047  0.520009139  0.520009139  1.52000914             0.75 1   0.6666667
#> 36 048  0.262511486  0.262511486  1.26251149             0.75 1   0.6666667
#> 37 049  1.340805674  1.340805674  2.34080567             0.75 1   0.6666667
#> 38 050 -1.466248992 -1.466248992 -0.46624899             0.75 1   0.6666667
#> 39 052  1.508366928  1.508366928  2.50836693             0.75 1   0.6666667
#> 40 053 -0.056882551 -0.056882551  0.94311745             0.75 0   0.3333333
#> 41 054 -0.473658670 -0.473658670  0.52634133             0.75 0   0.3333333
#> 42 055 -0.230609318 -0.230609318  0.76939068             0.75 0   0.3333333
#> 43 056 -0.975971783 -0.975971783  0.02402822             0.75 1   0.6666667
#> 44 057 -1.519560103 -1.519560103 -0.51956010             0.75 1   0.6666667
#> 45 058 -1.132115492 -1.132115492 -0.13211549             0.75 1   0.6666667
#> 46 059  1.732456622  1.732456622  2.73245662             0.75 1   0.6666667
#> 47 060  1.456327813  1.456327813  2.45632781             0.75 1   0.6666667
#> 48 061  0.136631624  0.136631624  1.13663162             0.75 1   0.6666667
#> 49 062 -2.527272469 -2.527272469 -1.52727247             0.75 1   0.6666667
#> 50 064 -1.922911742 -1.922911742 -0.92291174             0.75 0   0.3333333
#> 51 065 -2.166600121 -2.166600121 -1.16660012             0.75 1   0.6666667
#> 52 067 -1.224366171 -1.224366171 -0.22436617             0.75 1   0.6666667
#> 53 068 -2.367837184 -2.367837184 -1.36783718             0.75 0   0.3333333
#> 54 069  0.222758476  0.222758476  1.22275848             0.75 0   0.3333333
#> 55 071 -0.890538330 -0.890538330  0.10946167             0.75 0   0.3333333
#> 56 072  1.865303159  1.865303159  2.86530316             0.75 0   0.3333333
#> 57 073 -0.098196677 -0.098196677  0.90180332             0.75 1   0.6666667
#> 58 074 -0.397300713 -0.397300713  0.60269929             0.75 1   0.6666667
#> 59 075 -0.209680583 -0.209680583  0.79031942             0.75 1   0.6666667
#> 60 076 -0.345801617 -0.345801617  0.65419838             0.75 1   0.6666667
#> 61 077  0.005740578  0.005740578  1.00574058             0.75 1   0.6666667
#> 62 079  0.052601925  0.052601925  1.05260193             0.75 1   0.6666667
#> 63 080  0.089003583  0.089003583  1.08900358             0.75 1   0.6666667
#> 64 082 -0.986659530 -0.986659530  0.01334047             0.75 1   0.6666667
#> 65 083  0.454566178  0.454566178  1.45456618             0.75 0   0.3333333
#> 66 087  1.642000795  1.642000795  2.64200080             0.75 1   0.6666667
#> 67 088 -1.319461448 -1.319461448 -0.31946145             0.75 1   0.6666667
#> 68 090 -0.390412961 -0.390412961  0.60958704             0.75 0   0.3333333
#> 69 091 -1.301609944 -1.301609944 -0.30160994             0.75 0   0.3333333
#> 70 092 -0.390965912 -0.390965912  0.60903409             0.75 1   0.6666667
#> 71 095  0.526596714  0.526596714  1.52659671             0.75 1   0.6666667
#> 72 096  0.767680910  0.767680910  1.76768091             0.75 0   0.3333333
#> 73 097 -0.134420116 -0.134420116  0.86557988             0.75 1   0.6666667
#> 74 099  1.878533362  1.878533362  2.87853336             0.75 0   0.3333333
#> 75 100  0.434136018  0.434136018  1.43413602             0.75 0   0.3333333
#>              Y
#> 1   0.39019196
#> 2   1.87173382
#> 3   0.51151373
#> 4   0.75122885
#> 5   2.70614415
#> 6  -0.12398222
#> 7   1.58964233
#> 8   0.22092243
#> 9  -0.97697677
#> 10  1.52538486
#> 11  0.80436195
#> 12  1.18903563
#> 13  0.84425167
#> 14 -0.24872746
#> 15  1.43079870
#> 16  0.65896474
#> 17 -1.44404276
#> 18  1.68217763
#> 19  1.64969087
#> 20  0.96318632
#> 21  1.06124273
#> 22  2.46365733
#> 23 -2.45570630
#> 24  0.43756717
#> 25 -0.27714914
#> 26  1.10957480
#> 27  0.04851707
#> 28 -0.06693072
#> 29  0.83016129
#> 30  0.21783845
#> 31  1.10057950
#> 32  0.40435230
#> 33  1.33733099
#> 34  1.69165977
#> 35  1.52000914
#> 36  1.26251149
#> 37  2.34080567
#> 38 -1.46624899
#> 39  1.50836693
#> 40 -0.05688255
#> 41  0.52634133
#> 42  0.76939068
#> 43  0.02402822
#> 44 -1.51956010
#> 45 -0.13211549
#> 46  2.73245662
#> 47  2.45632781
#> 48  1.13663162
#> 49 -2.52727247
#> 50 -0.92291174
#> 51 -2.16660012
#> 52 -1.22436617
#> 53 -1.36783718
#> 54  1.22275848
#> 55  0.10946167
#> 56  1.86530316
#> 57  0.90180332
#> 58  0.60269929
#> 59  0.79031942
#> 60 -0.34580162
#> 61  1.00574058
#> 62  0.05260193
#> 63  0.08900358
#> 64  0.01334047
#> 65  0.45456618
#> 66  1.64200080
#> 67 -1.31946145
#> 68  0.60958704
#> 69 -0.30160994
#> 70  0.60903409
#> 71  1.52659671
#> 72  1.76768091
#> 73  0.86557988
#> 74  1.87853336
#> 75  0.43413602

draw_sample(design, data = dat)

#>     ID            u        Y_Z_0       Y_Z_1 S_inclusion_prob Z Z_cond_prob
#> 1  002  0.390191956  0.390191956  1.39019196                1 0   0.3333333
#> 2  003  1.871733816  1.871733816  2.87173382                1 0   0.3333333
#> 3  004 -0.488486270 -0.488486270  0.51151373                1 1   0.6666667
#> 4  005 -0.248771151 -0.248771151  0.75122885                1 1   0.6666667
#> 5  008  1.706144150  1.706144150  2.70614415                1 1   0.6666667
#> 6  011 -1.123982219 -1.123982219 -0.12398222                1 1   0.6666667
#> 7  012  1.589642326  1.589642326  2.58964233                1 0   0.3333333
#> 8  013 -0.779077571 -0.779077571  0.22092243                1 1   0.6666667
#> 9  014 -0.976976771 -0.976976771  0.02302323                1 0   0.3333333
#> 10 015  0.525384860  0.525384860  1.52538486                1 1   0.6666667
#> 11 018 -0.195638053 -0.195638053  0.80436195                1 1   0.6666667
#> 12 019  1.189035632  1.189035632  2.18903563                1 0   0.3333333
#> 13 020 -0.155748325 -0.155748325  0.84425167                1 1   0.6666667
#> 14 021 -1.248727456 -1.248727456 -0.24872746                1 1   0.6666667
#> 15 022  0.430798701  0.430798701  1.43079870                1 1   0.6666667
#> 16 024 -0.341035256 -0.341035256  0.65896474                1 1   0.6666667
#> 17 025 -2.444042764 -2.444042764 -1.44404276                1 1   0.6666667
#> 18 026  0.682177629  0.682177629  1.68217763                1 1   0.6666667
#> 19 027  1.649690872  1.649690872  2.64969087                1 0   0.3333333
#> 20 028 -0.036813680 -0.036813680  0.96318632                1 1   0.6666667
#> 21 030  0.061242727  0.061242727  1.06124273                1 1   0.6666667
#> 22 031  1.463657329  1.463657329  2.46365733                1 1   0.6666667
#> 23 033 -2.455706303 -2.455706303 -1.45570630                1 0   0.3333333
#> 24 034  0.437567170  0.437567170  1.43756717                1 0   0.3333333
#> 25 036 -1.277149137 -1.277149137 -0.27714914                1 1   0.6666667
#> 26 037  0.109574799  0.109574799  1.10957480                1 1   0.6666667
#> 27 038  0.048517072  0.048517072  1.04851707                1 0   0.3333333
#> 28 039 -1.066930718 -1.066930718 -0.06693072                1 1   0.6666667
#> 29 040 -0.169838706 -0.169838706  0.83016129                1 1   0.6666667
#> 30 041 -0.782161548 -0.782161548  0.21783845                1 1   0.6666667
#> 31 043  0.100579496  0.100579496  1.10057950                1 1   0.6666667
#> 32 044 -0.595647705 -0.595647705  0.40435230                1 1   0.6666667
#> 33 045  0.337330988  0.337330988  1.33733099                1 1   0.6666667
#> 34 046  0.691659765  0.691659765  1.69165977                1 1   0.6666667
#> 35 047  0.520009139  0.520009139  1.52000914                1 1   0.6666667
#> 36 048  0.262511486  0.262511486  1.26251149                1 1   0.6666667
#> 37 049  1.340805674  1.340805674  2.34080567                1 1   0.6666667
#> 38 050 -1.466248992 -1.466248992 -0.46624899                1 0   0.3333333
#> 39 052  1.508366928  1.508366928  2.50836693                1 0   0.3333333
#> 40 053 -0.056882551 -0.056882551  0.94311745                1 0   0.3333333
#> 41 054 -0.473658670 -0.473658670  0.52634133                1 1   0.6666667
#> 42 055 -0.230609318 -0.230609318  0.76939068                1 1   0.6666667
#> 43 056 -0.975971783 -0.975971783  0.02402822                1 1   0.6666667
#> 44 057 -1.519560103 -1.519560103 -0.51956010                1 0   0.3333333
#> 45 058 -1.132115492 -1.132115492 -0.13211549                1 1   0.6666667
#> 46 059  1.732456622  1.732456622  2.73245662                1 1   0.6666667
#> 47 060  1.456327813  1.456327813  2.45632781                1 1   0.6666667
#> 48 061  0.136631624  0.136631624  1.13663162                1 1   0.6666667
#> 49 062 -2.527272469 -2.527272469 -1.52727247                1 0   0.3333333
#> 50 064 -1.922911742 -1.922911742 -0.92291174                1 1   0.6666667
#> 51 065 -2.166600121 -2.166600121 -1.16660012                1 0   0.3333333
#> 52 067 -1.224366171 -1.224366171 -0.22436617                1 0   0.3333333
#> 53 068 -2.367837184 -2.367837184 -1.36783718                1 1   0.6666667
#> 54 069  0.222758476  0.222758476  1.22275848                1 1   0.6666667
#> 55 071 -0.890538330 -0.890538330  0.10946167                1 1   0.6666667
#> 56 072  1.865303159  1.865303159  2.86530316                1 0   0.3333333
#> 57 073 -0.098196677 -0.098196677  0.90180332                1 1   0.6666667
#> 58 074 -0.397300713 -0.397300713  0.60269929                1 1   0.6666667
#> 59 075 -0.209680583 -0.209680583  0.79031942                1 1   0.6666667
#> 60 076 -0.345801617 -0.345801617  0.65419838                1 0   0.3333333
#> 61 077  0.005740578  0.005740578  1.00574058                1 1   0.6666667
#> 62 079  0.052601925  0.052601925  1.05260193                1 0   0.3333333
#> 63 080  0.089003583  0.089003583  1.08900358                1 0   0.3333333
#> 64 082 -0.986659530 -0.986659530  0.01334047                1 1   0.6666667
#> 65 083  0.454566178  0.454566178  1.45456618                1 0   0.3333333
#> 66 087  1.642000795  1.642000795  2.64200080                1 0   0.3333333
#> 67 088 -1.319461448 -1.319461448 -0.31946145                1 0   0.3333333
#> 68 090 -0.390412961 -0.390412961  0.60958704                1 1   0.6666667
#> 69 091 -1.301609944 -1.301609944 -0.30160994                1 1   0.6666667
#> 70 092 -0.390965912 -0.390965912  0.60903409                1 1   0.6666667
#> 71 095  0.526596714  0.526596714  1.52659671                1 1   0.6666667
#> 72 096  0.767680910  0.767680910  1.76768091                1 1   0.6666667
#> 73 097 -0.134420116 -0.134420116  0.86557988                1 1   0.6666667
#> 74 099  1.878533362  1.878533362  2.87853336                1 0   0.3333333
#> 75 100  0.434136018  0.434136018  1.43413602                1 0   0.3333333
#>              Y
#> 1   0.39019196
#> 2   1.87173382
#> 3   0.51151373
#> 4   0.75122885
#> 5   2.70614415
#> 6  -0.12398222
#> 7   1.58964233
#> 8   0.22092243
#> 9  -0.97697677
#> 10  1.52538486
#> 11  0.80436195
#> 12  1.18903563
#> 13  0.84425167
#> 14 -0.24872746
#> 15  1.43079870
#> 16  0.65896474
#> 17 -1.44404276
#> 18  1.68217763
#> 19  1.64969087
#> 20  0.96318632
#> 21  1.06124273
#> 22  2.46365733
#> 23 -2.45570630
#> 24  0.43756717
#> 25 -0.27714914
#> 26  1.10957480
#> 27  0.04851707
#> 28 -0.06693072
#> 29  0.83016129
#> 30  0.21783845
#> 31  1.10057950
#> 32  0.40435230
#> 33  1.33733099
#> 34  1.69165977
#> 35  1.52000914
#> 36  1.26251149
#> 37  2.34080567
#> 38 -1.46624899
#> 39  1.50836693
#> 40 -0.05688255
#> 41  0.52634133
#> 42  0.76939068
#> 43  0.02402822
#> 44 -1.51956010
#> 45 -0.13211549
#> 46  2.73245662
#> 47  2.45632781
#> 48  1.13663162
#> 49 -2.52727247
#> 50 -0.92291174
#> 51 -2.16660012
#> 52 -1.22436617
#> 53 -1.36783718
#> 54  1.22275848
#> 55  0.10946167
#> 56  1.86530316
#> 57  0.90180332
#> 58  0.60269929
#> 59  0.79031942
#> 60 -0.34580162
#> 61  1.00574058
#> 62  0.05260193
#> 63  0.08900358
#> 64  0.01334047
#> 65  0.45456618
#> 66  1.64200080
#> 67 -1.31946145
#> 68  0.60958704
#> 69 -0.30160994
#> 70  0.60903409
#> 71  1.52659671
#> 72  1.76768091
#> 73  0.86557988
#> 74  1.87853336
#> 75  0.43413602