Get estimates, inquiries, assignment vectors, or samples from a design given data

Get estimates, inquiries, assignment vectors, or samples from a design given data

Usage

get_estimates(design, data = NULL, start = 1, end = length(design))

Arguments

design

A design object, typically created using the + operator

data

A data.frame object with sufficient information to get the data, estimates, inquiries, an assignment vector, or a sample.

start

(Defaults to 1) a scalar indicating which step in the design to begin with. By default all data steps are drawn, from step 1 to the last step of the design.

end

(Defaults to length(design)) a scalar indicating which step in the design to finish with.

Examples

design <- 
  declare_model(
    N = 100, 
    U = rnorm(N),
    potential_outcomes(Y ~ Z + U)
  ) +
  declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(S = complete_rs(N, n = 75)) +
  declare_assignment(Z = complete_ra(N, m = 50)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z)) +
  declare_estimator(Y ~ Z, inquiry = "ATE")

dat <- draw_data(design)

draw_data(design, data = dat, start = 2)
#>     ID           U       Y_Z_0      Y_Z_1 S Z           Y
#> 1  001 -0.87882941 -0.87882941  0.1211706 1 0 -0.87882941
#> 2  002  1.72273527  1.72273527  2.7227353 1 1  2.72273527
#> 3  005  1.31454786  1.31454786  2.3145479 1 0  1.31454786
#> 4  007 -0.27090867 -0.27090867  0.7290913 1 1  0.72909133
#> 5  009 -0.15784375 -0.15784375  0.8421562 1 1  0.84215625
#> 6  010 -0.14136779 -0.14136779  0.8586322 1 1  0.85863221
#> 7  011  0.72143442  0.72143442  1.7214344 1 1  1.72143442
#> 8  012 -0.76129734 -0.76129734  0.2387027 1 1  0.23870266
#> 9  013  0.54495115  0.54495115  1.5449511 1 1  1.54495115
#> 10 014 -0.42257532 -0.42257532  0.5774247 1 1  0.57742468
#> 11 015  0.77341969  0.77341969  1.7734197 1 1  1.77341969
#> 12 016 -1.42948589 -1.42948589 -0.4294859 1 1 -0.42948589
#> 13 018  0.06756114  0.06756114  1.0675611 1 1  1.06756114
#> 14 019 -1.12128149 -1.12128149 -0.1212815 1 0 -1.12128149
#> 15 020 -1.40300134 -1.40300134 -0.4030013 1 1 -0.40300134
#> 16 022 -2.03362499 -2.03362499 -1.0336250 1 0 -2.03362499
#> 17 023  0.39062693  0.39062693  1.3906269 1 1  1.39062693
#> 18 026 -1.23828408 -1.23828408 -0.2382841 1 0 -1.23828408
#> 19 027  1.48692259  1.48692259  2.4869226 1 1  2.48692259
#> 20 028 -2.21942940 -2.21942940 -1.2194294 1 1 -1.21942940
#> 21 029 -0.14122478 -0.14122478  0.8587752 1 1  0.85877522
#> 22 030  0.04489543  0.04489543  1.0448954 1 0  0.04489543
#> 23 032 -0.22008234 -0.22008234  0.7799177 1 0 -0.22008234
#> 24 034 -0.32520499 -0.32520499  0.6747950 1 1  0.67479501
#> 25 035 -1.89831056 -1.89831056 -0.8983106 1 1 -0.89831056
#> 26 037 -2.13205588 -2.13205588 -1.1320559 1 1 -1.13205588
#> 27 038 -0.22044023 -0.22044023  0.7795598 1 0 -0.22044023
#> 28 039 -0.21398788 -0.21398788  0.7860121 1 1  0.78601212
#> 29 040  1.10023709  1.10023709  2.1002371 1 1  2.10023709
#> 30 043  0.58184112  0.58184112  1.5818411 1 1  1.58184112
#> 31 045  0.32048821  0.32048821  1.3204882 1 1  1.32048821
#> 32 048 -1.36766032 -1.36766032 -0.3676603 1 1 -0.36766032
#> 33 049 -0.89620433 -0.89620433  0.1037957 1 1  0.10379567
#> 34 050 -0.49659735 -0.49659735  0.5034026 1 1  0.50340265
#> 35 051  1.71128742  1.71128742  2.7112874 1 1  2.71128742
#> 36 052  0.20509619  0.20509619  1.2050962 1 0  0.20509619
#> 37 053  2.24635016  2.24635016  3.2463502 1 1  3.24635016
#> 38 055 -0.37015267 -0.37015267  0.6298473 1 1  0.62984733
#> 39 057 -0.48434283 -0.48434283  0.5156572 1 0 -0.48434283
#> 40 058  0.19582465  0.19582465  1.1958247 1 1  1.19582465
#> 41 059  1.06382219  1.06382219  2.0638222 1 0  1.06382219
#> 42 061  1.27432412  1.27432412  2.2743241 1 0  1.27432412
#> 43 062  0.11220531  0.11220531  1.1122053 1 1  1.11220531
#> 44 063 -0.59326089 -0.59326089  0.4067391 1 0 -0.59326089
#> 45 064 -1.52633841 -1.52633841 -0.5263384 1 1 -0.52633841
#> 46 066  1.49182165  1.49182165  2.4918216 1 1  2.49182165
#> 47 067 -0.87372593 -0.87372593  0.1262741 1 1  0.12627407
#> 48 068 -0.92788770 -0.92788770  0.0721123 1 0 -0.92788770
#> 49 069  1.61444233  1.61444233  2.6144423 1 1  2.61444233
#> 50 070  0.35517142  0.35517142  1.3551714 1 0  0.35517142
#> 51 071  0.19758280  0.19758280  1.1975828 1 1  1.19758280
#> 52 073 -0.22804249 -0.22804249  0.7719575 1 1  0.77195751
#> 53 074  0.90613469  0.90613469  1.9061347 1 1  1.90613469
#> 54 075 -0.85552556 -0.85552556  0.1444744 1 0 -0.85552556
#> 55 076  0.21110636  0.21110636  1.2111064 1 0  0.21110636
#> 56 077 -0.68504003 -0.68504003  0.3149600 1 1  0.31495997
#> 57 078 -0.26997051 -0.26997051  0.7300295 1 0 -0.26997051
#> 58 079  1.92496731  1.92496731  2.9249673 1 1  2.92496731
#> 59 080 -1.22239227 -1.22239227 -0.2223923 1 1 -0.22239227
#> 60 081 -1.96710660 -1.96710660 -0.9671066 1 0 -1.96710660
#> 61 082 -0.39923655 -0.39923655  0.6007634 1 1  0.60076345
#> 62 083  0.40808053  0.40808053  1.4080805 1 0  0.40808053
#> 63 084  0.77153549  0.77153549  1.7715355 1 1  1.77153549
#> 64 086 -0.73482926 -0.73482926  0.2651707 1 0 -0.73482926
#> 65 087  1.56635794  1.56635794  2.5663579 1 1  2.56635794
#> 66 088 -2.07882778 -2.07882778 -1.0788278 1 0 -2.07882778
#> 67 090  1.15203958  1.15203958  2.1520396 1 1  2.15203958
#> 68 091 -0.08650010 -0.08650010  0.9134999 1 1  0.91349990
#> 69 093  1.02370257  1.02370257  2.0237026 1 0  1.02370257
#> 70 094 -0.10639568 -0.10639568  0.8936043 1 0 -0.10639568
#> 71 095 -0.24421579 -0.24421579  0.7557842 1 1  0.75578421
#> 72 096 -0.63353293 -0.63353293  0.3664671 1 1  0.36646707
#> 73 098 -0.27453732 -0.27453732  0.7254627 1 1  0.72546268
#> 74 099  0.53869217  0.53869217  1.5386922 1 0  0.53869217
#> 75 100  2.12792630  2.12792630  3.1279263 1 1  3.12792630

get_estimates(design, data = dat)
#>   estimator term  estimate std.error statistic    p.value  conf.low conf.high
#> 1 estimator    Z 0.7093093 0.2700566  2.626521 0.01050509 0.1710872  1.247531
#>   df outcome inquiry
#> 1 73       Y     ATE