Modify a design after the fact

Insert, delete and replace steps in an (already declared) design object.

insert_step(design, new_step, before, after)

delete_step(design, step)

replace_step(design, step, new_step)

Arguments

design
A design object, usually created using the + operator, expand_design, or the design library.

new_step
The new step; Either a function or a partial call.

before
The step before which to add steps.

after
The step after which to add steps.

step
The quoted label of the step to be deleted or replaced.

Value

A new design object.

Details

See modify_design for details.

Examples

 my_population <- declare_population(N = 100, noise = rnorm(N), label = "my_pop")

 my_potential_outcomes <-
   declare_potential_outcomes(Y_Z_0 = noise,
                              Y_Z_1 = noise + rnorm(N, mean = 2, sd = 2))

 my_assignment <- declare_assignment(m = 50)
 my_assignment_2 <- declare_assignment(m = 25)

 design <- my_population + my_potential_outcomes + my_assignment

 design

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, noise = rnorm(N), label = "my_pop") 
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: noise 
#>  min median  mean  max   sd N_missing N_unique
#>   -3  -0.12 -0.12 2.35 1.09         0      100
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y_Z_0 = noise, Y_Z_1 = noise + rnorm(N,     mean = 2, sd = 2)) 
#> 
#> Added variable: Y_Z_0 
#>  min median  mean  max   sd N_missing N_unique
#>   -3  -0.12 -0.12 2.35 1.09         0      100
#> 
#> Added variable: Y_Z_1 
#>    min median mean  max   sd N_missing N_unique
#>  -4.34   2.12 2.05 8.35 2.35         0      100
#> 
#> Step 3 (assignment): declare_assignment(m = 50) --------------------------------
#> 
#> Added variable: Z 
#>     0    1
#>    50   50
#>  0.50 0.50
#> 
#> Added variable: Z_cond_prob 
#>   0.5
#>   100
#>  1.00
#> 

 insert_step(design, declare_step(dplyr::mutate, income = noise^2), after = my_assignment)

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, noise = rnorm(N), label = "my_pop") 
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: noise 
#>    min median mean  max   sd N_missing N_unique
#>  -2.39   0.16 0.22 2.99 0.96         0      100
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y_Z_0 = noise, Y_Z_1 = noise + rnorm(N,     mean = 2, sd = 2)) 
#> 
#> Added variable: Y_Z_0 
#>    min median mean  max   sd N_missing N_unique
#>  -2.39   0.16 0.22 2.99 0.96         0      100
#> 
#> Added variable: Y_Z_1 
#>    min median mean  max   sd N_missing N_unique
#>  -3.79   2.15 2.41 7.75 2.26         0      100
#> 
#> Step 3 (assignment): declare_assignment(m = 50) --------------------------------
#> 
#> Added variable: Z 
#>     0    1
#>    50   50
#>  0.50 0.50
#> 
#> Added variable: Z_cond_prob 
#>   0.5
#>   100
#>  1.00
#> 
#> Step 4 (custom): declare_step(dplyr::mutate, income = noise^2) -----------------
#> 
#> Added variable: income 
#>  min median mean  max   sd N_missing N_unique
#>    0   0.54 0.96 8.93 1.46         0      100
#> 

 insert_step(design, declare_step(dplyr::mutate, income = noise^2), before = my_assignment)

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, noise = rnorm(N), label = "my_pop") 
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: noise 
#>   min median  mean max  sd N_missing N_unique
#>  -2.7  -0.14 -0.06 2.4 1.1         0      100
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y_Z_0 = noise, Y_Z_1 = noise + rnorm(N,     mean = 2, sd = 2)) 
#> 
#> Added variable: Y_Z_0 
#>   min median  mean max  sd N_missing N_unique
#>  -2.7  -0.14 -0.06 2.4 1.1         0      100
#> 
#> Added variable: Y_Z_1 
#>   min median mean  max   sd N_missing N_unique
#>  -3.6   1.83 1.86 7.62 2.24         0      100
#> 
#> Step 3 (custom): declare_step(dplyr::mutate, income = noise^2) -----------------
#> 
#> Added variable: income 
#>  min median mean  max   sd N_missing N_unique
#>    0   0.67 1.19 7.29 1.44         0      100
#> 
#> Step 4 (assignment): declare_assignment(m = 50) --------------------------------
#> 
#> Added variable: Z 
#>     0    1
#>    50   50
#>  0.50 0.50
#> 
#> Added variable: Z_cond_prob 
#>   0.5
#>   100
#>  1.00
#> 

 # If you are using a design created by a designer, for example from
 #   the DesignLibrary package, you will not have access to the step
 #   objects. Instead, you can always use the label of the step.
 
 # get the labels for the steps
 names(design)

#> [1] "my_pop"                "my_potential_outcomes" "my_assignment"        

 
 insert_step(design, declare_sampling(n = 50), after = "my_pop")

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, noise = rnorm(N), label = "my_pop") 
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: noise 
#>   min median  mean max   sd N_missing N_unique
#>  -2.1  -0.04 -0.04 2.1 0.99         0      100
#> 
#> Step 2 (sampling): declare_sampling(n = 50) ------------------------------------
#> 
#> N = 50 (50 subtracted) 
#> 
#> Added variable: S_inclusion_prob 
#>   0.5
#>    50
#>  1.00
#> 
#> Altered variable: ID 
#>   Before: 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#>   After:
#>  N_missing N_unique     class
#>          0       50 character
#> 
#> Altered variable: noise 
#>   Before: 
#>   min median  mean max   sd N_missing N_unique
#>  -2.1  -0.04 -0.04 2.1 0.99         0      100
#> 
#>   After:
#>   min median  mean  max  sd N_missing N_unique
#>  -2.1  -0.13 -0.12 1.74 0.9         0       50
#> 
#> Step 3 (potential outcomes): declare_potential_outcomes(Y_Z_0 = noise, Y_Z_1 = noise + rnorm(N,     mean = 2, sd = 2)) 
#> 
#> Added variable: Y_Z_0 
#>   min median  mean  max  sd N_missing N_unique
#>  -2.1  -0.13 -0.12 1.74 0.9         0       50
#> 
#> Added variable: Y_Z_1 
#>  min median mean  max sd N_missing N_unique
#>   -3   2.45 2.42 7.63  2         0       50
#> 
#> Step 4 (assignment): declare_assignment(m = 50) --------------------------------
#> 
#> Added variable: Z 
#>     1
#>    50
#>  1.00
#> 
#> Added variable: Z_cond_prob 
#>     1
#>    50
#>  1.00
#> 

 delete_step(design, my_assignment)

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, noise = rnorm(N), label = "my_pop") 
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: noise 
#>    min median  mean  max   sd N_missing N_unique
#>  -2.96   0.04 -0.11 3.79 1.06         0      100
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y_Z_0 = noise, Y_Z_1 = noise + rnorm(N,     mean = 2, sd = 2)) 
#> 
#> Added variable: Y_Z_0 
#>    min median  mean  max   sd N_missing N_unique
#>  -2.96   0.04 -0.11 3.79 1.06         0      100
#> 
#> Added variable: Y_Z_1 
#>    min median mean  max  sd N_missing N_unique
#>  -4.42   1.91 1.89 7.13 2.2         0      100
#> 

 replace_step(design, my_assignment, declare_step(dplyr::mutate, words = "income"))

#> 
#> Design Summary
#> 
#> Step 1 (population): declare_population(N = 100, noise = rnorm(N), label = "my_pop") 
#> 
#> N = 100 
#> 
#> Added variable: ID 
#>  N_missing N_unique     class
#>          0      100 character
#> 
#> Added variable: noise 
#>    min median mean  max   sd N_missing N_unique
#>  -2.18   0.07    0 2.43 1.03         0      100
#> 
#> Step 2 (potential outcomes): declare_potential_outcomes(Y_Z_0 = noise, Y_Z_1 = noise + rnorm(N,     mean = 2, sd = 2)) 
#> 
#> Added variable: Y_Z_0 
#>    min median mean  max   sd N_missing N_unique
#>  -2.18   0.07    0 2.43 1.03         0      100
#> 
#> Added variable: Y_Z_1 
#>   min median mean  max   sd N_missing N_unique
#>  -4.5   1.87 1.89 7.02 2.37         0      100
#> 
#> Step 3 (custom): declare_step(dplyr::mutate, words = "income") -----------------
#> 
#> Added variable: words 
#>  income
#>     100
#>    1.00
#>