Modify a design after the fact

Insert, delete and replace steps in an (already declared) design object.

Usage

insert_step(design, new_step, before, after)

delete_step(design, step)

replace_step(design, step, new_step)

Arguments

design

A design object, usually created using the + operator, expand_design, or the design library.

new_step

The new step; Either a function or a partial call.

before

The step before which to add steps.

after

The step after which to add steps.

step

The quoted label of the step to be deleted or replaced.

Value

A new design object.

Details

See modify_design for details.

Examples

 my_model <- 
   declare_model(
     N = 100, 
     U = rnorm(N),
     Y_Z_0 = U,
     Y_Z_1 = U + rnorm(N, mean = 2, sd = 2)
   )

 my_assignment <- declare_assignment(Z = complete_ra(N, m = 50))
 my_assignment_2 <- declare_assignment(Z = complete_ra(N, m = 25))

 design <- my_model + my_assignment

 draw_data(design)
#>      ID            U        Y_Z_0       Y_Z_1 Z
#> 1   001 -0.379568391 -0.379568391 -1.70149004 1
#> 2   002  0.406766274  0.406766274  5.96250997 1
#> 3   003  0.635518399  0.635518399  2.69003735 0
#> 4   004  1.377646989  1.377646989  2.47325042 0
#> 5   005 -0.949309023 -0.949309023 -2.06657632 1
#> 6   006  0.001483096  0.001483096 -0.61605409 1
#> 7   007  1.265388505  1.265388505  1.61778514 1
#> 8   008  1.383216470  1.383216470  4.91853090 0
#> 9   009  2.270837220  2.270837220  4.05471593 1
#> 10  010 -1.539723449 -1.539723449  2.71829460 1
#> 11  011 -0.220922290 -0.220922290  0.23262737 0
#> 12  012 -0.957452932 -0.957452932 -2.03697675 0
#> 13  013  0.010623493  0.010623493  2.98352387 0
#> 14  014 -0.862868768 -0.862868768  2.57242941 0
#> 15  015 -0.364097652 -0.364097652  2.36085372 1
#> 16  016  0.844807910  0.844807910  1.15004076 1
#> 17  017  0.385405052  0.385405052  4.35272856 1
#> 18  018 -1.274743074 -1.274743074  1.02557779 1
#> 19  019 -0.971967381 -0.971967381  1.07928376 0
#> 20  020  0.498837744  0.498837744  2.93234207 1
#> 21  021  1.314059303  1.314059303  2.09443182 1
#> 22  022 -1.985083683 -1.985083683  2.18721852 0
#> 23  023 -0.405596227 -0.405596227  4.30888861 1
#> 24  024 -1.651672781 -1.651672781  1.89341228 1
#> 25  025 -0.552992970 -0.552992970  3.10463317 1
#> 26  026 -0.455715564 -0.455715564 -0.06945505 0
#> 27  027 -0.751008900 -0.751008900  3.53098635 0
#> 28  028 -0.883535203 -0.883535203  1.11116891 0
#> 29  029 -1.880675462 -1.880675462 -0.11507762 1
#> 30  030  0.249372415  0.249372415 -1.14498039 1
#> 31  031 -0.505183721 -0.505183721 -1.21063836 0
#> 32  032 -1.617724125 -1.617724125 -3.48434726 1
#> 33  033 -0.257899954 -0.257899954  4.26503437 1
#> 34  034 -1.242673098 -1.242673098  1.46284441 0
#> 35  035 -0.187390467 -0.187390467  3.62033945 0
#> 36  036 -0.173101767 -0.173101767  1.39494047 1
#> 37  037 -0.825864827 -0.825864827  0.34170978 0
#> 38  038 -0.901651168 -0.901651168 -1.83317207 1
#> 39  039  0.911032290  0.911032290  5.67191337 0
#> 40  040  0.558938504  0.558938504  5.65790034 0
#> 41  041 -1.324874913 -1.324874913  0.03558427 1
#> 42  042  0.038117154  0.038117154  3.13768009 1
#> 43  043  0.129803737  0.129803737  3.50588199 0
#> 44  044 -0.990859811 -0.990859811  2.29695404 0
#> 45  045 -1.203503433 -1.203503433 -0.86321600 0
#> 46  046 -0.070436653 -0.070436653  4.23806751 0
#> 47  047  1.625755112  1.625755112  6.00928857 1
#> 48  048 -1.197501032 -1.197501032  0.81285128 1
#> 49  049  0.302689294  0.302689294  1.01710851 0
#> 50  050 -0.068992892 -0.068992892  1.85297894 0
#> 51  051 -0.447072183 -0.447072183  1.60013429 0
#> 52  052 -0.746975316 -0.746975316  2.73832355 1
#> 53  053 -0.740784370 -0.740784370  3.79101877 0
#> 54  054  0.899791381  0.899791381  0.59323229 1
#> 55  055 -0.951575040 -0.951575040 -0.23363669 1
#> 56  056 -0.277051477 -0.277051477  2.60918405 1
#> 57  057 -0.666980889 -0.666980889 -0.78139811 1
#> 58  058  1.133632398  1.133632398  4.61297494 1
#> 59  059  1.780580231  1.780580231  5.41160672 1
#> 60  060  0.482201587  0.482201587  3.20806312 1
#> 61  061 -0.450935005 -0.450935005  1.03681162 1
#> 62  062  0.138331359  0.138331359 -0.45284225 0
#> 63  063 -1.730432387 -1.730432387 -1.56015117 1
#> 64  064 -1.080988060 -1.080988060 -0.22276400 0
#> 65  065 -0.076419039 -0.076419039  2.97439738 0
#> 66  066  1.004580816  1.004580816  3.40287144 0
#> 67  067  2.163462312  2.163462312  6.26030023 0
#> 68  068 -0.052100067 -0.052100067  3.48578670 0
#> 69  069  0.986986528  0.986986528  1.80129012 1
#> 70  070 -0.333960016 -0.333960016  1.71646846 0
#> 71  071  0.808945604  0.808945604  3.50859261 0
#> 72  072 -0.015960774 -0.015960774  0.31333519 0
#> 73  073  0.772907916  0.772907916  3.28382928 0
#> 74  074 -1.263492721 -1.263492721 -0.70146670 0
#> 75  075 -0.401914035 -0.401914035  1.38541873 0
#> 76  076 -0.984409503 -0.984409503 -0.69454810 0
#> 77  077 -1.815560295 -1.815560295  1.32888671 0
#> 78  078 -0.391654509 -0.391654509  0.61340096 1
#> 79  079  0.133255792  0.133255792  3.13539426 0
#> 80  080  0.271107978  0.271107978  3.32665186 0
#> 81  081 -0.535931502 -0.535931502 -0.46565717 0
#> 82  082  0.920523180  0.920523180  3.08518396 0
#> 83  083  0.394260484  0.394260484  3.66659142 0
#> 84  084 -0.583224945 -0.583224945  0.57373352 1
#> 85  085  1.490864226  1.490864226  2.92841786 1
#> 86  086 -1.055366740 -1.055366740  0.08387676 0
#> 87  087  0.539046783  0.539046783  3.24593900 1
#> 88  088 -0.561928940 -0.561928940  0.44428106 1
#> 89  089  1.148482036  1.148482036  4.61627487 1
#> 90  090 -0.157653298 -0.157653298  1.78539796 0
#> 91  091 -0.210733756 -0.210733756 -0.67198217 1
#> 92  092 -0.530020446 -0.530020446  1.58691202 0
#> 93  093 -0.079104852 -0.079104852  1.95436205 1
#> 94  094  0.379870331  0.379870331  2.94961565 1
#> 95  095  0.044950545  0.044950545  4.96502788 1
#> 96  096 -0.642643568 -0.642643568 -1.13613456 1
#> 97  097 -1.584418904 -1.584418904 -0.91018170 0
#> 98  098 -0.611172816 -0.611172816 -2.89567655 0
#> 99  099  0.494782641  0.494782641  2.07728221 1
#> 100 100  0.089037186  0.089037186  2.50468575 1
 
 design_modified <- replace_step(design, 2, my_assignment_2)
 
 draw_data(design)
#>      ID            U        Y_Z_0         Y_Z_1 Z
#> 1   001  0.604254851  0.604254851  2.0797562494 0
#> 2   002  2.186169974  2.186169974  3.7961578426 0
#> 3   003 -0.280634005 -0.280634005  4.2494115275 0
#> 4   004 -1.592532323 -1.592532323 -0.7648860579 0
#> 5   005 -0.991304347 -0.991304347  2.1961651015 1
#> 6   006  1.065547797  1.065547797  5.5925503517 1
#> 7   007 -1.761821495 -1.761821495  0.1122381694 1
#> 8   008  2.450561987  2.450561987  0.1353294367 0
#> 9   009  0.520231576  0.520231576  2.8693187708 1
#> 10  010 -1.012289827 -1.012289827  0.7907472131 0
#> 11  011 -1.589376313 -1.589376313  0.6572767509 1
#> 12  012  1.073241675  1.073241675  1.3434772963 1
#> 13  013 -0.095206736 -0.095206736  2.3403440204 1
#> 14  014  0.059224263  0.059224263  4.9976925827 0
#> 15  015  0.145738212  0.145738212  2.9510191822 0
#> 16  016 -1.752043124 -1.752043124  0.9196061496 1
#> 17  017 -0.175707052 -0.175707052 -0.8141775308 0
#> 18  018  0.614079729  0.614079729  2.5271322895 0
#> 19  019 -0.888314380 -0.888314380  1.0258705263 1
#> 20  020  0.234803729  0.234803729  4.9226411760 0
#> 21  021  0.295913042  0.295913042  6.6246515679 0
#> 22  022 -1.283897184 -1.283897184  0.6861270610 1
#> 23  023 -0.149729141 -0.149729141  3.2385935386 0
#> 24  024 -0.231172969 -0.231172969  0.0375634022 1
#> 25  025  0.810851630  0.810851630  3.9952367963 1
#> 26  026 -1.126914423 -1.126914423 -1.3256846149 0
#> 27  027 -0.281475540 -0.281475540 -0.2783860213 1
#> 28  028  1.340672590  1.340672590  4.2497319065 0
#> 29  029 -0.968023504 -0.968023504  1.5567506139 0
#> 30  030 -1.266450217 -1.266450217  1.9530276244 1
#> 31  031 -0.204917117 -0.204917117 -1.8550990083 1
#> 32  032  1.275089116  1.275089116  7.7721103059 1
#> 33  033 -2.498898534 -2.498898534  0.3350015332 1
#> 34  034 -1.601436353 -1.601436353 -0.3225406657 0
#> 35  035 -1.037751743 -1.037751743  4.1214166070 1
#> 36  036  0.773967265  0.773967265  3.6484910104 1
#> 37  037 -0.743590056 -0.743590056  3.1660458029 0
#> 38  038  1.879258298  1.879258298  1.1375769751 1
#> 39  039  0.134507883  0.134507883  3.1741472115 0
#> 40  040 -0.978792244 -0.978792244  1.8818380454 0
#> 41  041  1.207612334  1.207612334  1.1438834091 0
#> 42  042 -2.356664626 -2.356664626 -1.3217762131 0
#> 43  043 -0.433331335 -0.433331335  0.1058622482 0
#> 44  044  0.653887475  0.653887475  2.3124323566 1
#> 45  045  1.004667464  1.004667464  0.2213698529 0
#> 46  046  1.701737014  1.701737014  4.9176691607 1
#> 47  047  0.183059910  0.183059910  4.5039645749 0
#> 48  048 -1.021550436 -1.021550436 -1.9411024457 1
#> 49  049 -0.822600254 -0.822600254 -1.0448871927 1
#> 50  050  1.345581610  1.345581610  5.6167068293 1
#> 51  051  0.968082881  0.968082881  3.4788264709 0
#> 52  052 -0.169753818 -0.169753818  1.3123966367 0
#> 53  053 -0.122878899 -0.122878899  3.2363375603 0
#> 54  054 -0.814080740 -0.814080740  3.7635796241 1
#> 55  055 -1.218250639 -1.218250639 -0.1349260126 1
#> 56  056  0.043704504  0.043704504  1.8538447110 1
#> 57  057  2.665426149  2.665426149  5.2822883890 1
#> 58  058  1.558047988  1.558047988  1.1589253220 0
#> 59  059 -0.507692487 -0.507692487  4.8914332644 0
#> 60  060 -0.812202934 -0.812202934  3.2726197305 1
#> 61  061 -0.539838668 -0.539838668 -2.4768904326 0
#> 62  062  0.854263061  0.854263061  1.5575498829 0
#> 63  063  0.401225391  0.401225391  3.9266756677 0
#> 64  064 -1.627239077 -1.627239077 -0.0001080399 1
#> 65  065 -0.848354922 -0.848354922 -0.4824135556 0
#> 66  066 -1.285403754 -1.285403754 -1.8028469883 0
#> 67  067 -1.942769824 -1.942769824  1.9273043574 1
#> 68  068 -0.047403856 -0.047403856  2.7865358104 0
#> 69  069  0.695899366  0.695899366  3.4692257230 1
#> 70  070  0.460945979  0.460945979  3.7166639024 0
#> 71  071 -0.246305337 -0.246305337  0.0067619114 0
#> 72  072 -0.915210207 -0.915210207 -0.5964750234 1
#> 73  073 -0.870312403 -0.870312403 -0.7766331644 1
#> 74  074  0.512753704  0.512753704 -0.3097186349 0
#> 75  075 -0.352989180 -0.352989180  1.4844562091 1
#> 76  076  0.155078084  0.155078084  0.0335268181 1
#> 77  077  0.485581445  0.485581445  2.6720252398 0
#> 78  078 -0.959558985 -0.959558985 -0.0679313890 0
#> 79  079 -0.160213172 -0.160213172  0.4689212367 0
#> 80  080  0.573866300  0.573866300  4.5238530719 1
#> 81  081 -0.187161994 -0.187161994  5.1902206035 0
#> 82  082  0.298418189  0.298418189  3.0025598949 1
#> 83  083  1.013033340  1.013033340  1.4069530228 1
#> 84  084 -1.871664210 -1.871664210  0.8337617052 1
#> 85  085 -1.813539983 -1.813539983  0.0490153913 1
#> 86  086 -0.002814439 -0.002814439  4.6714392315 1
#> 87  087  0.528232814  0.528232814  0.3902228089 1
#> 88  088 -0.779063491 -0.779063491 -0.3919188807 0
#> 89  089 -0.388825796 -0.388825796  4.3590022583 0
#> 90  090  0.020194567  0.020194567  3.0895444560 0
#> 91  091 -0.609701476 -0.609701476  3.0431395385 1
#> 92  092  0.612122364  0.612122364  3.8935865149 1
#> 93  093  1.370534207  1.370534207  6.8623756211 0
#> 94  094 -0.287433563 -0.287433563  3.2782415202 0
#> 95  095 -0.623593232 -0.623593232 -2.3490223983 0
#> 96  096  0.923566087  0.923566087  4.8468192254 1
#> 97  097 -0.470750692 -0.470750692  0.4670580655 0
#> 98  098 -2.769636432 -2.769636432  0.4709415882 1
#> 99  099 -1.618163428 -1.618163428  1.9022857526 1
#> 100 100  0.193219450  0.193219450 -0.7494953325 1
 
 if (FALSE) {
 
 design <- 
   declare_model(
     N = 100,
     U = rnorm(N),
     potential_outcomes(Y ~ 0.20 * Z + U)
   ) + 
     declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) + 
     declare_assignment(Z = complete_ra(N, m = N/2)) + 
     declare_measurement(Y = reveal_outcomes(Y ~ Z)) + 
     declare_estimator(Y ~ Z, inquiry = "ATE")
 
 insert_step(design, declare_sampling(S = complete_rs(N, n = 50)),
             after = 1)

 # If you are using a design created by a designer, for example from
 #   the DesignLibrary package, you will not have access to the step
 #   objects. Instead, you can always use the label of the step.
 
 design <- DesignLibrary::two_arm_designer()
 
 # get the labels for the steps
 names(design)
 
 insert_step(design, 
   declare_sampling(S = complete_rs(N, n = 50)),
   after = "potential_outcomes")
   
 }

design <- 
  declare_model(
    N = 100,
    U = rnorm(N),
    potential_outcomes(Y ~ 0.20 * Z + U)
  ) + 
    declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) + 
    declare_assignment(Z = complete_ra(N, m = N/2)) + 
    declare_measurement(Y = reveal_outcomes(Y ~ Z)) + 
    declare_estimator(Y ~ Z, inquiry = "ATE")
delete_step(design, step = 5)
#> 
#> Research design declaration summary
#> 
#> Step 1 (model): declare_model(N = 100, U = rnorm(N), potential_outcomes(Y ~ 0.2 * Z + U)) 
#> 
#> Step 2 (inquiry): declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) -------------------
#> 
#> Step 3 (assignment): declare_assignment(Z = complete_ra(N, m = N/2)) -----------
#> 
#> Step 4 (measurement): declare_measurement(Y = reveal_outcomes(Y ~ Z)) ----------
#> 
#> Run of the design:
#> 
#>  inquiry estimand
#>      ATE      0.2
#> 

design <- 
  declare_model(
    N = 100,
    U = rnorm(N),
    potential_outcomes(Y ~ 0.20 * Z + U)
  ) + 
    declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) + 
    declare_assignment(Z = complete_ra(N, m = N/2)) + 
    declare_measurement(Y = reveal_outcomes(Y ~ Z)) + 
    declare_estimator(Y ~ Z, inquiry = "ATE")

replace_step(
  design, 
  step = 3, 
  new_step = declare_assignment(Z = simple_ra(N, prob = 0.5)))
#> 
#> Research design declaration summary
#> 
#> Step 1 (model): declare_model(N = 100, U = rnorm(N), potential_outcomes(Y ~ 0.2 * Z + U)) 
#> 
#> Step 2 (inquiry): declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) -------------------
#> 
#> Step 3 (assignment): declare_assignment(Z = simple_ra(N, prob = 0.5)) ----------
#> 
#> Step 4 (measurement): declare_measurement(Y = reveal_outcomes(Y ~ Z)) ----------
#> 
#> Step 5 (estimator): declare_estimator(Y ~ Z, inquiry = "ATE") ------------------
#> 
#> Run of the design:
#> 
#>  inquiry estimand estimator term estimate std.error statistic p.value conf.low
#>      ATE      0.2 estimator    Z    0.154     0.179     0.862   0.391   -0.201
#>  conf.high df outcome
#>      0.509 98       Y
#>