Reveal Observed Outcomes

Reveal Observed Outcomes

reveal_outcomes(data = NULL, outcome_variable_names = Y,
  assignment_variable_names = Z, attrition_variable_name = NULL,
  outcome_function = NULL)

Arguments

data
A data.frame containing columns of potential outcomes and an assignment variable

outcome_variable_names
The outcome prefix(es) of the potential outcomes

assignment_variable_names
The bare (unquote) name(s) of the assignment variable

attrition_variable_name
The bare (unquote) name of the attrition variable

outcome_function
If specified, reveal_outcomes draws outcomes using outcome_function rather than the switching equation.

Details

Typically, a design includes a potential outcomes declaration and an assignment declaration. Reveal outcomes uses the random assignment to pluck out the correct potential outcomes. This is analogous to the "switching equation" (Gerber and Green 2012, Chapter 2).

Examples

my_population <- declare_population(N = 100, noise = rnorm(N))

my_potential_outcomes <- declare_potential_outcomes(
  Y_Z_0 = noise, Y_Z_1 = noise +
  rnorm(N, mean = 2, sd = 2))

my_assignment <- declare_assignment(m = 50)

design <- declare_design(my_population,
                         my_potential_outcomes,
                         my_assignment,
                         reveal_outcomes)

design

#> Error in if (!is.null(call) & !is.null(type) & type != "declare_step") {    args <- lang_args(call)    has_formula <- sapply(args, is_formula)    formulae <- args[has_formula]    if (length(formulae) == 1) {        return(formulae[[1]])    }    else {        return(NULL)    }} else {    return(NULL)}: argument is of length zero