Set the diagnosands for a design

A researcher often has a set of diagnosands in mind to appropriately assess the quality of a design. set_diagnosands sets the default diagnosands for a design, so that later readers can assess the design on the same terms as the original author. Readers can also use diagnose_design to diagnose the design using any other set of diagnosands.

set_diagnosands(x, diagnosands = default_diagnosands)

Arguments

x
A design typically created using the + operator, or a simulations data.frame created by simulate_design.

diagnosands
A set of diagnosands created by declare_diagnosands

Value

a design object with a diagnosand attribute

Examples

design <-
declare_population(data = sleep) +
  declare_estimand(mean_outcome = mean(extra)) +
  declare_sampling(n = 10) +
  declare_estimator(extra ~ 1, estimand = "mean_outcome",
     term = '(Intercept)', model = lm_robust)

diagnosands <- declare_diagnosands(
  median_bias = median(estimate - estimand))

design <- set_diagnosands(design, diagnosands)

if (FALSE) {
diagnose_design(design)

simulations_df <- simulate_design(design)

simulations_df <- set_diagnosands(simulations_df, design)

diagnose_design(simulations_df)

}