Tidy diagnosis

Tidy diagnosis

Usage

# S3 method for diagnosis
tidy(x, conf.int = TRUE, conf.level = 0.95, ...)

Arguments

x

A diagnosis object generated by diagnose_design.

conf.int

Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to ‘TRUE’.

conf.level

The confidence level to use for the confidence interval if ‘conf.int = TRUE’. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

...

extra arguments (not used)

Value

A data.frame with columns for diagnosand names, estimated diagnosand values, bootstrapped standard errors and confidence intervals

Examples

effect_size <- 0.1
design <-
  declare_model(
    N = 100,
    U = rnorm(N),
    X = rnorm(N),
    potential_outcomes(Y ~ effect_size * Z + X + U)
  ) +
  declare_inquiry(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_assignment(Z = complete_ra(N)) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z)) +
  declare_estimator(Y ~ Z, inquiry = "ATE", label = "unadjusted") + 
  declare_estimator(Y ~ Z + X, inquiry = "ATE", label = "adjusted")

diagnosis <- diagnose_design(design, sims = 100)

tidy(diagnosis)
#>    design inquiry  estimator outcome term    diagnosand     estimate  std.error
#> 1  design     ATE   adjusted       Y    Z mean_estimand 1.000000e-01 0.00000000
#> 2  design     ATE   adjusted       Y    Z mean_estimate 1.000850e-01 0.02046566
#> 3  design     ATE   adjusted       Y    Z          bias 8.497065e-05 0.02046566
#> 4  design     ATE   adjusted       Y    Z   sd_estimate 2.042096e-01 0.01520222
#> 5  design     ATE   adjusted       Y    Z          rmse 2.031860e-01 0.01522089
#> 6  design     ATE   adjusted       Y    Z         power 7.000000e-02 0.02792197
#> 7  design     ATE   adjusted       Y    Z      coverage 9.600000e-01 0.02035518
#> 8  design     ATE unadjusted       Y    Z mean_estimand 1.000000e-01 0.00000000
#> 9  design     ATE unadjusted       Y    Z mean_estimate 1.167920e-01 0.02441034
#> 10 design     ATE unadjusted       Y    Z          bias 1.679204e-02 0.02441034
#> 11 design     ATE unadjusted       Y    Z   sd_estimate 2.584291e-01 0.01538932
#> 12 design     ATE unadjusted       Y    Z          rmse 2.576814e-01 0.01564422
#> 13 design     ATE unadjusted       Y    Z         power 6.000000e-02 0.02308176
#> 14 design     ATE unadjusted       Y    Z      coverage 9.600000e-01 0.01821837
#>       conf.low  conf.high
#> 1   0.10000000 0.10000000
#> 2   0.05975347 0.13445326
#> 3  -0.04024653 0.03445326
#> 4   0.17638935 0.23351744
#> 5   0.17581545 0.23326839
#> 6   0.03000000 0.13525000
#> 7   0.93000000 0.99000000
#> 8   0.10000000 0.10000000
#> 9   0.06879978 0.15752500
#> 10 -0.03120022 0.05752500
#> 11  0.22775625 0.28086504
#> 12  0.22729166 0.28413356
#> 13  0.02000000 0.10525000
#> 14  0.92475000 0.99525000